% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIGER.R
\name{adj2el}
\alias{adj2el}
\title{Convert a bipartite adjacency matrix to an edgelist}
\usage{
adj2el(adj)
}
\arguments{
\item{adj}{An adjacency matrix, with rows as TFs and columns as genes.}
}
\value{
An edge list dataframe with three columns. First column is TF name,
second column is gene name, and third column is edge weight.
}
\description{
Convert a bipartite adjacency matrix to an edgelist
}
