% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIGER.R
\name{adj2regulon}
\alias{adj2regulon}
\title{Convert bipartite adjacency to regulon}
\usage{
adj2regulon(adj)
}
\arguments{
\item{adj}{An adjacency matrix, with rows as TFs and columns as genes.}
}
\value{
A VIPER required regulon object.
}
\description{
Convert bipartite adjacency to regulon
}
