% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANE.R
\name{adjMatToElist}
\alias{adjMatToElist}
\title{converts adjacency matrix to edge list}
\usage{
adjMatToElist(adj_mat)
}
\arguments{
\item{adj_mat}{adjacency matrix}
}
\value{
edge list
}
\description{
converts adjacency matrix to edge list
}
