% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPACAFunctions.R
\name{alpaca}
\alias{alpaca}
\title{Main ALPACA function}
\usage{
alpaca(net.table, file.stem, verbose = F)
}
\arguments{
\item{net.table}{A table of edges, with the first column representing the TFs ("from" nodes) and the second column representing the targets ("to" nodes). The third column contains the edge weights corresponding to the control or healthy network, and the fourth column contains the edge weights for the disease network or network of interest.}

\item{file.stem}{The folder location and title under which all results will be stored.}

\item{verbose}{Indicates whether the full differential modularity matrix should also be written to a file. Defaults to FALSE.}
}
\value{
List where first element is the membership vector and second element is the contribution score of each node to its module's total differential modularity
}
\description{
This function compares two networks and finds the sets of nodes that best characterize the change in modular structure
}
\examples{
alpaca()
}
\keyword{modularity}
