% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPACA.R
\name{alpaca.TestNodeRank}
\alias{alpaca.TestNodeRank}
\title{Enrichment in ranked list}
\usage{
alpaca.TestNodeRank(node.ordered, true.pos)
}
\arguments{
\item{node.ordered}{An ordered list of nodes (high-scoring to low-scoring).}

\item{true.pos}{The selected set of nodes being tested for enrichment among the ranked list.}
}
\value{
A vector of 4 values. 1) Wilcoxon p-value, 2) KS p-value, 3) Fisher p-value, 4) Fisher odds ratio.
}
\description{
This function computes the enrichment of selected nodes in a ranked list, using Wilcoxon, Kolmogorov-Smirnov, and Fisher exact tests.
}
\examples{
a <- 1 # example place holder
}
