% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANE.R
\name{alpacaComputeDifferentialScoreFromDWBM}
\alias{alpacaComputeDifferentialScoreFromDWBM}
\title{Compute Differential modularity score from differential modularity matrix}
\usage{
alpacaComputeDifferentialScoreFromDWBM(dwbm, louv.memb)
}
\arguments{
\item{dwbm}{differential modularity matrix}

\item{louv.memb}{louvain community membership}
}
\value{
Vector of differntial modularity score
}
\description{
This functions takes the precomputed differential modularity matrix and the genLouvain membership to compute the differential modularity score.
}
