% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANE.R
\name{alpacaCrane}
\alias{alpacaCrane}
\title{Find the robust nodes in ALPACA community using CRANE}
\usage{
alpacaCrane(input, alp, alpha = 0.1, beta = 0, iteration = 30, isParallel = F)
}
\arguments{
\item{input}{same input for alpaca: first column TF, second column Genes, third column edge weights from baseline condition, fourth column edge weights from disease condition.}

\item{alp}{alpca object in list format (output from alpaca package)}

\item{alpha}{alpha paramter perturbs each edge weights}

\item{beta}{beta parameter perturbs the strength of each node. Set this to 0 if you want nodes to have node strength identical to the orignal network.}

\item{iteration}{Number of CRANE distributions to create. Higher value leads to better ranking but longer runtime.}

\item{isParallel}{TRUE = use Multithread / FALSE = do not use Multithread}
}
\value{
list of data frames
}
\description{
Find the robust nodes in ALPACA community using CRANE
}
\examples{
\dontrun{

input=cbind(nonAng,ang[,3])
alp=alpaca(input,NULL,verbose = F)
alpListObject=alpacaCrane(input, alp, isParallel = T)

}
}
