% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANE.R
\name{alpacaGetMember}
\alias{alpacaGetMember}
\title{get the member vector from alpaca object}
\usage{
alpacaGetMember(alp, target = "all")
}
\arguments{
\item{alp}{alpaca object}

\item{target}{tf, gene, or all}
}
\value{
member vector
}
\description{
get the member vector from alpaca object
}
