% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPACA.R
\name{alpacaGoToGenes}
\alias{alpacaGoToGenes}
\title{Map GO terms to gene symbols}
\usage{
alpacaGoToGenes(go.term)
}
\arguments{
\item{go.term}{The GO Biological Process ID (string).}
}
\value{
A vector of all gene symbols associated with the GO term.
}
\description{
This function extracts all the gene symbols associated with a GO term and its descendants. (v1)
}
\examples{
a <- 1 # example place holder
}
