% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPACA.R
\name{alpacalisttogo}
\alias{alpacalisttogo}
\title{GO term enrichment for a list of gene sets}
\usage{
alpacalisttogo(gene.list, univ.vec, comm.nums)
}
\arguments{
\item{gene.list}{A list consisting of vectors of genes; genes must be identified by their official gene symbols.}

\item{univ.vec}{A vector of  all gene symbols that were present in the original network. This set is used as the universe for running the hypergeometric test in GOstats.}

\item{comm.nums}{A vector of names for the gene sets in the input parameter "gene.list". These are used to create the table of final results.}
}
\value{
A table whose rows represent enriched GO terms (p_adj<0.05) and columns describe useful properties, like the name of the GO term, the label of the gene set which is enriched in that GO term, the adjusted p-value and Odds Ratio.
}
\description{
GO term enrichment is run using the GOstats package, and corrected for multiple testing using the Benjamini-Hochberg method.
}
\examples{
a <- 1 # example place holder
}
