% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANE.R
\name{alpacaObjectToDfList}
\alias{alpacaObjectToDfList}
\title{Converts alpaca output into list of data frames}
\usage{
alpacaObjectToDfList(alp)
}
\arguments{
\item{alp}{alpaca object}
}
\value{
list of data frames
}
\description{
Converts alpaca output into list of data frames
}
