% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{AUCTrapezoid}
\alias{AUCTrapezoid}
\title{Given the values for axis X and axis Y, this function computes the AUC score
with linear interpolation between the points. This is a helper function for
ComputeAUC.}
\usage{
AUCTrapezoid(x, y, absoluteMin = NULL, absoluteMax = NULL)
}
\arguments{
\item{x}{The value on the X axis}

\item{y}{The value on the Y axis}

\item{absoluteMin}{The absolute minimum value. If NULL, it is the minimum
value across both X and Y.}

\item{absoluteMax}{The absolute maximum value. If NULL, it is the maximum
value across both X and Y.}
}
\value{
The AUC score
}
\description{
Given the values for axis X and axis Y, this function computes the AUC score
with linear interpolation between the points. This is a helper function for
ComputeAUC.
}
