% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regpredict.R
\name{bereFull}
\alias{bereFull}
\title{Bipartite Edge Reconstruction from Expression data 
(composite method with direct/indirect)}
\usage{
bereFull(motifs, exprData, alpha = 0.5, penalized = TRUE, lambda = 10,
  score = "motifincluded")
}
\arguments{
\item{motifs}{A motif dataset, a data.frame, matrix or exprSet 
containing 3 columns. Each row describes an motif associated 
with a transcription factor (column 1) a gene (column 2) 
and a score (column 3) for the motif.}

\item{exprData}{An expression dataset, as a genes (rows) by 
samples (columns) data.frame}

\item{alpha}{A weight parameter specifying proportion of weight 
to give to indirect compared to direct evidence.  See documentation.}

\item{penalized}{logical indicating whether or not to use penalized logistic regression}

\item{lambda}{if using penalized, the lambda parameter in the penalized logistic regression}

\item{score}{String to indicate whether motif information will 
be readded upon completion of the algorithm}

\item{verbose}{logical to indicate printing of output for 
algorithm progress.}

\item{method}{String to indicate algorithm method.  Must be 
one of "cd","lda", or "wcd". Default is correlation difference 
"cd".}
}
\value{
An matrix or data.frame
}
\description{
This function generates a complete bipartite network from 
gene expression data and sequence motif data
}
\keyword{keywords}

