% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARN.R
\docType{data}
\name{Bladder_data}
\alias{Bladder_data}
\title{Bladder RNA-seq data from the GTEx consortium}
\format{
An object of class \code{"ExpressionSet"}; see \code{\link[Biobase]{ExpressionSet}}.
}
\source{
GTEx Portal
}
\usage{
data(bladder)
}
\value{
ExpressionSet object
}
\description{
Bladder RNA-seq data from the GTEx consortium. V6 release.
}
\examples{
\donttest{data(bladder);
checkMisAnnotation(bladder, "GENDER");}
}
\references{
GTEx Consortium, 2015. The Genotype-Tissue Expression (GTEx) pilot analysis: Multitissue gene regulation in humans. Science, 348(6235), pp.648-660.
(\href{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC4547484/}{PubMed})
}
\keyword{datasets}
