% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNAGI.R
\name{BreadthFirstSearchU}
\alias{BreadthFirstSearchU}
\title{Find all edges adjacent to the starting nodes, excluding the nodes
specified.}
\usage{
BreadthFirstSearchU(
  networks,
  startingNodes,
  nodesToExclude,
  verbose = FALSE,
  topX = NULL
)
}
\arguments{
\item{networks}{A PANDA-like network}

\item{startingNodes}{The list of nodes from which to start.}

\item{nodesToExclude}{The list of nodes to exclude from the search.}

\item{verbose}{Whether or not to print detailed information about the run.}

\item{topX}{Select the X lowest significant p-values for each gene. NULL by default.}
}
\description{
Find all edges adjacent to the starting nodes, excluding the nodes
specified.
}
