% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{BuildComparisonObject}
\alias{BuildComparisonObject}
\title{Given the results, the source network, and the ingroup and outgroup, this function
builds a list of comparisons to make from the source network to the ingroup and
from the source network to the outgroup. If inhibitory or activating networks are specified,
then these comparisons are segmented out.}
\usage{
BuildComparisonObject(
  sourceNetwork,
  ingroupToCompare,
  outgroupToCompare,
  results
)
}
\arguments{
\item{sourceNetwork}{The name of the file with the source network.}

\item{ingroupToCompare}{A list of files to compare with the source network (the in-group).}

\item{outgroupToCompare}{A list of files to compare with the target network (the out-group).}

\item{results}{An object of type FERRET_Results.}
}
\value{
A network in the format of an adjacency list
}
\description{
Given the results, the source network, and the ingroup and outgroup, this function
builds a list of comparisons to make from the source network to the ingroup and
from the source network to the outgroup. If inhibitory or activating networks are specified,
then these comparisons are segmented out.
}
