% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TARDIGRADE.R
\name{BuildPathwayGraph}
\alias{BuildPathwayGraph}
\title{Builds the pathway graph from RaMP and stores in in the file specified by the user.}
\usage{
BuildPathwayGraph(analyteHasPathway, pathwaySizeLimit = -1)
}
\arguments{
\item{analyteHasPathway}{A data frame mapping analytes (column "Analyte") to 
pathways (column "Pathway").}

\item{pathwaySizeLimit}{The maximum number of analytes a pathway can contain.
Any pathways that are larger will be excluded from analysis. If -1, all are included.}
}
\value{
An object of type TARDIGRADE_KnowledgeGraph.
}
\description{
Builds the pathway graph from RaMP and stores in in the file specified by the user.
}
