% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLOBFISH.R, R/UNAGI.R
\name{BuildSubnetwork}
\alias{BuildSubnetwork}
\title{Find the subnetwork of significant edges connecting the genes.}
\usage{
BuildSubnetwork(
  geneSet,
  networks,
  alpha,
  hopConstraint,
  nullDistribution,
  verbose = FALSE,
  topX = NULL,
  doFDRAdjustment = TRUE
)

BuildSubnetwork(
  geneSet,
  networks,
  alpha,
  hopConstraint,
  nullDistribution,
  verbose = FALSE,
  topX = NULL,
  doFDRAdjustment = TRUE
)
}
\arguments{
\item{geneSet}{A character vector of genes comprising the targets of interest.}

\item{networks}{A list of  unipartite (PANDA-like) networks, where each network is a data frame with the following format:
tf,gene,score}

\item{alpha}{The significance cutoff for the statistical test.}

\item{hopConstraint}{The maximum number of hops to be considered between gene pairs.
Must be an even number.}

\item{nullDistribution}{The null distribution, specified as a vector of values.}

\item{verbose}{Whether or not to print detailed information about the run.}

\item{topX}{Select the X lowest significant p-values for each gene. NULL by default.}

\item{doFDRAdjustment}{Whether or not to perform FDR adjustment.
parameter allows the edges to be split into chunks to prevent memory errors.
saved and need to be recalculated.
Default is FALSE.}

\item{pValueChunks}{The number of chunks to split when calculating the p-value. This
parameter allows the edges to be split into chunks to prevent memory errors.}

\item{pValueFile}{The file where the p-values should be saved. If NULL, they are not
saved and need to be recalculated.}

\item{loadPValues}{Whether p-values should be loaded from pValueFile or re-generated.
Default is FALSE.}
}
\value{
A bipartite subnetwork in the same format as the original networks.

A  unipartite subnetwork in the same format as the original networks.
}
\description{
Find the subnetwork of significant edges connecting the genes.

Find the subnetwork of significant edges connecting the genes.
}
