% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TARDIGRADE.R
\name{CalculateFruchtermanReingoldLayout}
\alias{CalculateFruchtermanReingoldLayout}
\title{Calculates the Fruchterman-Reingold layout.}
\usage{
CalculateFruchtermanReingoldLayout(knowledgeGraph)
}
\arguments{
\item{knowledgeGraph}{An object of type TARDIGRADE_KnowledgeGraph.}
}
\value{
A numeric matrix with two (dim=2) or three (dim=3) columns, 
and as many rows as the number of vertices, the x, y and potentially z 
coordinates of the vertices.
}
\description{
Calculates the Fruchterman-Reingold layout.
}
