% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{CalculateOneTransitionMatrix}
\alias{CalculateOneTransitionMatrix}
\title{Calculates a single transition matrix given case and control data.}
\usage{
CalculateOneTransitionMatrix(
  i,
  mode,
  nullExpr,
  motif,
  ni_method,
  ni.coefficient.cutoff,
  alphaw,
  design,
  remove.diagonal,
  method,
  outputDir,
  logging,
  by.tfs = TRUE
)
}
\arguments{
\item{i}{Iteration}

\item{mode}{"buildNet" or "regNet"}

\item{nullExpr}{Gene expression dataset}

\item{motif}{Regulatory data.frame consisting of three columns.  For each row, a transcription factor (column 1) 
regulates a gene (column 2) with a defined strength (column 3), usually taken to be 0 or 1.
May also be NA, if MONSTER is being run on pre-computed gene regulatory networks, passed in the `expr` argument.
See also `domonster` to use this mode.}

\item{ni_method}{String to indicate algorithm method.  Must be one of "BERE","pearson",or "lda". Default is "BERE"}

\item{ni.coefficient.cutoff}{numeric to specify a p-value cutoff at the network
inference step.  Default is NA, indicating inclusion of all coefficients.}

\item{alphaw}{A weight parameter between 0 and 1 specifying proportion of weight 
to give to indirect compared to direct evidence. The default is 0.5 to give an 
equal weight to direct and indirect evidence.}

\item{design}{The design matrix}

\item{remove.diagonal}{#' Logical for returning a result containing 0s across the diagonal (default = TRUE).}

\item{method}{character specifying which algorithm to use, default='ols'.}

\item{outputDir}{character vector specifying a directory or path in which 
which to save MONSTER results, default is NA and results are not saved.}

\item{logging}{Whether or not to print logging messages for MONSTER (default is TRUE)}

\item{by.tfs}{logical indicating a transcription factor based transformation.    If 
false, gives gene by gene transformation matrix}
}
\value{
A list of matrices, each one corresponding to one null model
}
\description{
Calculates a single transition matrix given case and control data.
}
