% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLOBFISH.R
\name{CalculatePValues}
\alias{CalculatePValues}
\title{Calculate p-values for all edges in the network using a Wilcoxon two-sample test
for each edge.}
\usage{
CalculatePValues(
  network,
  nullDistribution,
  pValueChunks = 100,
  doFDRAdjustment = TRUE,
  pValueFile = "pvalues.RDS",
  verbose = FALSE
)
}
\arguments{
\item{network}{A combination of PANDA-like networks, with the following format 
(e.g., 3 networks), provided as a data frame:
tf,gene,score1,score2,score3}

\item{nullDistribution}{The null distribution, specified as a vector of values.}

\item{pValueChunks}{The number of chunks to split when calculating the p-value. This
parameter allows the edges to be split into chunks to prevent memory errors.}

\item{doFDRAdjustment}{Whether or not to perform FDR adjustment.}

\item{pValueFile}{The file where the p-values should be saved. If NULL, they are not
saved and need to be recalculated.}

\item{verbose}{Whether or not to print detailed information about the run.}
}
\value{
A vector of p-values, one for each edge.
}
\description{
Calculate p-values for all edges in the network using a Wilcoxon two-sample test
for each edge.
}
