% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{computeCorrelations}
\alias{computeCorrelations}
\title{Computes correlation for numeric phenotypes and ANOVA for categorical
phenotypes.}
\usage{
computeCorrelations(
  expression,
  phenotype,
  phenotype_dictionary,
  pathways,
  method,
  pval_adj_method
)
}
\arguments{
\item{expression}{: gene expression matrix (normalized, and filtered) 
with rows as genes and columns as samples.
Row and column names must be present.
Row names must be HGNC symbols.
Column names must match the row names of the phenotype matrix.}

\item{phenotype}{: phenotype matrix
with rows as samples and columns as phenotype variables.}

\item{phenotype_dictionary}{: a vector of strings
containing type of each phenotype.
Types can be either "numeric" or "categorical"}

\item{pathways}{: a list of pathways (e.g. KEGG, GO, Reactome etc. 
downloaded from http://www.gsea-msigdb.org/gsea/msigdb/human/collections.jsp)}

\item{method}{: One of "pearson", "spearman", or "kendall".}

\item{pval_adj_method}{Wrapper for p.adjust. Defaults to "none" (no adjustment). Other options are
"holm", "hochberg", "hommel", "bonferroni", "BH", "BY", and "fdr".}
}
\description{
Computes correlation for numeric phenotypes and ANOVA for categorical
phenotypes.
}
