% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{computeCorrelationsNetwork}
\alias{computeCorrelationsNetwork}
\title{Computes network-phenotype correlation for continuous phenotypes with continuous network features,
t-test for continuous phenotypes with dichotomous network features, ANOVA for categorical
phenotypes with continuous network features, t-test for dichotomous phenotypes with continuous network features,
and FFH for categorical or dichotomous phenotypes with dichotomous network features.}
\usage{
computeCorrelationsNetwork(
  network,
  phenotype,
  phenotype_dictionary,
  pathways,
  method,
  pval_adj_method,
  featureType
)
}
\arguments{
\item{network}{: A weighted matrix where rows are samples and columns are network features.}

\item{phenotype}{: phenotype matrix
with rows as samples and columns as phenotype variables.}

\item{phenotype_dictionary}{: a vector of strings
containing type of each phenotype.
Types can be either "numeric" or "categorical"}

\item{pathways}{: Pathways used for GSEA calculation}

\item{method}{: One of "pearson", "spearman", or "kendall".}

\item{pval_adj_method}{: Wrapper for p.adjust. Defaults to "none" (no adjustment). Other options are
"holm", "hochberg", "hommel", "bonferroni", "BH", "BY", and "fdr".}

\item{featureType}{: Whether network features are continuous or dichotomous.}
}
\description{
Computes network-phenotype correlation for continuous phenotypes with continuous network features,
t-test for continuous phenotypes with dichotomous network features, ANOVA for categorical
phenotypes with continuous network features, t-test for dichotomous phenotypes with continuous network features,
and FFH for categorical or dichotomous phenotypes with dichotomous network features.
}
