% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{computeLinearRegressionNetwork}
\alias{computeLinearRegressionNetwork}
\title{Computes linear regression for both numeric and categorical phenotypes with respect to network. 
Computes logistic regression when the outcome (numeric) is dichotomous.}
\usage{
computeLinearRegressionNetwork(
  network,
  phenotype,
  phenotype_dictionary,
  pval_adj_method,
  featureType
)
}
\arguments{
\item{network}{: A weighted matrix where rows are samples and columns are network features.}

\item{phenotype}{: phenotype matrix
with rows as samples and columns as phenotype variables.}

\item{phenotype_dictionary}{: a vector of strings
containing type of each phenotype.
Types can be either "numeric" or "categorical"}

\item{pval_adj_method}{Wrapper for p.adjust. Defaults to "none" (no adjustment). Other options are
"holm", "hochberg", "hommel", "bonferroni", "BH", "BY", and "fdr".}

\item{featureType}{"continuous" or "dichotomous" (for binary networks)}
}
\description{
Computes linear regression for both numeric and categorical phenotypes with respect to network. 
Computes logistic regression when the outcome (numeric) is dichotomous.
}
