% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{computePhenotypeCorrelations}
\alias{computePhenotypeCorrelations}
\title{Computes phenotype-phenotype correlation for continuous phenotypes, ANOVA for nominal
phenotypes, and t-test for dichotomous phenotypes}
\usage{
computePhenotypeCorrelations(
  phenotype,
  phenotype_dictionary,
  method,
  pval_adj_method
)
}
\arguments{
\item{phenotype}{: phenotype matrix
with rows as samples and columns as phenotype variables.}

\item{phenotype_dictionary}{: a vector of strings
containing type of each phenotype.
Types can be either "numeric" or "categorical"}

\item{method}{: One of "pearson", "spearman", or "kendall".}

\item{pval_adj_method}{Wrapper for p.adjust. Defaults to "none" (no adjustment). Other options are
"holm", "hochberg", "hommel", "bonferroni", "BH", "BY", and "fdr".}
}
\description{
Computes phenotype-phenotype correlation for continuous phenotypes, ANOVA for nominal
phenotypes, and t-test for dichotomous phenotypes
}
