% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{ComputeProjection}
\alias{ComputeProjection}
\title{Helper function to compute the projection of a network.}
\usage{
ComputeProjection(network, k, allNodes)
}
\arguments{
\item{network}{A network to project (N x N).}

\item{k}{Number of eigenvectors to include in the subspace projection.}

\item{allNodes}{Additional nodes to add to the adjacency matrix.}
}
\value{
The projection (a matrix of N x k)
}
\description{
Helper function to compute the projection of a network.
}
