% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{ComputeRobustnessForOneEdgeType}
\alias{ComputeRobustnessForOneEdgeType}
\title{Helper function that computes the AUC score for either all edges or
for activating and inhibitory edges separately.}
\usage{
ComputeRobustnessForOneEdgeType(
  results,
  comparisons,
  metric = c("jaccard", "degree", "modularity", "subspace"),
  k = 5,
  plotCurve = TRUE,
  xlab = "",
  ylab = "",
  mode = "Score",
  numberOfCutoffs = 10
)
}
\arguments{
\item{results}{An object of type FERRET_Results}

\item{comparisons}{An object of type FERRET_Comparisons. Note that the 
in-group and out-group names in FERRET_Comparisons MUST match the names
in FERRET_Results.}

\item{metric}{The metric to use for computing similarity. One or more elements
in the following list:
- jaccard: Computes the Jaccard similarity (intersection / union) of edges.
- degree: Computes the overlap in degree of each node.
- modularity: Computes the differential modularity from a "baseline" network
to each "perturbed" network, then scales each modularity value between 0 and 1.
- subspace: Computes the geodesic distance between each pair of networks
on the Grassmann manifold for a given input k, then scales each
distance between 0 and 1.
Default is all elements.}

\item{k}{is the number of eigenvectors to include in the subspace representation
of each network. Default is 5.}

\item{plotCurve}{Whether or not to generate a plot. Default is TRUE.}

\item{xlab}{The label for the X axis of the plot. Default is empty. This only
needs to be set when generating the plot.}

\item{ylab}{The label for the Y axis of the plot. Default is empty. This only
needs to be set when generating the plot.}

\item{mode}{"Percentile" or "Score".}

\item{numberOfCutoffs}{Number of cutoffs to evaluate.}
}
\value{
An AUC score, also generates a  plot as a side effect.
}
\description{
Helper function that computes the AUC score for either all edges or
for activating and inhibitory edges separately.
}
