% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{ComputeSimilarities}
\alias{ComputeSimilarities}
\title{Helper function to compute the similarities between the baseline network,
the ingroup, and the outgroup.}
\usage{
ComputeSimilarities(results, comparisons, cutoffs, metric, k = 5)
}
\arguments{
\item{results}{Object of type FERRET_Results}

\item{comparisons}{Object of type FERRET_Comparisons}

\item{cutoffs}{The list of cutoffs to use when filtering the networks}

\item{metric}{The metric to use for computing similarity. An element
in the following list:
- jaccard: Computes the Jaccard similarity (intersection / union) of edges.
- degree: Computes the overlap in degree of each node.
- modularity: Computes the differential modularity from a "baseline" network
to each "perturbed" network, then scales each modularity value between 0 and 1.
- subspace: Computes the geodesic distance between each pair of networks
on the Grassmann manifold for a given input k, then scales each
distance between 0 and 1.
Default is all elements.}

\item{k}{is the number of eigenvectors to include in the subspace representation
of each network. Default is 5.}
}
\value{
Object of type FERRET_Similarity
}
\description{
Helper function to compute the similarities between the baseline network,
the ingroup, and the outgroup.
}
