% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPACAFunctions.R
\name{computeWBMmat}
\alias{computeWBMmat}
\title{Compute modularity matrix for weighted bipartite network}
\usage{
computeWBMmat(edge.mat)
}
\arguments{
\item{edge.mat}{A table of edges, with the first column representing the TFs ("from" nodes) and the second column representing the targets ("to" nodes). The third column contains the edge weights corresponding to the network of interest.}
}
\value{
Modularity matrix with rows representing TFs ("from" nodes) and columns repesenting targets ("to" nodes)
}
\description{
This function computes the modularity matrix for a weighted bipartite network.
}
\examples{
computeWBMmat()
}
