% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condor.core.enrich.R
\name{condor.core.enrich}
\alias{condor.core.enrich}
\title{Compare qscore distribution of a subset of nodes to all other nodes.}
\usage{
condor.core.enrich(test_nodes, q, perm = FALSE, plot.hist = FALSE,
  nsamp = 1000)
}
\arguments{
\item{test_nodes}{is a list containing the subset of nodes (of one node class
--blue or red--only) to be tested}

\item{q}{is a two column data frame containing the node names in the
first column and the q-scores in the second column.}

\item{perm}{if TRUE, run permutation tests. Else, run
\code{\link[stats]{ks.test}} and \code{\link[stats]{wilcox.test}} only.}

\item{plot.hist}{if TRUE, produces two histograms of test statistics
from permutation tests, one for KS and one for
wilcoxon and a red dot for true labeling. Only works if perm=TRUE.}

\item{nsamp}{Number of permutation tests to run}
}
\value{
if \code{perm=FALSE}, the analytical p-values from
\code{\link[stats]{ks.test}} and \code{\link[stats]{wilcox.test}}

if \code{perm=TRUE}, the permutation p-values are provided in
addition to the analytical values.
}
\description{
Compute one-sided KS and wilcox tests to determine if a subset of nodes
has a stochastically larger qscore distribution.
}
\note{
\code{\link[stats]{ks.test}} and \code{\link[stats]{wilcox.test}}
will throw warnings due to the presence of ties, so the p-values will be
approximate. See those functions for further details.
}
\examples{
r = c(1,1,1,2,2,2,3,3,3,4,4);
b = c(1,2,3,1,2,4,2,3,4,3,4);
reds <- c("Alice","Sue","Janine","Mary")
blues <- c("Bob","John","Ed","Hank")
elist <- data.frame(red=reds[r],blue=blues[b])
condor.object <- create.condor.object(elist)
condor.object <- condor.cluster(condor.object)
condor.object <- condor.qscore(condor.object)
q_in <- condor.object$qscores$red.qscore
out <- condor.core.enrich(c("Alice","Mary"),q=q_in,perm=TRUE,plot.hist=TRUE)
}

