% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condor.plot.heatmap.R
\name{condor.plot.heatmap}
\alias{condor.plot.heatmap}
\title{Plot weighted adjacency matrix with links grouped by community}
\usage{
condor.plot.heatmap(condor.object, main = "", xlab = "blues",
  ylab = "reds")
}
\arguments{
\item{condor.object}{output of either \code{\link{condor.cluster}} or
\code{\link{condor.modularity.max}}}

\item{main}{plot title}

\item{xlab}{x axis label}

\item{ylab}{y axis label}
}
\value{
produces a \code{\link[graphics]{plot}} output.
}
\description{
This function will generate the network link 'heatmap' for a weighted network
}
\examples{
data(small1976)
condor.object <- create.condor.object(small1976)
condor.object <- condor.cluster(condor.object, project=FALSE)
condor.plot.heatmap(condor.object)
}

