% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDORFunctions.R
\name{condor.qscore}
\alias{condor.qscore}
\title{Calculate Qscore for all nodes}
\usage{
condor.qscore(condor.object)
}
\arguments{
\item{condor.object}{output of \code{\link{condor.cluster}} or 
\code{\link{condor.modularity.max}}}
}
\value{
condor.object list has \code{condor.object$qscores} added to it.
this includes two data.frames, \code{blue.qscore} and \code{red.qscore}
which have the qscore for each red and blue node.
}
\description{
Qscore is designed to calculate the fraction of the modularity 
contributed by each node to its community's modularity
}
\examples{
r = c(1,1,1,2,2,2,3,3,3,4,4);
b = c(1,2,3,1,2,4,2,3,4,3,4);
reds <- c("Alice","Sue","Janine","Mary")
blues <- c("Bob","John","Ed","Hank")
elist <- data.frame(red=reds[r],blue=blues[b])
condor.object <- create.condor.object(elist)
condor.object <- condor.cluster(condor.object)
condor.object <- condor.qscore(condor.object)   

}
