% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANE.R
\name{condorCreateObject}
\alias{condorCreateObject}
\title{creates condor object}
\usage{
condorCreateObject(elist)
}
\arguments{
\item{elist}{edge list}
}
\value{
condor object
}
\description{
creates condor object
}
