% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDOR.R
\name{condorPlotHeatmap}
\alias{condorPlotHeatmap}
\title{Plot weighted adjacency matrix with links grouped by community}
\usage{
condorPlotHeatmap(condor.object, main = "", xlab = "blues", ylab = "reds")
}
\arguments{
\item{condor.object}{output of either \code{\link{condorCluster}} or 
\code{\link{condorModularityMax}}}

\item{main}{plot title}

\item{xlab}{x axis label}

\item{ylab}{y axis label}
}
\value{
produces a \code{\link{plot}} output.
}
\description{
This function will generate the network link 'heatmap' for a weighted network
}
\examples{
data(small1976)
condor.object <- createCondorObject(small1976)
condor.object <- condorCluster(condor.object, project=FALSE)
condorPlotHeatmap(condor.object)
}
