% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANE.R
\name{condorRun}
\alias{condorRun}
\title{Run CONDOR clustering}
\usage{
condorRun(elist, qscore = F)
}
\arguments{
\item{elist}{edge list}

\item{qscore}{TRUE = output qscore / FALSE = do not output qscore}
}
\value{
condor object
}
\description{
Run CONDOR clustering
}
