% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{ConsolidateRobustnessHelper}
\alias{ConsolidateRobustnessHelper}
\title{This is a helper function for ConsolidateRobustness().}
\usage{
ConsolidateRobustnessHelper(resultList, functionToApply, metric)
}
\arguments{
\item{resultList}{A list of FERRET_ROC_AUC objects to consolidate.}

\item{functionToApply}{One of "mean", "min", or "max".}

\item{metric}{The metric to consider.}
}
\value{
a vector of new results.
}
\description{
This is a helper function for ConsolidateRobustness().
}
