% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{ConsolidateRobustnessHelperOld}
\alias{ConsolidateRobustnessHelperOld}
\title{This is a helper function for ConsolidateRobustness().}
\usage{
ConsolidateRobustnessHelperOld(resultList, functionToApply, metric, group)
}
\arguments{
\item{resultList}{A list of FERRET_ROC_AUC objects to consolidate.}

\item{functionToApply}{One of "mean", "min", or "max".}

\item{metric}{The metric to consider.}

\item{group}{One of "ingroup" or "outgroup"}
}
\value{
a vector of new results.
}
\description{
This is a helper function for ConsolidateRobustness().
}
