% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertgmt.R
\name{convertgmt}
\alias{convertgmt}
\title{Convert .gmt files into a binary matrix.}
\usage{
convertgmt(signature, cagenes, ...)
}
\arguments{
\item{signature}{A file containing gene sets (signatures) in .gmt format. These gene sets will be used to de-sparsify the gene-level mutation scores.}

\item{cagenes}{A vector of genes, for example of cancer-associated genes. This will be used to subset the gene-level mutation data to.}
}
\value{
A matrix containing gene set mutation scores.
}
\description{
Convert .gmt files into a binary matrix.
}
