% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMBARFunctions.R
\name{corgenelength}
\alias{corgenelength}
\title{Normalize gene mutation scores by gene length.}
\usage{
corgenelength(x, cagenes, exonsize, ...)
}
\arguments{
\item{x}{Mutation data, in the format of a matrix, including the number of mutations for samples (rows) and genes (columns).}

\item{cagenes}{A vector of genes, for example of cancer-associated genes. This will be used to subset the gene-level mutation data to.}

\item{exonsize}{A vector of gene lengths. This will be used to normalize the gene mutation scores.}
}
\value{
Mutation rate-adjusted gene mutation scores.
}
\description{
Normalize gene mutation scores by gene length.
}
