% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANE.R
\name{craneBipartite}
\alias{craneBipartite}
\title{Pertrubs the bipartite network with fixed node strength}
\usage{
craneBipartite(df, alpha = 0.1, beta = 0, getAdj = F, randomStart = F)
}
\arguments{
\item{df}{Adjacency Matrix or Edge list}

\item{alpha}{alpha paramter perturbs each edge weights}

\item{beta}{beta parameter perturbs the strength of each node. Set this to 0 if you want nodes to have node strength identical to the orignal network.}

\item{getAdj}{TRUE = this will return adjacency matrix instead of edge list}

\item{randomStart}{FALSE = initialize the first row with completely random edge weights.}
}
\value{
edge list
}
\description{
Pertrubs the bipartite network with fixed node strength
}
\examples{
\dontrun{

# Using Edge list as input
elist=craneBipartite(nonAng)
elist=craneBipartite(nonAng,alpha=0.3)

# Using Edge list as input and Adjcency Matrix as output
adjMatrix=craneBipartite(nonAng,alpha=0.1,getAdj=T)

# Using Edge list as input and Adjcency Matrix as output
A=elistToAdjMat(nonAng)
elist=craneBipartite(A)

}
}
