% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANE.R
\name{craneUnipartite}
\alias{craneUnipartite}
\title{Pertrubs the unipartite network with fixed node strength from adjacency matrix}
\usage{
craneUnipartite(A, alpha = 0.1, isSelfLoop = F)
}
\arguments{
\item{A}{Adjacency Matrix}

\item{alpha}{alpha paramter perturbs each edge weights}

\item{isSelfLoop}{TRUE/FALSE if self loop exists. co-expression matrix will have a self-loop of 1. Thus TRUE}
}
\value{
adjacency matrix
}
\description{
Pertrubs the unipartite network with fixed node strength from adjacency matrix
}
