% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.panda.style.R
\name{create.panda.style}
\alias{create.panda.style}
\title{Create a Cytoscape visual style for PANDA network}
\usage{
create.panda.style(style_name = "PandaStyle")
}
\arguments{
\item{style.name}{Character string indicating the style name. Defaults to "PandaStyle"}
}
\value{
a visual style in Cytoscape Control Panel under "Style" button.
}
\description{
This function is able to create a Cytoscape visual style for any PANDA network output.
}
