% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDiffPandaStyle.R
\name{createDiffPandaStyle}
\alias{createDiffPandaStyle}
\title{Create a Cytoscape visual style for two PANDA networks}
\usage{
createDiffPandaStyle(style_name = "Diff.PandaStyle", condition_name = "cond.1")
}
\arguments{
\item{style_name}{Character string indicating the style name. Defaults to "Diff.PandaStyle"}

\item{condition_name}{string vector indicating the same condition name used in \code{\link{pandaDiffEdges}} and \code{\link{visDiffPandaInCytoscape}},
and all edges belong to this condition or phenotype will be represented as purple, otherwise, green.}
}
\value{
a visual style in Cytoscape Control Panel under "Style" button.
}
\description{
This function is able to create a Cytoscape visual style for two PANDA networks output.
}
