% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPandaStyle.R
\name{createPandaStyle}
\alias{createPandaStyle}
\title{Create a Cytoscape visual style for PANDA network}
\usage{
createPandaStyle(style.name = "PandaStyle")
}
\arguments{
\item{style.name}{Character string indicating the style name. Defaults to "PandaStyle"}
}
\value{
a visual style in Cytoscape Control Panel under "Style" button.
}
\description{
This function is able to create a Cytoscape visual style for any PANDA network output.
}
