% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIDER.R
\name{degreeAdjust}
\alias{degreeAdjust}
\title{Function to adjust the degree so that the hub nodes are not penalized in z-score transformation}
\usage{
degreeAdjust(A)
}
\arguments{
\item{A}{Input adjacency matrix}
}
\description{
Function to adjust the degree so that the hub nodes are not penalized in z-score transformation
}
