% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPACAFunctions.R
\name{DeltaZAnalysis.Louvain}
\alias{DeltaZAnalysis.Louvain}
\title{Edge subtraction method (Louvain optimizaton)}
\usage{
DeltaZAnalysis.Louvain(net.table, file.stem)
}
\arguments{
\item{net.table}{A table of edges, with the first column representing the TFs ("from" nodes) and the second column representing the targets ("to" nodes). The third column contains the edge weights corresponding to the control or healthy network, and the fourth column contains the edge weights for the disease network or network of interest.}

\item{file.stem}{The folder location and title under which all results will be stored.}
}
\value{
List where first element is the membership vector and second element is the contribution score of each node to its community's modularity in the final edge-subtracted network
}
\description{
Takes two networks, subtracts edges individually, and then clusters the subtracted network using Louvain method.
}
\examples{
DeltaZAnalysis.Louvain()
}
