% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desparsify.R
\name{desparsify}
\alias{desparsify}
\title{De-sparsify gene-level mutation scores into gene set-level mutation scores.}
\usage{
desparsify(edgx, mutratecorx, ...)
}
\arguments{
\item{edgx}{A binary matrix containing information on which genes belong to which gene sets. Output from the convertgmt function.}

\item{mutratecorx}{Gene-level mutation scores corrected for the number of gene sets each gene belongs to (from sambar function).}
}
\value{
De-sparsified mutation data.
}
\description{
De-sparsify gene-level mutation scores into gene set-level mutation scores.
}
