% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TARDIGRADE.R
\name{diffuseHeat}
\alias{diffuseHeat}
\title{Performs heat diffusion for a single network, given the knowledge graph.
This happens by multiplying the seed values in a network by the conductance
matrix, which is the inverse of a negative normalized graph Laplacian.}
\usage{
diffuseHeat(knowledgeGraphs = list(), scaledAnalyteMeasurements = list())
}
\arguments{
\item{knowledgeGraphs}{An list of TARDIGRADE_KnowledgeGraphs, one modified
graph for each sample.}

\item{scaledAnalyteMeasurements}{A list of analyte measurement matrices (e.g. DNA methylation,
gene expression, metabolite abundance, protein abundance), with each analyte
scaled between 0 and 1.}
}
\description{
Performs heat diffusion for a single network, given the knowledge graph.
This happens by multiplying the seed values in a network by the conductance
matrix, which is the inverse of a negative normalized graph Laplacian.
}
