% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TARDIGRADE.R
\name{doPathwayAnalysis}
\alias{doPathwayAnalysis}
\title{Performs pathway cluster enrichment analysis using a GSEA-like procedure.
Specifically, given the heat values of each node for each sample for 
both original and permuted data, (1) Find differential heat score for
each node and then rank them, (2) Find the composite ranking score for each
pathway cluster, and (3) Find the significance of that score.}
\usage{
doPathwayAnalysis(
  knowledgeGraphsControlsHeat = list(),
  knowledgeGraphsCasesHeat = list(),
  pvalue = 1,
  permCount = 1000
)
}
\arguments{
\item{knowledgeGraphsControlsHeat}{A list of heat diffusion results for all
samples in the control group.}

\item{knowledgeGraphsCasesHeat}{A list of heat diffusion results for all samples
in the case group.}

\item{pvalue}{FDR-adjusted p-value cutoff for significance.}

\item{permCount}{Number of permutations.}
}
\value{
A data frame containing p-values for each pathway cluster.
}
\description{
Performs pathway cluster enrichment analysis using a GSEA-like procedure.
Specifically, given the heat values of each node for each sample for 
both original and permuted data, (1) Find differential heat score for
each node and then rank them, (2) Find the composite ranking score for each
pathway cluster, and (3) Find the significance of that score.
}
