% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TARDIGRADE.R
\name{DoRaMPSetup}
\alias{DoRaMPSetup}
\title{Implements the setup for RaMP.}
\usage{
DoRaMPSetup(version = NULL)
}
\arguments{
\item{version}{The version to use for setup.}
}
\value{
A data frame mapping from Analyte (rampId:common_name) to Pathway
(type:pathwayName)
}
\description{
Implements the setup for RaMP.
}
