% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIGER.R
\name{.get_cmdstanr_fun}
\alias{.get_cmdstanr_fun}
\title{This function exists as a workaround for calling packages from cmdstanr without
triggering package warnings. We cannot use cmdstanr::function() because cmdstanr
is not included in Suggests under the DESCRIPTION file. It cannot be included
under Suggests because it is not supported by Bioconductor.}
\usage{
.get_cmdstanr_fun(fun)
}
\description{
This function exists as a workaround for calling packages from cmdstanr without
triggering package warnings. We cannot use cmdstanr::function() because cmdstanr
is not included in Suggests under the DESCRIPTION file. It cannot be included
under Suggests because it is not supported by Bioconductor.
}
\keyword{internal}
