% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{DownloadFERRETData}
\alias{DownloadFERRETData}
\title{Download the data to use in FERRET. This method downloads data from an in-built URL.
Users must specify the data set of interest and a path.}
\usage{
DownloadFERRETData(datasetName, destinationDir)
}
\arguments{
\item{datasetName}{One of "CPTAC_RCC" or "CPTAC_GBM".}

\item{destinationDir}{A directory where the data should be saved.}
}
\description{
Download the data to use in FERRET. This method downloads data from an in-built URL.
Users must specify the data set of interest and a path.
}
