% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{dTFIPlot}
\alias{dTFIPlot}
\title{This function plots the Off diagonal mass of an 
observed Transition Matrix compared to a set of null TMs}
\usage{
dTFIPlot(monsterObj, rescale = FALSE, plot.title = NA,
  highlight.tfs = NA)
}
\arguments{
\item{monsterObj}{monsterAnalysis Object}

\item{rescale}{logical indicating whether to reorder transcription
factors according to their statistical significance and to 
rescale the values observed to be standardized by the null
distribution}

\item{plot.title}{String specifying the plot title}

\item{highlight.tfs}{vector specifying a set of transcription 
factors to highlight in the plot}
}
\value{
ggplot2 object for transition matrix comparing observed 
distribution to that estimated under the null
}
\description{
This function plots the Off diagonal mass of an 
observed Transition Matrix compared to a set of null TMs
}
\examples{
# data(yeast)
# design <- c(rep(0,20),rep(NA,10),rep(1,20))
# monsterRes <- monster(yeast$exp.cc, design, yeast$motif, nullPerms=100, numMaxCores=4)#' 
data(monsterRes)
dTFIPlot(monsterRes)
}
