% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIGER.R
\name{el2adj}
\alias{el2adj}
\title{Convert bipartite edge list to adjacency mat}
\usage{
el2adj(el)
}
\arguments{
\item{el}{An edge list dataframe with three columns. First column is TF name,
second column is gene name, and third column is edge weight.}
}
\value{
An adjacency matrix with rows as TFs and columns as genes.
}
\description{
Convert bipartite edge list to adjacency mat
}
