% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANE.R
\name{elistAddTags}
\alias{elistAddTags}
\title{Adds "_A" to first column and "_B" to second column}
\usage{
elistAddTags(elist)
}
\arguments{
\item{elist}{edge list}
}
\value{
edge list
}
\description{
Adds "_A" to first column and "_B" to second column
}
