% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANE.R
\name{elistIsEdgeOrderEqual}
\alias{elistIsEdgeOrderEqual}
\title{check if first two columns are identical}
\usage{
elistIsEdgeOrderEqual(elist1, elist2)
}
\arguments{
\item{elist1}{edge list}

\item{elist2}{edge list}
}
\value{
boolean
}
\description{
check if first two columns are identical
}
