% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANE.R
\name{elistSort}
\alias{elistSort}
\title{Sorts the edge list based on first two columns in alphabetical order}
\usage{
elistSort(elist)
}
\arguments{
\item{elist}{edge list}
}
\value{
edge list
}
\description{
Sorts the edge list based on first two columns in alphabetical order
}
