% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANE.R
\name{elistToAdjMat}
\alias{elistToAdjMat}
\title{Converts edge list to adjacency matrix}
\usage{
elistToAdjMat(elist, isBipartite = F)
}
\arguments{
\item{elist}{edge list}

\item{isBipartite}{TRUE = for bipartite / FALSE = for unipartite}
}
\value{
Adjcency Matrix
}
\description{
Converts edge list to adjacency matrix
}
