% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLOBFISH.R
\name{evaluateSubnetworkAgainstNull}
\alias{evaluateSubnetworkAgainstNull}
\title{Calculates and outputs empirical p-values for a single network based on null models.}
\usage{
evaluateSubnetworkAgainstNull(
  file,
  network,
  logging = TRUE,
  inputType = "matrix",
  outputType = "matrix",
  n,
  pvalFile,
  alpha = 0.05,
  adjustmentMethod = "none"
)
}
\arguments{
\item{file}{The file containing the null networks}

\item{network}{The input network, either a matrix with TFs in rows and genes in columns or
a list with TF names in the first column, gene names in the second, and scores in the third.}

\item{logging}{Whether or not to print logging messages (default is TRUE)}

\item{inputType}{"matrix" or "list"}

\item{outputType}{"matrix or "list"}

\item{n}{Number of null networks to include in your analysis. Must be less than the
number of networks in the file.}

\item{pvalFile}{File where you wish to store the calculated p-values. Must be RDS,
and the directory must exist.}

\item{alpha}{p-value cutoff for binarization.}

\item{adjustmentMethod}{Method for p-value adjustment. Default is "none".}
}
\description{
Calculates and outputs empirical p-values for a single network based on null models.
}
