% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{ExpandNetworks}
\alias{ExpandNetworks}
\title{Helper function to expand the networks to include zero-weighted edges.}
\usage{
ExpandNetworks(network1, network2)
}
\arguments{
\item{network1}{A network for comparison.}

\item{network2}{Another network for comparison.}
}
\value{
A list of two expanded networks.
}
\description{
Helper function to expand the networks to include zero-weighted edges.
}
