% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{exportNetwork}
\alias{exportNetwork}
\title{Network exporter}
\usage{
exportNetwork(x, file, format = "ncol")
}
\arguments{
\item{x}{An object of class "panda" or an incidence network matrix.}

\item{file}{A connection or a string giving the file name to output the network.}

\item{format}{Character string giving the file format supported by igraph.}
}
\description{
exportNetwork writes the panda network (RegNet) to foreign file formats.
}
\examples{
\donttest{
data(pandaToyData)
pandaRes <- panda(pandaToyData$motif, pandaToyData$expression, pandaToyData$ppi)
topPandaRes <- topedges(pandaRes, 1000)
exportNetwork(topPandaRes, 'topPandaRes.txt')
}
}
\keyword{keywords}

