% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPACAFunctions.R
\name{ExtractTopGenes}
\alias{ExtractTopGenes}
\title{Extract core target genes in differential modules}
\usage{
ExtractTopGenes(module.result, set.lengths)
}
\arguments{
\item{module.result}{A table of edges, with the first column representing the TFs ("from" nodes) and the second column representing the targets ("to" nodes). The third column contains the edge weights corresponding to the control or healthy network, and the fourth column contains the edge weights for the disease network or network of interest.}

\item{set.lengths}{The desired lengths of the top gene lists.}
}
\value{
List with two elements. First element is a list of the top target genes in each cluster. Second element is a vector with the names of the gene sets. The names are in the format "number_length", where number is the module number label and length is the length of the gene set.
}
\description{
This function outputs the top target genes in each module, ranked by their contribution to the differential modularity of the particular module in which they belong.
}
\examples{
ExtractTopGenes()
}
