% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\docType{class}
\name{FERRET_Comparison-class}
\alias{FERRET_Comparison-class}
\title{FERRET_Comparison class
 This object type represents a single comparison as a source-target pair.
 
 @name FERRET_Comparison-class
 @rdname FERRET_Comparison-class
 @exportClass FERRET_Comparison
 @slot source The name of a source network as represented in a FERRET_Results object.
 @slot target The name of a target network as represented in a FERRET_Results object.}
\description{
FERRET_Comparison class
 This object type represents a single comparison as a source-target pair.
 
 @name FERRET_Comparison-class
 @rdname FERRET_Comparison-class
 @exportClass FERRET_Comparison
 @slot source The name of a source network as represented in a FERRET_Results object.
 @slot target The name of a target network as represented in a FERRET_Results object.
}
