% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\docType{class}
\name{FERRET_Comparisons-class}
\alias{FERRET_Comparisons-class}
\title{FERRET_Comparisons class
 This object type represents the list of comparisons to make for the ingroup
 and the outgroup.
 
 @name FERRET_Comparisons-class
 @rdname FERRET_Comparisons-class
 @exportClass FERRET_Comparisons
 @slot ingroup A list of FERRET_Comparison objects
 @slot outgroup A list of FERRET_Comparison objects}
\description{
FERRET_Comparisons class
 This object type represents the list of comparisons to make for the ingroup
 and the outgroup.
 
 @name FERRET_Comparisons-class
 @rdname FERRET_Comparisons-class
 @exportClass FERRET_Comparisons
 @slot ingroup A list of FERRET_Comparison objects
 @slot outgroup A list of FERRET_Comparison objects
}
