% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\docType{class}
\name{FERRET_Results-class}
\alias{FERRET_Results-class}
\title{FERRET_Results class
 This object type represents the list of results in a given directory.
 
 @name FERRET_Results-class
 @rdname FERRET_Results-class
 @exportClass FERRET_Results
 @slot results A named list of inferred networks, where each network is represented as adjacency list.
 Names correspond to the input file.
 @slot directory The name of the source directory.
 @slot interpretationOfNegative Whether to interpret a negative weight as a regulatory
 relationship with poor evidence or as an inhibitory relationship.}
\description{
FERRET_Results class
 This object type represents the list of results in a given directory.
 
 @name FERRET_Results-class
 @rdname FERRET_Results-class
 @exportClass FERRET_Results
 @slot results A named list of inferred networks, where each network is represented as adjacency list.
 Names correspond to the input file.
 @slot directory The name of the source directory.
 @slot interpretationOfNegative Whether to interpret a negative weight as a regulatory
 relationship with poor evidence or as an inhibitory relationship.
}
