% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\docType{class}
\name{FERRET_ROC_AUC-class}
\alias{FERRET_ROC_AUC-class}
\title{FERRET_ROC_AUC
 This object type includes ROC and AUC scores.
 
 @name FERRET_ROC_AUC-class
 @rdname FERRET_ROC_AUC-class
 @exportClass FERRET_ROC_AUC
 @slot auc A scalar AUC score.
 @slot roc A named list of results, where the names are the types of metrics
 and each item in the list is an object of type FERRET_Similarities.}
\description{
FERRET_ROC_AUC
 This object type includes ROC and AUC scores.
 
 @name FERRET_ROC_AUC-class
 @rdname FERRET_ROC_AUC-class
 @exportClass FERRET_ROC_AUC
 @slot auc A scalar AUC score.
 @slot roc A named list of results, where the names are the types of metrics
 and each item in the list is an object of type FERRET_Similarities.
}
