% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\docType{class}
\name{FERRET_Similarities-class}
\alias{FERRET_Similarities-class}
\title{FERRET_Similarities class
 This object type represents the list of similarities across cutoffs for
 the ingroup and outgroup.
 
 @name FERRET_Similarities-class
 @rdname FERRET_Similarities-class
 @exportClass FERRET_Similarities
 @slot ingroup A vector of similarity values across cutoffs for the ingroup.
 @slot outgroup A vector of similarity values across cutoffs for the outgroup.}
\description{
FERRET_Similarities class
 This object type represents the list of similarities across cutoffs for
 the ingroup and outgroup.
 
 @name FERRET_Similarities-class
 @rdname FERRET_Similarities-class
 @exportClass FERRET_Similarities
 @slot ingroup A vector of similarity values across cutoffs for the ingroup.
 @slot outgroup A vector of similarity values across cutoffs for the outgroup.
}
