% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARN.R
\name{filterMissingGenes}
\alias{filterMissingGenes}
\title{Filter genes not expressed in any sample}
\usage{
filterMissingGenes(obj, threshold = 0)
}
\arguments{
\item{obj}{ExpressionSet object.}

\item{threshold}{Minimum sum of gene counts across samples -- defaults to zero.}
}
\value{
Filtered ExpressionSet object
}
\description{
The main use case for this function is the removal of missing genes.
}
\examples{
\donttest{
u <- 'https://netzoo.s3.us-east-2.amazonaws.com/netZooR/unittest_datasets/yarn/'
skin <- tryCatch({tmp<-tempfile(); download.file(paste0(u,'skin.rdata'),tmp,quiet=TRUE); local({load(tmp);get('skin')})}, error=function(e) NULL)
if(!is.null(skin)) filterMissingGenes(skin)
}

}
