% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARN.R
\name{filterSamples}
\alias{filterSamples}
\title{Filter samples}
\usage{
filterSamples(obj, ids, groups = colnames(obj), keepOnly = FALSE)
}
\arguments{
\item{obj}{ExpressionSet object.}

\item{ids}{Names found within the groups labels corresponding to samples to be removed}

\item{groups}{Vector of labels for each sample or a column name of the phenoData slot
for the ids to filter. Default is the column names.}

\item{keepOnly}{Filter or keep only the samples with those labels.}
}
\value{
Filtered ExpressionSet object
}
\description{
Filter samples
}
\examples{
\donttest{
u <- 'https://netzoo.s3.us-east-2.amazonaws.com/netZooR/unittest_datasets/yarn/'
skin <- tryCatch({tmp<-tempfile(); download.file(paste0(u,'skin.rdata'),tmp,quiet=TRUE); local({load(tmp);get('skin')})}, error=function(e) NULL)
if(!is.null(skin)) filterSamples(skin,ids="Skin - Not Sun Exposed (Suprapubic)",groups="SMTSD")
if(!is.null(skin)) filterSamples(skin,ids=c("GTEX-OHPL-0008-SM-4E3I9","GTEX-145MN-1526-SM-5SI9T"))
}

}
