% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TARDIGRADE.R
\name{FindClusters}
\alias{FindClusters}
\title{Finds clusters in the pathway knowledge graph using Louvain clustering.}
\usage{
FindClusters(knowledgeGraph, clusterCount)
}
\arguments{
\item{knowledgeGraph}{An object of type TARDIGRADE_KnowledgeGraph.}

\item{clusterCount}{The number of clusters.}
}
\value{
An object of type TARDIGRADE_KnowledgeGraph.
}
\description{
Finds clusters in the pathway knowledge graph using Louvain clustering.
}
