% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLOBFISH.R
\name{FindConnectionsForAllHopCounts}
\alias{FindConnectionsForAllHopCounts}
\title{For all hop counts up to the maximum, find subnetworks connecting each pair of
genes by exactly that number of hops. For instance, find each}
\usage{
FindConnectionsForAllHopCounts(subnetworks, verbose = FALSE)
}
\arguments{
\item{subnetworks}{A list of bipartite (PANDA-like) subnetworks for each gene, 
containing only the significant edges meeting the hop count criteria and
where each network is a data frame with the following format:
tf,gene}

\item{verbose}{Whether or not to print detailed information about the run.}
}
\value{
A bipartite subnetwork in the same format as the original networks.
}
\description{
For all hop counts up to the maximum, find subnetworks connecting each pair of
genes by exactly that number of hops. For instance, find each
}
