% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNAGI.R
\name{FindConnectionsForAllHopCountsU}
\alias{FindConnectionsForAllHopCountsU}
\title{For all hop counts up to the maximum, find subnetworks connecting each pair of
genes by exactly that number of hops. For instance, find each
containing only the significant edges meeting the hop count criteria and
where each network is a data frame with the following format:}
\usage{
FindConnectionsForAllHopCountsU(subnetworks, odd = FALSE, verbose = FALSE)
}
\arguments{
\item{subnetworks}{The subnetworks, generated using FindEdgesForHopU().}

\item{odd}{Whether or not we are looking at an odd number of hops (in which case
we should subtract 1 from the hop count)}

\item{verbose}{Whether or not to print detailed information about the run.}
}
\description{
For all hop counts up to the maximum, find subnetworks connecting each pair of
genes by exactly that number of hops. For instance, find each
containing only the significant edges meeting the hop count criteria and
where each network is a data frame with the following format:
}
