% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNAGI.R
\name{FindEdgesForHopU}
\alias{FindEdgesForHopU}
\title{Find the subnetwork of significant edges n / 2 hops away from each gene.}
\usage{
FindEdgesForHopU(geneSet, network, hopConstraint, verbose = FALSE, topX = NULL)
}
\arguments{
\item{geneSet}{A character vector of genes comprising the targets of interest.}

\item{network}{A network with the first column ("source") consisting of source nodes,
the second ("target") consisting of target nodes, and the third ("score") consisting of scores}

\item{hopConstraint}{The maximum number of hops to be considered for a gene.}

\item{verbose}{Whether or not to print detailed information about the run.}

\item{topX}{Select the X lowest significant p-values for each gene. NULL by default.}
}
\description{
Find the subnetwork of significant edges n / 2 hops away from each gene.
}
