% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{formatTissueNames}
\alias{formatTissueNames}
\title{Formats the tissue names by removing the .RDS extension from the input files.}
\usage{
formatTissueNames(names)
}
\arguments{
\item{names}{The list of tissue names.}
}
\value{
A list of names.
}
\description{
Formats the tissue names by removing the .RDS extension from the input files.
}
