% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{GenerateNullFromControl}
\alias{GenerateNullFromControl}
\title{Generates null model for MONSTER by sampling "case" networks from a distribution
where the control network value is the mean and the provided null networks are
used to calculate the variance. Sampling is done on an edgewise basis, using
edgewise means and variances.}
\usage{
GenerateNullFromControl(concatNet, iterations, design, nullNetworks)
}
\arguments{
\item{concatNet}{The concatenated case and control networks}

\item{iterations}{Number of null models to generate}

\item{design}{The design matrix}

\item{nullNetworks}{The null networks from which to calculate the variance}
}
\value{
A list of matrices, each one corresponding to one null model
}
\description{
Generates null model for MONSTER by sampling "case" networks from a distribution
where the control network value is the mean and the provided null networks are
used to calculate the variance. Sampling is done on an edgewise basis, using
edgewise means and variances.
}
