% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLOBFISH.R
\name{GenerateNullPANDADistribution}
\alias{GenerateNullPANDADistribution}
\title{Generate a null distribution of edge scores for PANDA-like networks; that is,
the set of edges where (1) the TF does not have a binding motif in the gene region,
(2) the TF does not form a complex with any other TF that has a binding motif in
the gene region, and (3) the genes regulated by the TF are not coexpressed with the
gene in question. We obtain this by inputting an empty prior and an identity coexpression
matrix.}
\usage{
GenerateNullPANDADistribution(
  ppiFile,
  motifFile,
  sampSize = 20,
  numberOfPandas = 10,
  outputType = "vector"
)
}
\arguments{
\item{ppiFile}{The location of the protein-protein interaction network between transcription factors.
This should be a TSV file where the first two columns are the transcription
factors and the third is whether there is a PPI between them.}

\item{motifFile}{The location of the motif prior from genes to transcription factors. This should
be a TSV file where the first column is the transcription factors, the
second is the genes, and the third is whether the transcription factor's
binding motif is in the gene promoter region.}

\item{sampSize}{Number of samples to simulate}

\item{numberOfPandas}{Number of null PANDA networks to generate}

\item{outputType}{Type of output desired. Default is "vector", which returns a vector
of null edge weights. Another acceptable output is "network", which outputs an edge list
with the edge values across all null networks. Use "vector" if you wish to filter
edge weights agnostic of TF-gene pair and "network" if you wish to account for a
TF-gene pair's proximity to the prior.}
}
\description{
Generate a null distribution of edge scores for PANDA-like networks; that is,
the set of edges where (1) the TF does not have a binding motif in the gene region,
(2) the TF does not form a complex with any other TF that has a binding motif in
the gene region, and (3) the genes regulated by the TF are not coexpressed with the
gene in question. We obtain this by inputting an empty prior and an identity coexpression
matrix.
}
