% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{GetCommonNetwork}
\alias{GetCommonNetwork}
\title{This function finds the common network between the common networks for two
different scenarios (e.g., the ingroup and the outgroup).}
\usage{
GetCommonNetwork(results, comparisons)
}
\arguments{
\item{results}{An object of type FERRET_Results}

\item{comparisons}{A FERRET_Comparisons object.}
}
\value{
A single network in the same format as the result networks.
}
\description{
This function finds the common network between the common networks for two
different scenarios (e.g., the ingroup and the outgroup).
}
