% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{GetCommonNetworkAcrossComparisons}
\alias{GetCommonNetworkAcrossComparisons}
\title{This function finds the network common to all folds and the original network.
the weight of each edge is the minimum magnitude weight over all networks.}
\usage{
GetCommonNetworkAcrossComparisons(results, comparisons, useTargetsOnly = FALSE)
}
\arguments{
\item{results}{An object of type FERRET_Results}

\item{comparisons}{A list of FERRET_Comparisons objects. 
Note that the in-group and out-group names in FERRET_Comparisons MUST match the names
in FERRET_Results.}

\item{useTargetsOnly}{Whether to include the source network in the evaluation.}
}
\value{
A single network in the same format as the result networks.
}
\description{
This function finds the network common to all folds and the original network.
the weight of each edge is the minimum magnitude weight over all networks.
}
