% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNAGI.R
\name{GetConnectingSubnetworkHelper}
\alias{GetConnectingSubnetworkHelper}
\title{A helper function for obtaining the connecting subnetwork given two spheres
of influence. Called from FindConnectionsForAllHopCountsU().}
\usage{
GetConnectingSubnetworkHelper(
  subnetwork1,
  subnetwork2,
  gene1,
  gene2,
  hops,
  subnetworks,
  odd = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{subnetwork1}{The sphere of influence for the first gene in a pair.}

\item{subnetwork2}{The sphere of influence for the second gene in a pair.}

\item{gene1}{The first gene in a pair.}

\item{gene2}{The second gene in a pair.}

\item{hops}{The current hop.}

\item{subnetworks}{The full subnetworks list.}

\item{odd}{Whether or not the maximum number of hops is an odd number.}

\item{verbose}{Whether or not to print detailed information about the run.}
}
\description{
A helper function for obtaining the connecting subnetwork given two spheres
of influence. Called from FindConnectionsForAllHopCountsU().
}
