% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{GetDifferentialNetworks}
\alias{GetDifferentialNetworks}
\title{This function finds the differential network between the common networks for
two different scenarios (e.g., the ingroup and outgroup). We note that, in this
scenario, we only report edges where the magnitude in the ingroup network is larger
than in the outgroup network and both edges are in the same direction. 
That is, negative means that a more significant inhibitory edge is present in the
ingroup network than in the outgroup network. It does NOT mean that
the magnitude is greater in the outgroup network than in the ingroup network.}
\usage{
GetDifferentialNetworks(results, comparisons)
}
\arguments{
\item{results}{An object of type FERRET_Results}

\item{comparisons}{A list of FERRET_Comparisons object.}
}
\value{
A single network in the same format as the result networks.
}
\description{
This function finds the differential network between the common networks for
two different scenarios (e.g., the ingroup and outgroup). We note that, in this
scenario, we only report edges where the magnitude in the ingroup network is larger
than in the outgroup network and both edges are in the same direction. 
That is, negative means that a more significant inhibitory edge is present in the
ingroup network than in the outgroup network. It does NOT mean that
the magnitude is greater in the outgroup network than in the ingroup network.
}
