% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{GetEdgesAboveCutoff}
\alias{GetEdgesAboveCutoff}
\title{This function returns each network filtered to only include those edges with
scores above a specified cutoff.}
\usage{
GetEdgesAboveCutoff(network, cutoff)
}
\arguments{
\item{network}{The network to filter.}

\item{cutoff}{The cutoff to use when filtering the network.}
}
\value{
A network in the format of an adjacency list.
}
\description{
This function returns each network filtered to only include those edges with
scores above a specified cutoff.
}
