% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{GetFullNetworkFromPCANetwork}
\alias{GetFullNetworkFromPCANetwork}
\title{This function takes in a network of associations between PCs and reconstructs
a network of associations between genes using the PC loadings.}
\usage{
GetFullNetworkFromPCANetwork(pcNetwork, PC, zeroCutoff = 0.001)
}
\arguments{
\item{pcNetwork}{The network of associations between PCs, formatted as a list with
source, target, and score info.}

\item{PC}{The principal components, formatted as a matrix with rows equal to the number
of genes and columns equal to the PCs.}

\item{zeroCutoff}{The zero cutoff.}
}
\value{
A single network in the same format as the input network.
}
\description{
This function takes in a network of associations between PCs and reconstructs
a network of associations between genes using the PC loadings.
}
