% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{GetNegative}
\alias{GetNegative}
\title{Helper function to return a subnetwork with only negative edges.}
\usage{
GetNegative(results)
}
\arguments{
\item{results}{A FERRET_Results object.}
}
\value{
A new adjacency list with only the negative scores, converted to positive.
}
\description{
Helper function to return a subnetwork with only negative edges.
}
