% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TARDIGRADE.R
\name{getNormalizedEnrichmentScore}
\alias{getNormalizedEnrichmentScore}
\title{Obtains the normalized enrichment score for each pathway cluster based on
the node correlation rankings and the cluster definitions.}
\usage{
getNormalizedEnrichmentScore(
  nodeCorrelationRank = NULL,
  pathwayClusterMapping = NULL,
  pathwayClusterNames = NULL,
  pathwayClusterSizes = NULL
)
}
\arguments{
\item{nodeCorrelationRank}{A ranked vector of nodes (analytes) by correlation
with case/control status.}

\item{pathwayClusterMapping}{The mapping from gene to pathway cluster.}

\item{pathwayClusterNames}{A list of names for each pathway cluster.}

\item{pathwayClusterSizes}{The sizes of each pathway cluster.}
}
\value{
A normalized enrichment score for each pathway cluster.
}
\description{
Obtains the normalized enrichment score for each pathway cluster based on
the node correlation rankings and the cluster definitions.
}
