% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TARDIGRADE.R
\name{getNormalizedLaplacian}
\alias{getNormalizedLaplacian}
\title{Obtains the normalized Laplacian for a knowledge graph. This is a helper
function for diffuseHeat.}
\usage{
getNormalizedLaplacian(knowledgeGraph)
}
\arguments{
\item{knowledgeGraphs}{An TARDIGRADE_KnowledgeGraph.}
}
\value{
The normalized Laplacian as a matrix.
}
\description{
Obtains the normalized Laplacian for a knowledge graph. This is a helper
function for diffuseHeat.
}
