% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{GetOverlapWithPrior}
\alias{GetOverlapWithPrior}
\title{This function computes the percent weight of the input network that overlaps with the prior.}
\usage{
GetOverlapWithPrior(network, prior)
}
\arguments{
\item{network}{The input network, where the transcription factors in the first column 
are gene symbols.}

\item{prior}{The prior}
}
\value{
The Jaccard index (scalar)
}
\description{
This function computes the percent weight of the input network that overlaps with the prior.
}
