% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{GetPositive}
\alias{GetPositive}
\title{Helper function to return a subnetwork with only positive edges.}
\usage{
GetPositive(results)
}
\arguments{
\item{results}{A FERRET_Results object.}
}
\value{
A new adjacency list with only the positive scores.
}
\description{
Helper function to return a subnetwork with only positive edges.
}
