% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPACAFunctions.R
\name{go.to.genes}
\alias{go.to.genes}
\title{Map GO terms to gene symbols}
\usage{
go.to.genes(go.term)
}
\arguments{
\item{go.term}{The GO Biological Process ID (string).}
}
\value{
A vector of all gene symbols associated with the GO term.
}
\description{
This function extracts all the gene symbols associated with a GO term and its descendants. (v1)
}
\examples{
go.to.genes()
}
