% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{gsea_numeric}
\alias{gsea_numeric}
\title{Function to run GSEA for a numeric phenotype}
\usage{
gsea_numeric(expression, pheno, pathways)
}
\arguments{
\item{expression}{: gene expression matrix (normalized, and filtered) 
with rows as genes and columns as samples.
Row and column names must be present.
Row names must be HGNC symbols.
Column names must match the row names of the phenotype matrix.}

\item{pheno}{: phenotype matrix
with rows as samples and columns as phenotype variables.}

\item{pathways}{: a list of pathways (e.g. KEGG, GO, Reactome etc. 
downloaded from http://www.gsea-msigdb.org/gsea/msigdb/human/collections.jsp)}
}
\description{
Function to run GSEA for a numeric phenotype
}
