% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{JaccardSim}
\alias{JaccardSim}
\title{Helper function to compute the Jaccard similarities between two networks.
This is a weighted version of the Jaccard similarity; i.e.,}
\usage{
JaccardSim(network1, network2)
}
\arguments{
\item{network1}{A network for comparison.}

\item{network2}{Another network for comparison.}
}
\value{
The similarity value.
}
\description{
Helper function to compute the Jaccard similarities between two networks.
This is a weighted version of the Jaccard similarity; i.e.,
}
