% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANE.R
\name{jutterDegree}
\alias{jutterDegree}
\title{CRANE Beta perturbation function. This function will add noice to the node strength sequence.}
\usage{
jutterDegree(nodeD, beta, beta_slope = T)
}
\arguments{
\item{nodeD}{Vector of node strength}

\item{beta}{beta}

\item{beta_slope}{TRUE=use predetermined slope to add noise / FALSE = use constant value for noise}
}
\value{
vector with new strength distribution
}
\description{
CRANE Beta perturbation function. This function will add noice to the node strength sequence.
}
