% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{LoadResults}
\alias{LoadResults}
\title{This function loads all results from a given directory and stores them in
a FERRET_Results object. Results are expected to be in an adjacency list
format, where the first column is the list transcription factors regulating genes,
the second column is the list of regulated genes, and the third column is the
score for each regulatory relationship.}
\usage{
LoadResults(
  resultDirectory,
  interpretationOfNegative = "poor",
  firstColumnIsRowname = TRUE
)
}
\arguments{
\item{resultDirectory}{The name of the directory.}

\item{interpretationOfNegative}{Negatives can refer to either poor associations ("poor")
or inhibitory ("inhibitory") regulation, depending on the meaning.}

\item{firstColumnIsRowname}{Whether to consider the first column as the list of
row names.}
}
\value{
An object of type FERRET_Results.
}
\description{
This function loads all results from a given directory and stores them in
a FERRET_Results object. Results are expected to be in an adjacency list
format, where the first column is the list transcription factors regulating genes,
the second column is the list of regulated genes, and the third column is the
score for each regulatory relationship.
}
