% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPACAFunctions.R
\name{metanetwork}
\alias{metanetwork}
\title{Create metanetwork for Louvain optimization}
\usage{
metanetwork(J, S)
}
\arguments{
\item{J}{The modularity matrix}

\item{S}{The community membership vector from the previous round of agglomeration.}
}
\value{
The differential modularity matrix, with rows representing "from" nodes and columns representing "to" nodes.
}
\description{
Computes the "effective" adjacency matrix of a metanetwork whose nodes represent communities in the larger input matrix.
}
\examples{
metanetwork()
}
