% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TARDIGRADE.R
\name{ModifyGraph}
\alias{ModifyGraph}
\title{Modifies the edges in the graph given a data-driven network, with the following rules:
- If an edge exists in both the knowledge graph and the data-driven network, increase
  the edge weight to 1 and make no other changes.
- If an edge exists in the knowledge graph and CANNOT be present in the data-driven
  network (i.e. the nodes do not exist or are not of the correct analyte types to be
  connected), make no changes.
- If an edge exists in the knowledge graph and COULD be present in the data-driven network
  (but isn't), remove it.
- If an edge does not exist in the knowledge graph but does exist in the data-driven
  network, add it with no pathways to the dict.}
\usage{
ModifyGraph(
  knowledgeGraph = NULL,
  network = NULL,
  sourceAnalytes = NULL,
  targetAnalytes = NULL
)
}
\arguments{
\item{knowledgeGraph}{An object of type TARDIGRADE_KnowledgeGraph.}

\item{network}{A data frame representing binary edges in a network.}

\item{sourceAnalytes}{The list of source analytes (e.g., TFs) evaluated.}

\item{targetAnalytes}{The list of target analytes (e.g., genes) evaluated.}
}
\value{
An adjacency matrix.
}
\description{
Modifies the edges in the graph given a data-driven network, with the following rules:
- If an edge exists in both the knowledge graph and the data-driven network, increase
  the edge weight to 1 and make no other changes.
- If an edge exists in the knowledge graph and CANNOT be present in the data-driven
  network (i.e. the nodes do not exist or are not of the correct analyte types to be
  connected), make no changes.
- If an edge exists in the knowledge graph and COULD be present in the data-driven network
  (but isn't), remove it.
- If an edge does not exist in the knowledge graph but does exist in the data-driven
  network, add it with no pathways to the dict.
}
