\name{MONSTER-package}
\alias{MONSTER-package}
\alias{MONSTER}
\docType{package}
\title{
\packageTitle{MONSTER}
}
\description{
\packageDescription{MONSTER}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MONSTER}
\packageIndices{MONSTER}

MONSTER takes in sequence motif data linking transcription factors (TFs) to genes and gene expression from two conditions. The goal is generate bipartite networks from the gene expression data which quantify evidence of the regulatory roles of each of the TFs to each of the genes. Next, critical TFs are identified by computing a transition matrix, which maps the gene regulatory network in the first state to the gene regulatory network in the second state.

}
\author{
\packageAuthor{MONSTER}

Maintainer: \packageMaintainer{MONSTER}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }

\examples{
##
}
