% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monster.checkDataType}
\alias{monster.checkDataType}
\title{Checks that data is something MONSTER can handle}
\usage{
monster.checkDataType(expr)
}
\arguments{
\item{expr}{Gene Expression dataset}
}
\value{
expr Gene Expression dataset in the proper form (may be the same as input)
}
\description{
Checks that data is something MONSTER can handle
}
\examples{
expr.matrix <- matrix(rnorm(2000),ncol=20)
monster.checkDataType(expr.matrix)
#TRUE
data(yeast)
class(yeast$exp.cc)
monster.checkDataType(yeast$exp.cc)
#TRUE

}
