% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monster.hcl.heatmap.plot}
\alias{monster.hcl.heatmap.plot}
\title{Transformation matrix plot}
\usage{
monster.hcl.heatmap.plot(monsterObj, method = "pearson")
}
\arguments{
\item{monsterObj}{monsterAnalysis Object}

\item{method}{distance metric for hierarchical clustering.
Default is "Pearson correlation"}
}
\value{
ggplot2 object for transition matrix heatmap
}
\description{
This function plots a hierachically clustered heatmap and 
corresponding dendrogram of a transaction matrix
}
\examples{
# data(yeast)
# design <- c(rep(0,20),rep(NA,10),rep(1,20))
# monsterRes <- monster(yeast$exp.cc, design, yeast$motif, nullPerms=100, numMaxCores=4)
data(monsterRes)
monster.hcl.heatmap.plot(monsterRes)
}
