% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monster.monsterNI}
\alias{monster.monsterNI}
\title{Bipartite Edge Reconstruction from Expression data}
\usage{
monster.monsterNI(motif, expr.data, verbose = FALSE,
  randomize = "none", method = "bere", ni.coefficient.cutoff = NA,
  alphaw = 1, regularization = "none", score = "motifincluded",
  cpp = FALSE)
}
\arguments{
\item{motif}{A motif dataset, a data.frame, matrix or exprSet containing 
3 columns. Each row describes an motif associated with a transcription 
factor (column 1) a gene (column 2) and a score (column 3) for the motif.}

\item{expr.data}{An expression dataset, as a genes (rows) by samples (columns)}

\item{verbose}{logical to indicate printing of output for algorithm progress.}

\item{randomize}{logical indicating randomization by genes, within genes or none}

\item{method}{String to indicate algorithm method.  Must be one of 
"bere","pearson","cd","lda", or "wcd". Default is "bere"}

\item{ni.coefficient.cutoff}{numeric to specify a p-value cutoff at the network
inference step.  Default is NA, indicating inclusion of all coefficients.}

\item{alphaw}{A weight parameter specifying proportion of weight 
to give to indirect compared to direct evidence.  See documentation.}

\item{regularization}{String parameter indicating one of "none", "L1", "L2"}

\item{score}{String to indicate whether motif information will be 
readded upon completion of the algorithm}

\item{cpp}{logical use C++ for maximum speed, set to false if unable to run.}
}
\value{
matrix for inferred network between TFs and genes
}
\description{
This function generates a complete bipartite network from 
gene expression data and sequence motif data
}
\examples{
data(yeast)
cc.net <- monster.monsterNI(yeast$motif,yeast$exp.cc)
}
