% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monster.transformation.matrix}
\alias{monster.transformation.matrix}
\title{Bi-partite network analysis tools}
\usage{
monster.transformation.matrix(network.1, network.2, by.tfs = TRUE,
  standardize = FALSE, remove.diagonal = TRUE, method = "ols")
}
\arguments{
\item{network.1}{starting network, a genes by transcription factors data.frame with scores 
for the existence of edges between}

\item{network.2}{final network, a genes by transcription factors data.frame with scores 
for the existence of edges between}

\item{by.tfs}{logical indicating a transcription factor based transformation.    If 
false, gives gene by gene transformation matrix}

\item{standardize}{logical indicating whether to standardize the rows and columns}

\item{remove.diagonal}{logical for returning a result containing 0s across the diagonal}

\item{method}{character specifying which algorithm to use, default='ols'}
}
\value{
matrix object corresponding to transition matrix
}
\description{
This function analyzes a bi-partite network.
}
\examples{
\donttest{
data(yeast)
cc.net.1 <- monster.monsterNI(yeast$motif,yeast$exp.cc[1:1000,1:20])
cc.net.2 <- monster.monsterNI(yeast$motif,yeast$exp.cc[1:1000,31:50])
monster.transformation.matrix(cc.net.1, cc.net.2)
}
}
