% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monster.transitionNetworkPlot}
\alias{monster.transitionNetworkPlot}
\title{This function uses igraph to plot the transition matrix as a network}
\usage{
monster.transitionNetworkPlot(monsterObj, numEdges = 100, numTopTFs = 10)
}
\arguments{
\item{monsterObj}{monsterAnalysis Object}

\item{numEdges}{The number of edges to display}

\item{numTopTFs}{The number of TFs to display, ranked by largest dTFI}
}
\value{
igraph object for transition matrix
}
\description{
This function uses igraph to plot the transition matrix as a network
}
\examples{
# data(yeast)
# yeast$exp.cc[is.na(yeast$exp.cc)] <- mean(as.matrix(yeast$exp.cc),na.rm=T)
# design <- c(rep(0,20),rep(NA,10),rep(1,20))
# monsterRes <- monster(yeast$exp.cc, design, yeast$motif, nullPerms=100, numMaxCores=4)#' 
data(monsterRes)
monster.transitionNetworkPlot(monsterRes)

}
