\name{monsterAnalysis}
\docType{class}
\alias{monsterAnalysis-class}

\title{Class "monsterAnalysis" -- a list containing the transition matrix results from a monster analysis}
\description{This is the main class for MONSTER.}
\section{Extends}{
    Class \code{list}
}
\note{ 
    Note: This is a summary for reference. For an explanation of the actual usage, see the vignette.
    
    'monsterAnalysis' obj are the main class in use by the MONSTER package. The class extends \code{\link{list}} and provides the key results from the results of the MONSTER analysis.
    TODO: more...
    

}

\section{Methods}{
    Class-specific methods.
    TODO: more...
}

\examples{
    # See vignette
}