% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monsterCalculateTmStatsPerCell}
\alias{monsterCalculateTmStatsPerCell}
\title{Calculate statistics for a transformation matrix at the matrix cell level (TF_a -> TF_b)}
\usage{
monsterCalculateTmStatsPerCell(monsterObj, method = "z-score")
}
\arguments{
\item{monsterObj}{monsterAnalysis Object}

\item{method}{one of 'z-score' or 'non-parametric'}
}
\value{
p-values, adjusted p-values, z-scores (if applicable)
}
\description{
This function powers both the p-value and t-value (or z-score) calculations
for each cell in a transformation matrix.
}
