% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monsterdTFIPlot}
\alias{monsterdTFIPlot}
\title{This function plots the Off diagonal mass of an 
observed Transition Matrix compared to a set of null TMs}
\usage{
monsterdTFIPlot(
  monsterObj,
  rescale = "none",
  plot.title = NA,
  highlight.tfs = NA,
  nTFs = -1
)
}
\arguments{
\item{monsterObj}{monsterAnalysis Object}

\item{rescale}{string indicating whether to reorder transcription
factors according to their statistical significance and to 
rescale the values observed to be standardized by the null
distribution ('significance'), to reorder transcription
factors according to the largest dTFIs ('magnitude') with the TF x axis labels proportional to their significance
, or finally without ordering them ('none'). When rescale is set to 'significance', 
the results can be different between two MONSTER runs if the number of permutations is not large enough to sample 
the null, that is why it is the seed should be set prior to calling MONSTER to get reproducible results. 
If rescale is set to another value such as 'magnitude' or 'none', it will produce deterministic results 
between two identical MONSTER runs.}

\item{plot.title}{String specifying the plot title}

\item{highlight.tfs}{vector specifying a set of transcription 
factors to highlight in the plot}

\item{nTFs}{number of TFs to plot in x axis. -1 takes all TFs.}
}
\value{
ggplot2 object for transition matrix comparing observed 
distribution to that estimated under the null
}
\description{
This function plots the Off diagonal mass of an 
observed Transition Matrix compared to a set of null TMs
}
\examples{
# data(yeast)
# yeast$exp.cc[is.na(yeast$exp.cc)] <- mean(as.matrix(yeast$exp.cc),na.rm=TRUE)
# design <- c(rep(0,20),rep(NA,10),rep(1,20))
# monsterRes <- monster(yeast$exp.cc, design, yeast$motif, nullPerms=100, numMaxCores=4)#' 
data(monsterRes)
slot(monsterRes, "logging") <- FALSE
monsterdTFIPlot(monsterRes)
}
