% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monsterGetTm}
\alias{monsterGetTm}
\title{monsterGetTm}
\usage{
monsterGetTm(x)
}
\arguments{
\item{x}{an object of class "monsterAnalysis"}
}
\value{
Transition matrix
}
\description{
acessor for the transition matrix in MONSTER object
}
\examples{
data(monsterRes)
tm <- monsterGetTm(monsterRes)
}
