% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monsterMonsterNI}
\alias{monsterMonsterNI}
\title{Bipartite Edge Reconstruction from Expression data}
\usage{
monsterMonsterNI(
  motif.data,
  expr.data,
  verbose = FALSE,
  randomize = "none",
  method = "BERE",
  ni.coefficient.cutoff = NA,
  alphaw = 1,
  regularization = "none",
  score = "motifincluded",
  cpp = FALSE,
  logging = TRUE
)
}
\arguments{
\item{motif.data}{A motif dataset, a data.frame, matrix or exprSet containing 
3 columns. Each row describes an motif associated with a transcription 
factor (column 1) a gene (column 2) and a score (column 3) for the motif.}

\item{expr.data}{An expression dataset, as a genes (rows) by samples (columns)}

\item{verbose}{logical to indicate printing of output for algorithm progress.}

\item{randomize}{logical indicating randomization by genes, within genes or none}

\item{method}{String to indicate algorithm method.  Must be one of 
"BERE","pearson",or "lda". Default is "BERE".
Important note: the direct regulatory network observed from gene expression is currently
implemented as a regular correlation as opposed to the partial correlation described 
in the paper (please see Schlauch et al., 2017,  https://doi.org/10.1186/s12918-017-0517-y)}

\item{ni.coefficient.cutoff}{numeric to specify a p-value cutoff at the network
inference step.  Default is NA, indicating inclusion of all coefficients.}

\item{alphaw}{A weight parameter between 0 and 1 specifying proportion of weight 
to give to indirect compared to direct evidence. The default is 0.5 to give an 
equal weight to direct and indirect evidence.}

\item{regularization}{String parameter indicating one of "none", "L1", "L2"}

\item{score}{String to indicate whether motif information will be 
readded upon completion of the algorithm
to give to indirect compared to direct evidence.  See documentation.}

\item{cpp}{logical use C++ for maximum speed, set to false if unable to run.}

\item{logging}{Whether or not to log progress (default = TRUE)}
}
\value{
matrix for inferred network between TFs and genes
}
\description{
This function generates a complete bipartite network from 
gene expression data and sequence motif data
}
\examples{
data(yeast)
cc.net <- monsterMonsterNI(yeast$motif,yeast$exp.cc)
}
