% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monsterPrintMonsterAnalysis}
\alias{monsterPrintMonsterAnalysis}
\title{monsterPrintMonsterAnalysis}
\usage{
monsterPrintMonsterAnalysis(x, ...)
}
\arguments{
\item{x}{an object of class "monster"}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Description of transition matrices in object
}
\description{
summarizes the results of a MONSTER analysis
}
\examples{
data(yeast)
yeast$exp.cc[is.na(yeast$exp.cc)] <- mean(as.matrix(yeast$exp.cc),na.rm=TRUE)
design <- c(rep(1,25),rep(0,10),rep(NA,15))
#monster(yeast$exp.cc,design,yeast$motif, nullPerms=10, numMaxCores=1)
}
