% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monsterTransformationMatrix}
\alias{monsterTransformationMatrix}
\title{Bi-partite network analysis tools}
\usage{
monsterTransformationMatrix(
  network.1,
  network.2,
  by.tfs = TRUE,
  standardize = FALSE,
  remove.diagonal = TRUE,
  method = "ols",
  logging = TRUE
)
}
\arguments{
\item{network.1}{starting network, a genes by transcription factors data.frame with scores 
for the existence of edges between}

\item{network.2}{final network, a genes by transcription factors data.frame with scores 
for the existence of edges between}

\item{by.tfs}{logical indicating a transcription factor based transformation.    If 
false, gives gene by gene transformation matrix}

\item{standardize}{logical indicating whether to standardize the rows and columns}

\item{remove.diagonal}{logical for returning a result containing 0s across the diagonal}

\item{method}{character specifying which algorithm to use, default='ols'}

\item{logging}{Whether or not to log progress (default = TRUE)}
}
\value{
matrix object corresponding to transition matrix
}
\description{
This function analyzes a bi-partite network.
}
\examples{
data(yeast)
cc.net.1 <- monsterMonsterNI(yeast$motif,yeast$exp.cc[1:1000,1:20])
cc.net.2 <- monsterMonsterNI(yeast$motif,yeast$exp.cc[1:1000,31:50])
monsterTransformationMatrix(cc.net.1, cc.net.2)
}
