% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monsterTransitionPCAPlot}
\alias{monsterTransitionPCAPlot}
\title{Principal Components plot of transformation matrix}
\usage{
monsterTransitionPCAPlot(
  monsterObj,
  title = "PCA Plot of Transition",
  clusters = 1,
  alpha = 1
)
}
\arguments{
\item{monsterObj}{a monsterAnalysis object resulting from a monster analysis}

\item{title}{The title of the plot}

\item{clusters}{A vector indicating the number of clusters to compute}

\item{alpha}{A vector indicating the level of transparency to be plotted}
}
\value{
ggplot2 object for transition matrix PCA
}
\description{
This function plots the first two principal components for a 
transaction matrix
}
\examples{
# data(yeast)
# design <- c(rep(0,20),rep(NA,10),rep(1,20))
# yeast$exp.cc[is.na(yeast$exp.cc)] <- mean(as.matrix(yeast$exp.cc),na.rm=TRUE)
# monsterRes <- monster(yeast$exp.cc, design, yeast$motif, nullPerms=100, numMaxCores=4)#' 
data(monsterRes)
slot(monsterRes, "logging") <- FALSE
# Color the nodes according to cluster membership
clusters <- kmeans(monsterGetTm(monsterRes),3)$cluster 
monsterTransitionPCAPlot(monsterRes, 
title="PCA Plot of Transition - Cell Cycle vs Stress Response", 
clusters=clusters)
}
