% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{networkChisq}
\alias{networkChisq}
\title{Function to run associations between categorical phenotypes and dichotomous network features.
If at least 5 samples exist in each group, run a Chi-square test and compute Cramer's V.
Otherwise, run a Fisher-Freeman-Halton Test.}
\usage{
networkChisq(network, phenotype)
}
\arguments{
\item{network}{The network features to test.}

\item{phenotype}{The phenotype of interest}
}
\value{
A data frame with two vectors: "cor", which lists the Chi-square or Fisher-Freeman-Halton Test p-values, 
and "V" which lists the Cramer's V statistics (will be NA if Fisher-Freeman-Halton Test is computed)
}
\description{
Function to run associations between categorical phenotypes and dichotomous network features.
If at least 5 samples exist in each group, run a Chi-square test and compute Cramer's V.
Otherwise, run a Fisher-Freeman-Halton Test.
}
