% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{networkContinuous}
\alias{networkContinuous}
\title{Function to find network correlations with continuous phenotypes.}
\usage{
networkContinuous(network, pheno, method)
}
\arguments{
\item{network}{: A weighted matrix where rows are samples and columns are network features.}

\item{pheno}{: phenotype matrix
with rows as samples and columns as phenotype variables.}

\item{method}{: One of "pearson", "spearman", or "kendall".}
}
\description{
Function to find network correlations with continuous phenotypes.
}
