% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{networkTTest}
\alias{networkTTest}
\title{Function to compute Welch's t-test statistics on dichotomous phenotypes with
continuous network features or on continuous phenotypes with dichotomous network features.}
\usage{
networkTTest(network, phenotype, dichotomousVar)
}
\arguments{
\item{network}{The network features to test.}

\item{phenotype}{A phenotype vector against which to run associations.}

\item{dichotomousVar}{Whether the phenotype is dichotomous or continuous.}
}
\value{
A data frame with two vectors: "cor", which lists the t-test p-values, 
and "V" which is set to NA.
}
\description{
Function to compute Welch's t-test statistics on dichotomous phenotypes with
continuous network features or on continuous phenotypes with dichotomous network features.
}
