% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPACAFunctions.R
\name{node.to.gene}
\alias{node.to.gene}
\title{Remove tags from gene names}
\usage{
node.to.gene(x)
}
\arguments{
\item{x}{Tagged node identifier}
}
\value{
Untagged node name
}
\description{
In gene regulatory networks, transcription factors can act as both "from" nodes (regulators) and "to" nodes (target genes), so the network analysis functions automatically tag the two columns to differentiate them. This function removes those tags from the gene identifiers.
}
\examples{
node.to.gene()
}
