% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{OutDegreeSim}
\alias{OutDegreeSim}
\title{Helper function to compute the Out-Degree similarities between two networks.}
\usage{
OutDegreeSim(network1, network2)
}
\arguments{
\item{network1}{A network for comparison.}

\item{network2}{Another network for comparison.}
}
\value{
The similarity value.
}
\description{
Helper function to compute the Out-Degree similarities between two networks.
}
