% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panda.diff.edges.R
\name{panda.diff.edges}
\alias{panda.diff.edges}
\title{Identify differential edges in two PANDA networks}
\usage{
panda.diff.edges(
  panda.net1,
  panda.net2,
  threshold = 0.8,
  condition_name = "cond.1"
)
}
\arguments{
\item{panda.net1}{vector indicating the PANDA networks of one condition or phenotype.}

\item{panda.net2}{vector indicating the PANDA networks of another compared condition orphenotype.}

\item{threshold}{numerical vector indicating a threshold probability to select select edges.}

\item{condition_name}{string vector indicating the condition name of net1}
}
\value{
a data.frame with five columns: tf, gene, motif, Score and defined condition name("T"in this column means egde belongs to first condition or phenotype,
"F" means edge belongs to the second condition or phenotype)
}
\description{
To determine the probability that an edge is “different” between the networks, 
we first subtracted the z-score weight values estimated by PANDA for the two networks and then determined the value of the inverse cumulative distribution for this difference. 
The product of these two probabilities represents the probability than an edge is both “supported” and “different.” 
We select edges for which this combined probability is greater than a threshold probability (default value is 0.8).
}
