% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandaR.R
\docType{package}
\name{pandaR}
\alias{pandaR}
\alias{pandaR-package}
\title{pandaR: Passing Messages between Biological Networks to Refine Predicted Interactions.}
\description{
The PANDA approach is to model the regulatory network as a bipartite network and estimate edge weights based on the evidence that information from a particular transcription factor i is successfully being passed to a particular gene j. This package provides a straightforward tool for applying this established method.
}

