\name{pandaResultPairs}
\docType{data}
\alias{pandaResultPairs}
\title{Analysis result from PANDA algorithm on toy data converted into pairs}
\description{
  This data panda object resulting from running the PANDA algorithm on the supplied toy dataset. The data 
  consists of a matrix of TF, Gene, initial link, and final Z.
  data(pandaResultPairs)
}
\usage{pandaResultPairs}
\format{A matrix}
\value{A matrix}
\references{
Glass K, Huttenhower C, Quackenbush J, Yuan GC. Passing Messages Between Biological Networks to Refine Predicted Interactions. PLoS One. 2013 May 31;8(5):e64832.
}
\keyword{datasets}
