\name{pandaToyData}
\docType{data}
\alias{pandaToyData}
\title{Toy gene expression, motif, and ppi data}
\description{
  This data is a list containing three data.frames.  The motif data.frame describes a set of pairwise connections where a specific known sequence motif of a transcription factor was found upstream of the corresponding gene.  The expression data.frame is a set of 1000 gene expression levels measured across 50 samples.  Finally, the ppi data.frame describes a set of known pairwise protein interactions.
}
\usage{pandaToyData}
\format{A list containing 3 data.frames}
\value{A list of length 3}
\source{}
\references{
Glass K, Huttenhower C, Quackenbush J, Yuan GC. Passing Messages Between Biological Networks to Refine Predicted Interactions. PLoS One. 2013 May 31;8(5):e64832.
}
\keyword{datasets}
