% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{phenotype_cor}
\alias{phenotype_cor}
\title{Function to compute correlation between continuous phenotypes.}
\usage{
phenotype_cor(phenotype, phenotypesToCompare, method)
}
\arguments{
\item{phenotype}{The phenotype vector to test.}

\item{phenotypesToCompare}{A data frame containing all phenotypes against which to run associations.
phenotypes against which to compare.}

\item{method}{: One of "pearson", "spearman", or "kendall".}
}
\value{
A data frame with two vectors: "cor", which lists the t-test p-values, 
and "V" which is set to NA.
}
\description{
Function to compute correlation between continuous phenotypes.
}
