% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plot.monsterAnalysis}
\alias{plot.monsterAnalysis}
\title{plot.monsterAnalysis}
\usage{
\method{plot}{monsterAnalysis}(x, ...)
}
\arguments{
\item{x}{an object of class "monsterAnalysis"}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Plot of the dTFI for each TF against null distribution
}
\description{
plots the sum of squares of off diagonal mass (differential TF Involvement)
}
\examples{
\donttest{
data(yeast)
monsterRes <- monster(yeast$exp.ko,c(rep(1,42),rep(0,49),rep(NA,15)),yeast$motif, nullPerms=10, numMaxCores=4)
plot(monsterRes)
}
}

