% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{plotAssociation}
\alias{plotAssociation}
\title{Plots the association between two variables}
\usage{
plotAssociation(
  expression,
  phenotype,
  phenotype_dictionary,
  variableX,
  variableY,
  variableTypeX,
  variableTypeY
)
}
\arguments{
\item{expression}{The gene expression data, formatted as for SEAHORSE}

\item{phenotype}{The phenotype data, formatted as for SEAHORSE}

\item{phenotype_dictionary}{The phenotype dictionary, formatted as for SEAHORSE}

\item{variableX}{The variable of interest to plot on the x-axis (a phenotype or gene)}

\item{variableY}{The variable of interest to plot on the y-axis (a phenotype or gene)}

\item{variableTypeX}{Either "gene" or "phenotype" (x-axis variable)}

\item{variableTypeY}{Either "gene" or "phenotype" (y-axis variable)}
}
\description{
Plots the association between two variables
}
