% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{plotBoxplot}
\alias{plotBoxplot}
\title{A helper function for plotAssociation(). Called when one is a dichotomous phenotype and
the other is a continuous variable.}
\usage{
plotBoxplot(x, y, xName, yName)
}
\arguments{
\item{x}{The data for the x axis}

\item{y}{The data for the y axis}

\item{xName}{The variable name to use for the x axis}

\item{yName}{The variable name to use for the y axis}
}
\description{
A helper function for plotAssociation(). Called when one is a dichotomous phenotype and
the other is a continuous variable.
}
