% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/algorithm_functions.R
\name{plotGraph}
\alias{plotGraph}
\title{Plot graph}
\usage{
plotGraph(x)
}
\arguments{
\item{x}{an object of class "panda"}
}
\value{
An matrix describing the subsetted bipartite network.
}
\description{
plotGraph plots a bipartite graph
}
\examples{
\donttest{
data(pandaToyData)
pandaRes <- panda(pandaToyData$motif,
           pandaToyData$expression,pandaToyData$ppi,hamming=.001,progress=TRUE)
topPandaRes <- topedges(pandaRes,1000)
subnet.pandaRes <- subnetwork(topPandaRes,c("AR","ARID3A","ELK1"))
plotGraph(subnet.pandaRes)
}
data(pandaResult)
topPandaRes <- topedges(pandaResult, 1000)
subnet.pandaRes <- subnetwork(topPandaRes,c("AR","ARID3A","ELK1"))
plotGraph(subnet.pandaRes)
}
\keyword{keywords}

