% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{plotHeat}
\alias{plotHeat}
\title{A helper function for plotAssociation(). Called when both variables are dichotomous phenotypes.}
\usage{
plotHeat(x, y, xName, yName)
}
\arguments{
\item{x}{The data for the x axis}

\item{y}{The data for the y axis}

\item{xName}{The variable name to use for the x axis}

\item{yName}{The variable name to use for the y axis}
}
\description{
A helper function for plotAssociation(). Called when both variables are dichotomous phenotypes.
}
