% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TARDIGRADE.R
\name{PlotKnowledgeGraph}
\alias{PlotKnowledgeGraph}
\title{Plots the knowledge graph using a Kamada-Kawai spatial layout. Edges are not shown,
as the location represents the orientation of the node in the graph and the
knowledge graph is expected to be dense. Nodes are single-pixel with low alpha
values by default, to show overlap. Clusters are color-coded if present. Otherwise,
no color is shown.}
\usage{
PlotKnowledgeGraph(
  knowledgeGraph,
  nodeSize = 1,
  alpha = 0.1,
  edgeWidth = 0,
  layout = NULL
)
}
\arguments{
\item{knowledgeGraph}{An object of type TARDIGRADE_KnowledgeGraph.}

\item{nodeSize}{The size of the node in pixels.}

\item{alpha}{The opacity of each node.}

\item{edgeWidth}{The width of the edges.}
}
\description{
Plots the knowledge graph using a Kamada-Kawai spatial layout. Edges are not shown,
as the location represents the orientation of the node in the graph and the
knowledge graph is expected to be dense. Nodes are single-pixel with low alpha
values by default, to show overlap. Clusters are color-coded if present. Otherwise,
no color is shown.
}
