% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLOBFISH.R
\name{PlotNetwork}
\alias{PlotNetwork}
\title{Plot the networks, using different colors for transcription factors, genes of interest,
and additional genes.}
\usage{
PlotNetwork(
  network,
  genesOfInterest,
  geneOfInterestColor = "red",
  tfColor = "blue",
  otherGenesColor = "gray",
  nodeSize = 1,
  edgeWidth = 0.5,
  vertexLabels = NA,
  vertexLabelSize = 0.7,
  vertexLabelOffset = 0.5,
  layoutBipartite = TRUE,
  edgeColor = "gray",
  edgeWidthHop1 = 0.5,
  edgeColorHop1 = "gray"
)
}
\arguments{
\item{network}{A data frame with the following format:
tf,gene}

\item{genesOfInterest}{Which genes of interest to highlight}

\item{geneOfInterestColor}{Color for the genes of interest}

\item{tfColor}{Color for the transcription factors}

\item{otherGenesColor}{Color for the other genes}

\item{nodeSize}{Size of node}

\item{edgeWidth}{Width of edges}

\item{vertexLabels}{Which vertex labels to include. By default, none are included.}

\item{vertexLabelSize}{The size of label to use for the vertex, as a fraction of the default.}

\item{vertexLabelOffset}{Number of pixels in the offset when plotting labels.
Default is TRUE.}

\item{layoutBipartite}{Whether or not to layout as a bipartite graph.}

\item{edgeColor}{Color of edges}

\item{edgeWidthHop1}{Width of edges 1 hop away from genes of interest.}

\item{edgeColorHop1}{Color of edges 1 hop away from genes of interest.}
}
\value{
A bipartite plot of the network
}
\description{
Plot the networks, using different colors for transcription factors, genes of interest,
and additional genes.
}
