% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNAGI.R
\name{PlotNetworkU}
\alias{PlotNetworkU}
\title{Plot the networks, using different colors for transcription factors, genes of interest,
and additional genes.}
\usage{
PlotNetworkU(
  network,
  nodeSize = 1,
  edgeWidth = 0.5,
  vertexLabels = NA,
  vertexLabelSize = 0.7,
  vertexLabelOffset = 0.5,
  geneColorMapping = NULL
)
}
\arguments{
\item{network}{A data frame with the following format:
tf,gene}

\item{nodeSize}{Size of node}

\item{edgeWidth}{Width of edges}

\item{vertexLabels}{Vector of labels to include. Labels that you do not wish to include
should be set to NA.}

\item{vertexLabelSize}{The size of label to use for the vertex, as a fraction of the default.}

\item{vertexLabelOffset}{Number of pixels in the offset when plotting labels.
Default is TRUE.}

\item{geneColorMapping}{Color mapping from a set of genes to a color. The
nodes and edges connected to them will be this color. If NULL, all genes and
their edges will be gray. The format is a data frame, where the first column ("gene")
is the name of the gene and the second ("color") is the color. All nodes must
be included in the color mapping.}
}
\description{
Plot the networks, using different colors for transcription factors, genes of interest,
and additional genes.
}
