% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPanda.R
\name{plotPanda}
\alias{plotPanda}
\title{Plot PANDA network}
\usage{
plotPanda(top = 100, file = "panda_top_100.png")
}
\arguments{
\item{top}{Numeric vector indicating the amount of edges selected to plot by decreasting order of egde weights. Defaults to 100.}

\item{file}{Character string indicating the name of output .png file. Defaults to 'panda_top_100.png'.}
}
\value{
a message showing the path of output plot file.
}
\description{
This function is able to plot specified amount of egdes in PANDA network, after generate PANDA network \code{\link{runPanda}}.
}
\examples{
# refer to the input datasets files of control in inst/extdat as example
control_expression_file_path <- system.file("extdata", "expr10.txt", package = "netZoo", mustWork = TRUE)
motif_file_path <- system.file("extdata", "chip.txt", package = "netZoo", mustWork = TRUE)
ppi_file_path <- system.file("extdata", "ppi.txt", package = "netZoo", mustWork = TRUE)

# Run PANDA algorithm
control_all_panda_result <- runPanda(e = control_expression_file_path, m = motif_file_path, ppi = ppi_file_path, rm_missing = TRUE )

# access PANDA regulatory network
control_net <- control_all_panda_result$panda

# plot the top 100 edges of the PANDA network
plotPanda(top =100, file="top100_panda.png")

}
