% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPANDAinCytoscape.R
\name{plotPANDAinCytoscape}
\alias{plotPANDAinCytoscape}
\title{Plot PANDA network in Cytoscape}
\usage{
plotPANDAinCytoscape(panda.net, network.name = "PANDA")
}
\arguments{
\item{panda.net}{Character string indicating the input PANDA network in data frame structure type.}

\item{network.name}{Character string indicating the name of Cytoscape network.}
}
\description{
This function is able to modify PANDA network and plot in Cytoscape.
}
