% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{PlotROC}
\alias{PlotROC}
\title{This function plots the ROC curve for a given similarity index.}
\usage{
PlotROC(
  averageSims,
  lowestSims = NULL,
  highestSims = NULL,
  auc,
  xlab,
  ylab,
  main,
  absoluteMin = NULL,
  absoluteMax = NULL,
  showCutoffs = TRUE,
  minTextDistAsPercentage = 10
)
}
\arguments{
\item{averageSims}{The average similarities for the in-groups and out-groups.
This is a FERRET_Similarities object.}

\item{lowestSims}{The lower boundary for the curve. This is a FERRET_Similarities object.}

\item{highestSims}{The upper boundary for the curve. This is a FERRET_Similarities object.}

\item{auc}{This is a total AUC score to include on the curve as text.}

\item{xlab}{This is the label to include on the X axis.}

\item{ylab}{This is the label to include on the Y axis.}

\item{main}{This is the title of the plot.}

\item{absoluteMin}{The absolute minimum similarity in the results.}

\item{absoluteMax}{The absolute maximum similarity in the results.}

\item{showCutoffs}{Whether or not to show the cutoffs on the plot. Default is TRUE.}

\item{minTextDistAsPercentage}{The minimum distance between text as a percentage
of the distance in the plot. This prevents annotations from overlapping.}
}
\value{
Nothing, but generates a plot as a side effect.
}
\description{
This function plots the ROC curve for a given similarity index.
}
