% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotZbyTF}
\alias{plotZbyTF}
\title{Plot Z by TF outdegree quartiles}
\usage{
plotZbyTF(x, y, motif, hasPrior = TRUE)
}
\arguments{
\item{x}{PANDA object - output of the \code{panda} function.}

\item{y}{PANDA object - second PANDA object.}

\item{motif}{Motif used to construct the networks.}

\item{hasPrior}{TRUE/FALSE, If TRUE plots the edges that are given a weight > 0 in the motif,
else plot those given a weight of 0}
}
\value{
ggplot heatmapfor each TF, get outdegree in regulatory prior
}
\description{
Generates a Z-score scatterplot for edges according to the TF outdegree in prior.
The two PANDA objects should only differ in the gene expression used for the network 
constructions or other parameters.
}
\examples{

data(pandaResult)
data(pandaToyData)
regnet = slot(pandaResult,"regNet")
plotZbyTF(regnet,regnet, pandaToyData$motif, hasPrior=TRUE)

}

