% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{print.monsterAnalysis}
\alias{print.monsterAnalysis}
\title{print.monsterAnalysis}
\usage{
print.monsterAnalysis(x, ...)
}
\arguments{
\item{x}{an object of class "monster"}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Description of transition matrices in object
}
\description{
summarizes the results of a MONSTER analysis
}
\examples{
\donttest{
data(yeast)
monster(yeast$exp.ko,c(rep(1,42),rep(0,49),rep(NA,15)),yeast$motif, nullPerms=10, numMaxCores=4)
}
}
