% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIGER.R
\name{prior.pp}
\alias{prior.pp}
\title{Filter low confident edge signs in the prior network using GeneNet}
\usage{
prior.pp(prior, expr)
}
\arguments{
\item{prior}{A prior network (adjacency matrix) with rows as TFs and columns as genes.}

\item{expr}{A normalized log-transformed gene expression matrix.}
}
\value{
A filtered prior network (adjacency matrix).
}
\description{
Filter low confident edge signs in the prior network using GeneNet
}
