% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARN.R
\name{qstats}
\alias{qstats}
\title{Compute quantile statistics}
\source{
\href{https://raw.githubusercontent.com/kokrah/qsmooth/master/R/qstats.r}{Kwame Okrah's qsmooth R package}
Compute quantile statistics
}
\usage{
qstats(exprs, groups, window)
}
\arguments{
\item{exprs}{for counts use log2(raw counts + 1)), for MA use log2(raw intensities)}

\item{groups}{groups to which samples belong (character vector)}

\item{window}{window size for running median as a fraction on the number of rows of exprs}
}
\value{
list of statistics
}
\description{
This function was directly borrowed from github user kokrah.
}
