% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{ReadRobustnessAUC}
\alias{ReadRobustnessAUC}
\title{This reads a file generated by WriteRobustnessAUC().}
\usage{
ReadRobustnessAUC(fileName)
}
\arguments{
\item{fileName}{A valid file name.}
}
\value{
An object of type FERRET_ROC_AUC
}
\description{
This reads a file generated by WriteRobustnessAUC().
}
