% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{rugPlot}
\alias{rugPlot}
\title{Generates a rug plot for a single pathway-phenotype association.}
\usage{
rugPlot(result, pathwayName, pathways, phenotypeName)
}
\arguments{
\item{result}{The SEAHORSE result}

\item{pathwayName}{The pathway of interest}

\item{pathways}{The pathways used as input to SEAHORSE}

\item{phenotypeName}{The phenotype of interest}
}
\description{
Generates a rug plot for a single pathway-phenotype association.
}
