% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLOBFISH.R
\name{RunBLOBFISH}
\alias{RunBLOBFISH}
\title{Given a set of genes of interest, full bipartite networks with scores (one network for each sample), a significance
cutoff for statistical testing, and a hop constraint, BLOBFISH finds a subnetwork of
significant edges connecting the genes.}
\usage{
RunBLOBFISH(
  geneSet,
  networks,
  alpha,
  hopConstraint,
  nullDistribution,
  verbose = FALSE,
  topX = NULL,
  doFDRAdjustment = TRUE,
  pValueChunks = 100,
  loadPValues = FALSE,
  pValueFile = "pvalues.RDS"
)
}
\arguments{
\item{geneSet}{A character vector of genes comprising the targets of interest.}

\item{networks}{A list of bipartite (PANDA-like) networks, where each network is a data frame with the following format:
tf,gene,score}

\item{alpha}{The significance cutoff for the statistical test.}

\item{hopConstraint}{The maximum number of hops to be considered between gene pairs.
Must be an even number.}

\item{nullDistribution}{The null distribution, specified as a vector of values.}

\item{verbose}{Whether or not to print detailed information about the run.}

\item{topX}{Select the X lowest significant p-values for each gene. NULL by default.}

\item{doFDRAdjustment}{Whether or not to perform FDR adjustment.}

\item{pValueChunks}{The number of chunks to split when calculating the p-value. This
parameter allows the edges to be split into chunks to prevent memory errors.}

\item{loadPValues}{Whether p-values should be loaded from pValueFile or re-generated.
Default is FALSE.}

\item{pValueFile}{The file where the p-values should be saved. If NULL, they are not
saved and need to be recalculated.}
}
\value{
A bipartite subnetwork in the same format as the original networks.
}
\description{
Given a set of genes of interest, full bipartite networks with scores (one network for each sample), a significance
cutoff for statistical testing, and a hop constraint, BLOBFISH finds a subnetwork of
significant edges connecting the genes.
}
