% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLioness.R
\name{runLioness}
\alias{runLioness}
\title{Run python implementation of LIONESS}
\usage{
runLioness(e = expression, m = motif, ppi = ppi,
  rm_missing = FALSE)
}
\arguments{
\item{e}{Character String indicatining the file path of expression values file, as each gene (row) by samples (columns) \emph{required}}

\item{m}{Character String indicatining the file path of pair file of motif edges,
when not provided, analysis continues with Pearson correlation matrix. \emph{optional}}

\item{ppi}{Character String indicatining the pair file path of Protein-Protein interaction dataset. \emph{optional}}

\item{rm_missing}{Boolean indicatining whether to remove missing values. If TRUE, removes missing values.
if FALSE, keep missing values. THe default value is FALSE. \emph{optional}}
}
\value{
A data frame with columns representing each sample, rows representing the regulator-target pair in PANDA network generated by \code{\link{runPanda}}. 
        Each cell filled with the related score, representing the estimated contribution of a sample to the aggregate network.
}
\description{
\strong{LIONESS}(Linear Interpolation to Obtain Network Estimates for Single Samples) is a method to estimate sample-specific regulatory networks.
 \href{https://arxiv.org/abs/1505.06440}{[(LIONESS arxiv paper)])}.
}
\examples{
# refer to the input datasets files of control in inst/extdat as example
control_expression_file_path <- system.file("extdata", "expr10.txt", package = "netZoo", mustWork = TRUE)
motif_file_path <- system.file("extdata", "chip.txt", package = "netZoo", mustWork = TRUE)
ppi_file_path <- system.file("extdata", "ppi.txt", package = "netZoo", mustWork = TRUE)

# Run PANDA algorithm
control_lioness_result <- runLioness(e = control_expression_file_path, m = motif_file_path, ppi = ppi_file_path, rm_missing = TRUE )

}
