% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPanda.R
\name{runPanda}
\alias{runPanda}
\title{Run pypanda in R}
\usage{
runPanda(e = expression, m = motif, ppi = ppi, rm_missing = FALSE)
}
\arguments{
\item{e}{Character String indicatining the file path of expression values file, as each gene (row) by samples (columns) \emph{required}}

\item{m}{Character String indicatining the file path of pair file of motif edges,
when not provided, analysis continues with Pearson correlation matrix. \emph{optional}}

\item{ppi}{Character String indicatining the pair file path of Protein-Protein interaction dataset. \emph{optional}}

\item{rm_missing}{Boolean indicatining whether to remove missing values. If TRUE, removes missing values.
if FALSE, keep missing values. THe default value is FALSE. \emph{optional}}
}
\value{
List of three items：
 Use \code{$panda} to access the entire data frame of PANDA output network consisting of four columns: 
"tf", "gene", "motif", and "force".

Use \code{$indegree} to access the data frame of indegree of PANDA network, consisting of two columns: "gene", "force".

Use \code{$outdegree} to access the data frame of outdegree of PANDA network, consisting of two columns: "tf", "force".

\strong{Note}: If there is any duplicate name of nodes in first two columns,
prefix the content in regulator column with "reg_" and content in target column with"tar_"

Please re-name the colnames if needed.
}
\description{
\strong{PANDA}(Passing Attributes between Networks for Data Assimilation) is a message-passing model 
to gene regulatory network reconstruction. It integrates multiple sources of biological data, 
including protein-protein interaction, gene expression, and sequence motif information,
in order to reconstruct genome-wide, condition-specific regulatory networks.
\href{http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0064832}{[(Glass et al. 2013)])}
This function is able to run \href{https://github.com/davidvi/pypanda}{pypanda} -- Python implementation of PANDA in R enviroment.
}
\examples{
# refer to four input datasets files in inst/extdat
treated_expression_file_path <- system.file("extdata", "expr4.txt", package = "netZoo", mustWork = TRUE)
control_expression_file_path <- system.file("extdata", "expr10.txt", package = "netZoo", mustWork = TRUE)
motif_file_path <- system.file("extdata", "chip.txt", package = "netZoo", mustWork = TRUE)
ppi_file_path <- system.file("extdata", "ppi.txt", package = "netZoo", mustWork = TRUE)


# Run PANDA for treated and control network
treated_all_panda_result <- runPanda(e = treated_expression_file_path, m = motif_file_path, ppi = ppi_file_path, rm_missing = TRUE )
control_all_panda_result <- runPanda(e = control_expression_file_path, m = motif_file_path, ppi = ppi_file_path, rm_missing = TRUE )

# access PANDA regulatory network
treated_net <- treated_all_panda_result$panda
control_net <- control_all_panda_result$panda

# access PANDA regulatory indegree network result
indegree_net <- treated_all_panda_result$indegree

# access PANDA regulatory outdegree network result
outdegree_net <- treated_all_panda_resultt$outdegree

}
