% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNAGI.R
\name{RunUNAGI}
\alias{RunUNAGI}
\title{Given a set of genes of interest, full unipartite networks with scores (one network for each sample), a significance
cutoff for statistical testing, and a hop constraint, UNAGI finds a subnetwork of
significant edges connecting the genes.}
\usage{
RunUNAGI(nodeSet, network, hopConstraint, verbose = FALSE, topX = NULL)
}
\arguments{
\item{nodeSet}{A character vector of genes comprising the targets of interest.}

\item{network}{A network with the first column ("source") consisting of source nodes,
the second ("target") consisting of target nodes, and the third ("score") consisting of scores}

\item{hopConstraint}{The maximum number of hops to be considered between gene pairs.}

\item{verbose}{Whether or not to print detailed information about the run.}

\item{topX}{Select the X lowest significant p-values for each gene. NULL by default.
Default is FALSE.}
}
\value{
A unipartite subnetwork in the same format as the original networks.
}
\description{
Given a set of genes of interest, full unipartite networks with scores (one network for each sample), a significance
cutoff for statistical testing, and a hop constraint, UNAGI finds a subnetwork of
significant edges connecting the genes.
}
