% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{ScaleNetworks}
\alias{ScaleNetworks}
\title{ScaleNetworks}
\usage{
ScaleNetworks(networks, min)
}
\arguments{
\item{networks}{A list of unscaled networks}

\item{min}{Minimum value to consider}
}
\value{
A list of scaled networks
}
\description{
ScaleNetworks
}
