% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{ScaleNetworksByPercentile}
\alias{ScaleNetworksByPercentile}
\title{ScaleNetworksByPercentile}
\usage{
ScaleNetworksByPercentile(networks, numberOfCutoffs)
}
\arguments{
\item{networks}{A list of unscaled networks}

\item{numberOfCutoffs}{Number of quantiles to consider}
}
\value{
A list of scaled networks
}
\description{
ScaleNetworksByPercentile
}
