% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLOBFISH.R
\name{SignificantBreadthFirstSearch}
\alias{SignificantBreadthFirstSearch}
\title{Find all significant edges adjacent to the starting nodes, excluding the nodes
specified.}
\usage{
SignificantBreadthFirstSearch(
  networks,
  alpha,
  startingNodes,
  nodesToExclude,
  startFromTF,
  doFDRAdjustment = FALSE,
  nullDistribution,
  verbose = FALSE,
  topX = NULL
)
}
\arguments{
\item{networks}{A concatenation of n PANDA-like networks with the following format:
tf,gene,score_net1, score_net2, ... , score_netn
Edges must be specified as "tf__gene".}

\item{alpha}{The significance cutoff for the statistical test.}

\item{startingNodes}{The list of nodes from which to start.}

\item{nodesToExclude}{The list of nodes to exclude from the search.}

\item{startFromTF}{Whether to start from transcription factors (TRUE) or genes (FALSE).}

\item{doFDRAdjustment}{Whether or not to adjust the p-values using FDR.}

\item{nullDistribution}{The null distribution, specified as a vector of values.}

\item{verbose}{Whether or not to print detailed information about the run.}

\item{topX}{Select the X lowest significant p-values for each gene. NULL by default.}
}
\value{
A bipartite subnetwork in the same format as the original networks.
}
\description{
Find all significant edges adjacent to the starting nodes, excluding the nodes
specified.
}
