% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPACAFunctions.R
\name{simulateNetwork}
\alias{simulateNetwork}
\title{Simulated networks}
\usage{
simulateNetwork(comm.sizes, edge.mat, num.module, size.module, dens.module)
}
\arguments{
\item{comm.sizes}{A two-column matrix indicating the number of "from" nodes (left column) and number of "to" nodes (right column) in each community (row).}

\item{edge.mat}{A matrix indicating the number of edges from the TFs in community i (rows) to target genes in community j (columns).}

\item{num.module}{The number of modules that will be added to simulate the disease network.}

\item{size.module}{A two-column matrix indicating the number of "from" and "to" nodes in each new module (row) that will be added to simulate the disease network.}

\item{dens.module}{A vector of length num.module, indicating the edge density of each added module.}
}
\value{
A list with two elements. The first element is a four-column edge table of the same form that is input into the differential modularity function. The second element is a list of all the new nodes in the modules that were added to create the disease network.
}
\description{
This function creates a pair of networks given user-defined parameters for the modular structure of the first (healthy) network and the type of added module in the second (disease) network.
}
\examples{
simulateNetwork()
}
