% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARN.R
\docType{data}
\name{skin}
\alias{skin}
\title{Skin dataset}
\format{
An object of class "ExpressionSet".
}
\usage{
data(skin)
}
\value{
An object of class "ExpressionSet".
}
\description{
Skin RNA-seq data from the GTEx consortium. V6 release.
}
\references{
GTEx Consortium, 2015. The Genotype-Tissue Expression (GTEx) 
pilot analysis: Multitissue gene regulation in humans. Science, 348(6235), 
pp.648-660.
}
\keyword{datasets}
