% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDOR.R
\docType{data}
\name{small1976}
\alias{small1976}
\title{Pollinator-plant interactions}
\format{
A data frame with 442 rows and 3 variables
}
\usage{
data(small1976)
}
\description{
A dataset containing the number of interactions 34 plants and 13 pollinators. The variables are as follows:
}
\details{
\itemize{
  \item pollinator. Species name of insect pollinator
  \item plant. Species name of plant
  \item interactions. Number of visitors caught on each plant species
}
}
\references{
Small (1976) Pollinator-plant interactions dataset.
}
\keyword{datasets}
