% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{spearmanPFromRho}
\alias{spearmanPFromRho}
\title{Approximates the p-value given correlation and the number of samples}
\usage{
spearmanPFromRho(rho, n)
}
\arguments{
\item{rho}{The correlation vector for a given phenotype}

\item{n}{The vector of sample counts used to calculate rho}
}
\value{
A vector of p-values
}
\description{
Approximates the p-value given correlation and the number of samples
}
