% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{SubspaceSim}
\alias{SubspaceSim}
\title{Helper function to compute the Subspace similarities between two networks.}
\usage{
SubspaceSim(projection1, projection2, k)
}
\arguments{
\item{projection1}{The projection of network 1.}

\item{projection2}{The projection of network 2.}

\item{k}{Number of eigenvectors to include in the subspace projection.}
}
\value{
The similarity value.
}
\description{
Helper function to compute the Subspace similarities between two networks.
}
