% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{summary.panda}
\alias{summary.panda}
\title{Summary.panda}
\usage{
\method{summary}{panda}(object, ...)
}
\arguments{
\item{object}{an object of class "panda"}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Summary description of panda S4 object
}
\description{
summarizes the results of a PANDA analysis
}
\examples{
\donttest{
data(pandaToyData)
panda.res <- panda(pandaToyData$motif,
           pandaToyData$expression,pandaToyData$ppi,hamming=.001,progress=TRUE)
summary(panda.res)}
}
\keyword{keywords}

