% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{summaryHistogramGene}
\alias{summaryHistogramGene}
\title{Generates a summary histogram for a gene.}
\usage{
summaryHistogramGene(expression, geneName, breaks = 20)
}
\arguments{
\item{expression}{The gene expression data, formatted as for SEAHORSE}

\item{breaks}{The number of breaks in the histogram}
}
\description{
Generates a summary histogram for a gene.
}
