% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{summaryHistogramPhenotype}
\alias{summaryHistogramPhenotype}
\title{Generates a summary histogram for a phenotype.}
\usage{
summaryHistogramPhenotype(
  phenotype,
  phenotypeName,
  phenotype_dictionary,
  breaks = 20
)
}
\arguments{
\item{phenotype}{The phenotype data, formatted as for SEAHORSE}

\item{phenotypeName}{The name of the phenotype}

\item{phenotype_dictionary}{The phenotype dictionary, formatted as for SEAHORSE}

\item{breaks}{The number of breaks in the histogram}
}
\description{
Generates a summary histogram for a phenotype.
}
