% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TARDIGRADE.R
\docType{class}
\name{TARDIGRADE_KnowledgeGraph-class}
\alias{TARDIGRADE_KnowledgeGraph-class}
\title{TARDIGRADE_KnowledgeGraph class
 This object type represents the list of results in a given directory.
 
 @name TARDIGRADE_KnowledgeGraph-class
 @rdname TARDIGRADE_KnowledgeGraph-class
 @exportClass TARDIGRADE_KnowledgeGraph
 @slot adjacencyMatrix The adjacency matrix for the graph.
 @slot edgeDict The dictionary mapping edges to pathways. This is a list.
 @slot clusters A named list of clusters.
 @slot analytePathwayMapping The original data frame mapping analytes to pathways.
 @slot clusterPathwayDict The dictionary mapping clusters to pathways.}
\description{
TARDIGRADE_KnowledgeGraph class
 This object type represents the list of results in a given directory.
 
 @name TARDIGRADE_KnowledgeGraph-class
 @rdname TARDIGRADE_KnowledgeGraph-class
 @exportClass TARDIGRADE_KnowledgeGraph
 @slot adjacencyMatrix The adjacency matrix for the graph.
 @slot edgeDict The dictionary mapping edges to pathways. This is a list.
 @slot clusters A named list of clusters.
 @slot analytePathwayMapping The original data frame mapping analytes to pathways.
 @slot clusterPathwayDict The dictionary mapping clusters to pathways.
}
