% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPACAFunctions.R
\name{tidyconfig}
\alias{tidyconfig}
\title{Renumbering community membership vector}
\usage{
tidyconfig(S)
}
\arguments{
\item{S}{The community membership vector derived from the previous round of agglomeration.}

\item{J}{The modularity matrix}
}
\value{
The renumbered membership vector.
}
\description{
This is a helper function genlouvain. It re-numbers the communities so that they run from 1 to N increasing through the vector.
}
\examples{
tidyconfig()
}
