% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIGER.R
\docType{data}
\name{TIGER_expr}
\alias{TIGER_expr}
\title{TIGER example expression matrix}
\format{
## `TIGER_expr`
A gene expression matrix with 1780 rows (genes) and 16 columns (samples)
}
\source{
<https://zenodo.org/record/7425777>
}
\usage{
data(TIGER_expr)
}
\description{
TIGER example expression matrix
}
\keyword{datasets}
