% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIGER.R
\docType{data}
\name{TIGER_prior}
\alias{TIGER_prior}
\title{TIGER example prior network}
\format{
## `TIGER_prior`
A prior network matrix with 14 rows (TFs) and 1772 columns (genes)
}
\source{
<https://zenodo.org/record/7425777>
}
\usage{
data(TIGER_prior)
}
\description{
TIGER example prior network
}
\keyword{datasets}
