% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPACAFunctions.R
\name{TopEnsembl.to.TopSym}
\alias{TopEnsembl.to.TopSym}
\title{Translating gene identifiers to gene symbols}
\usage{
TopEnsembl.to.TopSym(mod.top, annot.vec)
}
\arguments{
\item{mod.top}{A list of gene sets (gene identifiers)}

\item{annot.vec}{A vector of gene symbols with gene identifiers as the names of the vector, that defines the translation between annotations.}
}
\value{
A list of sets of gene symbols.
}
\description{
Takes a list of gene sets named using gene identifiers and converts them to a list of symbols given a user-defined annotation table.
}
\examples{
TopEnsembl.to.TopSym()
}
