% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{transitionPCAPlot}
\alias{transitionPCAPlot}
\title{Principal Components plot of transformation matrix}
\usage{
transitionPCAPlot(monsterObj, title = "PCA Plot of Transition",
  clusters = 1, alpha = 1)
}
\arguments{
\item{title}{The title of the plot}

\item{clusters}{A vector indicating the colors to be plotted for each node}

\item{alpha}{A vector indicating the level of transparency to be plotted 
for each node}

\item{tm}{a transition matrix for a bipartite network}
}
\value{
ggplot2 object for transition matrix PCA
}
\description{
This function plots the first two principal components for a 
transaction matrix
}
\examples{
# data(yeast)
# design <- c(rep(0,20),rep(NA,10),rep(1,20))
# monsterRes <- monster(yeast$exp.cc, design, yeast$motif, nullPerms=100, numMaxCores=4)#' 
load(file="./data/monsterRes.rda")
# Color the nodes according to cluster membership
clusters <- kmeans(slot(monsterRes, 'tm'),3)$cluster 
transitionPCAPlot(monsterRes, 
title="PCA Plot of Transition - Cell Cycle vs Stress Response", 
clusters=clusters)
}
\keyword{keywords}

