% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.diff.panda.in.cytoscape.R
\name{vis.diff.panda.in.cytoscape}
\alias{vis.diff.panda.in.cytoscape}
\title{Plot two PANDA networsk in Cytoscape}
\usage{
vis.diff.panda.in.cytoscape(net1, net2,
  network.name = "DiffPANDAnetwork")
}
\arguments{
\item{net1}{Character string indicating the input PANDA network in data frame structure type.}

\item{net2}{Character string indicating the another input PANDA network in data frame structure type.}

\item{network.name}{Character string indicating the name of Cytoscape network.}
}
\description{
This function is able to modify PANDA network and plot in Cytoscape.
}
\examples{
\dontrun{
# run this function to create a network in Cytoscape.
vis.diff.panda.in.cytoscape(panda.net1, panda.net1, network.name="diff.PANDA")
}
}
