% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visPandaInCytoscape.R
\name{visPandaInCytoscape}
\alias{visPandaInCytoscape}
\title{Plot PANDA network in Cytoscape}
\usage{
visPandaInCytoscape(panda.net, network_name = "PANDA")
}
\arguments{
\item{panda.net}{Character string indicating the input PANDA network in data frame structure type.}

\item{network_name}{Character string indicating the name of Cytoscape network.}
}
\value{
PANDA network in Cytoscape
}
\description{
This function is able to modify PANDA network and plot in Cytoscape. Please make sure that Cytoscape 
is installed and open it before calling this function.
}
