% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPACAFunctions.R
\name{WBM.louvain}
\alias{WBM.louvain}
\title{Generalized Louvain method for bipartite networks}
\usage{
WBM.louvain(net.frame)
}
\arguments{
\item{net.frame}{A table of edges, with the first column representing the TFs ("from" nodes) and the second column representing the targets ("to" nodes). The third column contains the edge weights corresponding to the network of interest.}
}
\value{
List where first element is the community membership vector and second element is the contribution score of each node to its community's portion of the bipartite modularity.
}
\description{
This function implements a generalized form of the Louvain method for weighted bipartite networks.
}
\examples{
WBM.louvain()
}
