% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{writeExpression}
\alias{writeExpression}
\title{Generates the expression data in the format expected by the UI. Writes a
data frame formatted as ENSG    SAMPID  GENE_EXPRESSION as .tsv.gz.}
\usage{
writeExpression(inFiles, file)
}
\arguments{
\item{inFiles}{The files to input to SEAHORSE, formatted as RDS.}

\item{file}{The name of the file to write. .tsv.gz. will be appended.}
}
\description{
Generates the expression data in the format expected by the UI. Writes a
data frame formatted as ENSG    SAMPID  GENE_EXPRESSION as .tsv.gz.
}
