% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{writeMapping}
\alias{writeMapping}
\title{Generates the gene mapping data in the format expected by the UI. Writes as
.tsv.gz formatted as ALIAS	ENSEMBL	SYMBOL	ENTREZID.}
\usage{
writeMapping(inFiles, file)
}
\arguments{
\item{inFiles}{The files to input to SEAHORSE, formatted as RDS.}

\item{file}{The name of the file to write. .tsv.gz. will be appended.}
}
\description{
Generates the gene mapping data in the format expected by the UI. Writes as
.tsv.gz formatted as ALIAS	ENSEMBL	SYMBOL	ENTREZID.
}
