% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{writePhenotype}
\alias{writePhenotype}
\title{Generates the phenotype data in the format expected by the UI.
Format is SAMPID	tissue	VARNAME	VALUE}
\usage{
writePhenotype(inFiles, file)
}
\arguments{
\item{inFiles}{The files to input to SEAHORSE, formatted as RDS.}

\item{file}{The file where results should be stored. .tsv.gz will be appended.}
}
\description{
Generates the phenotype data in the format expected by the UI.
Format is SAMPID	tissue	VARNAME	VALUE
}
