% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{WriteRobustnessAUC}
\alias{WriteRobustnessAUC}
\title{This function saves the results of ComputeRobustnessAUC to a file.}
\usage{
WriteRobustnessAUC(results, fileName)
}
\arguments{
\item{results}{An object of type FERRET_ROC_AUC}

\item{fileName}{A valid file name.}
}
\description{
This function saves the results of ComputeRobustnessAUC to a file.
}
