% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{writeToGz}
\alias{writeToGz}
\title{Writes a data frame as .csv.gz.}
\usage{
writeToGz(data, file)
}
\arguments{
\item{data}{The data frame}

\item{file}{The name of the file to write. .tsv.gz. will be appended.}
}
\description{
Writes a data frame as .csv.gz.
}
