% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPACA.R
\name{alpacatidyconfig}
\alias{alpacatidyconfig}
\title{Renumbering community membership vector}
\usage{
alpacatidyconfig(S)
}
\arguments{
\item{S}{The community membership vector derived from the previous round of agglomeration.}
}
\value{
The renumbered membership vector.
}
\description{
This is a helper function alpacagenlouvain. It re-numbers the communities so that they run from 1 to N increasing through the vector.
}
\examples{
a <- 1 # example place holder
}
