% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTERFunctions.R
\name{bereFull}
\alias{bereFull}
\title{Bipartite Edge Reconstruction from Expression data 
(composite method with direct/indirect)}
\usage{
bereFull(motif.data, expr.data, alpha = 0.5, lambda = 10,
  score = "motifincluded")
}
\arguments{
\item{motif.data}{A motif dataset, a data.frame, matrix or exprSet 
containing 3 columns. Each row describes an motif associated 
with a transcription factor (column 1) a gene (column 2) 
and a score (column 3) for the motif.}

\item{expr.data}{An expression dataset, as a genes (rows) by 
samples (columns) data.frame}

\item{alpha}{A weight parameter specifying proportion of weight 
to give to indirect compared to direct evidence.  See documentation.}

\item{lambda}{if using penalized, the lambda parameter in the penalized logistic regression}

\item{score}{String to indicate whether motif information will 
be readded upon completion of the algorithm}
}
\value{
An matrix or data.frame
}
\description{
This function generates a complete bipartite network from 
gene expression data and sequence motif data. This NI method
serves as a default method for inferring bipartite networks
in MONSTER.  Running bereFull can generate these networks
independently from the larger MONSTER method.
}
\examples{
data(yeast)
monsterRes <- bereFull(yeast$motif, yeast$exp.cc, alpha=.5)
}
