% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLOBFISH.R
\name{BuildSubnetwork}
\alias{BuildSubnetwork}
\title{Find the subnetwork of significant edges connecting the genes.}
\usage{
BuildSubnetwork(
  geneSet,
  networks,
  alpha,
  hopConstraint,
  nullDistribution,
  verbose = FALSE,
  topX = NULL
)
}
\arguments{
\item{geneSet}{A character vector of genes comprising the targets of interest.}

\item{networks}{A list of bipartite (PANDA-like) networks, where each network is a data frame with the following format:
tf,gene,score}

\item{alpha}{The significance cutoff for the statistical test.}

\item{hopConstraint}{The maximum number of hops to be considered between gene pairs.
Must be an even number.}

\item{nullDistribution}{The null distribution, specified as a vector of values.}

\item{verbose}{Whether or not to print detailed information about the run.}

\item{topX}{Select the X lowest significant p-values for each gene. NULL by default.}
}
\value{
A bipartite subnetwork in the same format as the original networks.
}
\description{
Find the subnetwork of significant edges connecting the genes.
}
