% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{CalculateTransitionMatricesFromFile}
\alias{CalculateTransitionMatricesFromFile}
\title{Calculates transition matrices using the following steps (to save memory)
for the real expression and each null expression matrix:
   1. Read in the column of the file corresponding to the null expression.
.   2. Center the null expression around the control values for a "noisy control".
.   3. Find the transition matrix between the noisy control and the cases.}
\usage{
CalculateTransitionMatricesFromFile(
  file,
  mode,
  expr,
  iterations,
  design,
  logging,
  numMaxCores,
  motif,
  ni_method,
  ni.coefficient.cutoff,
  alphaw,
  remove.diagonal,
  method,
  outputDir,
  by.tfs
)
}
\arguments{
\item{file}{The file containing the null networks}

\item{mode}{A parameter telling whether to build the regulatory networks ('buildNet') or to use provided regulatory networks
('regNet'). If set to 'regNet', then the parameters motif, ni_method, ni.coefficient.cutoff, and alphaw will be set to NA. Gene regulatory
networks are supplied in the 'expr' variable as a TF-by-Gene matrix, by concatenating the TF-by-Gene matrices of case and control, expr has size nTFs x 2nGenes.}

\item{expr}{The concatenated gene expression data / case and control networks}

\item{iterations}{Number of null models to generate}

\item{design}{The design matrix}

\item{logging}{Whether or not to print logging messages for MONSTER (default is TRUE)}

\item{numMaxCores}{requires doParallel, foreach.  Runs MONSTER in parallel computing 
environment.  Set to 1 to avoid parallelization, NA will take the default parallel pool in the computer.}

\item{motif}{Regulatory data.frame consisting of three columns.  For each row, a transcription factor (column 1) 
regulates a gene (column 2) with a defined strength (column 3), usually taken to be 0 or 1.
May also be NA, if MONSTER is being run on pre-computed gene regulatory networks, passed in the `expr` argument.
See also `domonster` to use this mode.}

\item{ni_method}{String to indicate algorithm method.  Must be one of "BERE","pearson",or "lda". Default is "BERE"}

\item{ni.coefficient.cutoff}{numeric to specify a p-value cutoff at the network
inference step.  Default is NA, indicating inclusion of all coefficients.}

\item{alphaw}{A weight parameter between 0 and 1 specifying proportion of weight 
to give to indirect compared to direct evidence. The default is 0.5 to give an 
equal weight to direct and indirect evidence.}

\item{remove.diagonal}{#' Logical for returning a result containing 0s across the diagonal (default = TRUE).}

\item{method}{character specifying which algorithm to use, default='ols'.}

\item{outputDir}{character vector specifying a directory or path in which 
which to save MONSTER results, default is NA and results are not saved.}
}
\value{
A list of matrices, each one corresponding to one null model
}
\description{
Calculates transition matrices using the following steps (to save memory)
for the real expression and each null expression matrix:
   1. Read in the column of the file corresponding to the null expression.
.   2. Center the null expression around the control values for a "noisy control".
.   3. Find the transition matrix between the noisy control and the cases.
}
