% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{checkDataType}
\alias{checkDataType}
\title{Checks that data is something MONSTER can handle}
\usage{
checkDataType(expr)
}
\arguments{
\item{expr}{Gene Expression dataset}
}
\value{
expr Gene Expression dataset in the proper form (may be the same as input)
}
\description{
This function runs the MONSTER algorithm
}

