% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{computeLinearRegression}
\alias{computeLinearRegression}
\title{Computes linear regression for both numeric and categorical phenotypes. Returns p-values
computed from moderated t-statistics using ebayes() and uses t-statistics as input to the
GSEA analysis.}
\usage{
computeLinearRegression(
  expression,
  phenotype,
  phenotype_dictionary,
  pathways,
  transform_expression
)
}
\arguments{
\item{expression}{: gene expression matrix (normalized, and filtered) 
with rows as genes and columns as samples.
Row and column names must be present.
Row names must be HGNC symbols.
Column names must match the row names of the phenotype matrix.}

\item{phenotype}{: phenotype matrix
with rows as samples and columns as phenotype variables.}

\item{phenotype_dictionary}{: a vector of strings
containing type of each phenotype.
Types can be either "numeric" or "categorical"}

\item{pathways}{: a list of pathways (e.g. KEGG, GO, Reactome etc. 
downloaded from http://www.gsea-msigdb.org/gsea/msigdb/human/collections.jsp)}

\item{transform_expression}{: Whether or not to transform gene expression data into logCPM.
This parameter is only used for the "linear" association method. Use if your data are raw RNA-seq
counts. Default is TRUE.}
}
\description{
Computes linear regression for both numeric and categorical phenotypes. Returns p-values
computed from moderated t-statistics using ebayes() and uses t-statistics as input to the
GSEA analysis.
}
