% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{ComputeRobustnessForOneEdgeType}
\alias{ComputeRobustnessForOneEdgeType}
\title{Helper function that computes the AUC score for either all edges or
for activating and inhibitory edges separately.}
\usage{
ComputeRobustnessForOneEdgeType(
  results,
  comparisons,
  metric = c("jaccard", "in-degree", "out-degree"),
  plotCurve = TRUE,
  xlab = "",
  ylab = "",
  mode = "Score",
  numberOfCutoffs = 10
)
}
\arguments{
\item{results}{An object of type FERRET_Results}

\item{comparisons}{An object of type FERRET_Comparisons. Note that the 
in-group and out-group names in FERRET_Comparisons MUST match the names
in FERRET_Results.}

\item{metric}{The metric to use for computing similarity. One or more elements
in the following list:
- jaccard: Computes the Jaccard similarity (intersection / union) of edges.
- in-degree: Computes the overlap in in-degree of each node.
- out-degree: Computes the overlap in out-degree of each node.}

\item{plotCurve}{Whether or not to generate a plot. Default is TRUE.}

\item{xlab}{The label for the X axis of the plot. Default is empty. This only
needs to be set when generating the plot.}

\item{ylab}{The label for the Y axis of the plot. Default is empty. This only
needs to be set when generating the plot.}

\item{mode}{"Percentile" or "Score".}

\item{numberOfCutoffs}{Number of cutoffs to evaluate.}
}
\value{
An AUC score, also generates a  plot as a side effect.
}
\description{
Helper function that computes the AUC score for either all edges or
for activating and inhibitory edges separately.
}
