% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{ComputeSimilarities}
\alias{ComputeSimilarities}
\title{Helper function to compute the similarities between the baseline network,
the ingroup, and the outgroup.}
\usage{
ComputeSimilarities(results, comparisons, cutoffs, metric)
}
\arguments{
\item{results}{Object of type FERRET_Results}

\item{comparisons}{Object of type FERRET_Comparisons}

\item{cutoffs}{The list of cutoffs to use when filtering the networks}

\item{metric}{The metric to use for computing similarity. An element
in the following list:
- jaccard: Computes the Jaccard similarity (intersection / union) of edges.
- in-degree: Computes the overlap in in-degree of each node.
- out-degree: Computes the overlap in out-degree of each node.}
}
\value{
Object of type FERRET_Similarity
}
\description{
Helper function to compute the similarities between the baseline network,
the ingroup, and the outgroup.
}
