% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FERRET.R
\name{ComputeSimilarityForGroup}
\alias{ComputeSimilarityForGroup}
\title{Helper function to compute the similarities within the ingroup or the outgroup}
\usage{
ComputeSimilarityForGroup(results, comparisons, metric)
}
\arguments{
\item{results}{Object of type FERRET_Results}

\item{comparisons}{The ingroup or outgroup slot from an object of type
FERRET_Comparisons.}

\item{metric}{The metric to use for computing similarity. An element
in the following list:
- jaccard: Computes the Jaccard similarity (intersection / union) of edges.
- in-degree: Computes the overlap in in-degree of each node.
- out-degree: Computes the overlap in out-degree of each node.}
}
\value{
A scalar similarity value.
}
\description{
Helper function to compute the similarities within the ingroup or the outgroup
}
