% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDOR.R
\name{condorPlotCommunities}
\alias{condorPlotCommunities}
\title{Plot adjacency matrix with links grouped and colored by community}
\usage{
condorPlotCommunities(
  condor.object,
  color_list,
  point.size = 0.01,
  xlab = "SNP",
  ylab = "Gene"
)
}
\arguments{
\item{condor.object}{output of either \code{\link{condorCluster}} or 
\code{\link{condorModularityMax}}}

\item{color_list}{vector of colors accepted by \code{col} inside the 
\code{\link[graphics]{plot}} function. There must be as many colors as 
communities.}

\item{point.size}{passed to \code{cex} in the 
\code{\link[graphics]{plot}}}

\item{xlab}{x axis label}

\item{ylab}{y axis label}
}
\value{
produces a \code{\link[graphics]{plot}} output.
}
\description{
This function will generate the network link 'heatmap' with colored dots
representing within-community links and black dots between-community 
links
}
\note{
For the condor paper \url{https://arxiv.org/abs/1509.02816}, I used
  35 colors from the "Tarnish" palette with "hard" clustering
}
\examples{
r = c(1,1,1,2,2,2,3,3,3,4,4);
b = c(1,2,3,1,2,4,2,3,4,3,4);
reds <- c("Alice","Sue","Janine","Mary")
blues <- c("Bob","John","Ed","Hank")
elist <- data.frame(red=reds[r],blue=blues[b])
condor.object <- createCondorObject(elist)
condor.object <- condorCluster(condor.object)
condorPlotCommunities(condor.object,
color_list=c("darkgreen","darkorange"),point.size=2,
xlab="Women",ylab="Men")
}
\references{
\url{https://medialab.github.io/iwanthue/} for 
 a nice color generator at
}
