% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.diff.panda.style.R
\name{create.diff.panda.style}
\alias{create.diff.panda.style}
\title{Create a Cytoscape visual style for two PANDA networks}
\usage{
create.diff.panda.style(style.name = "Diff.PandaStyle")
}
\arguments{
\item{style.name}{Character string indicating the style name. Defaults to "Diff.PandaStyle"}
}
\value{
a visual style in Cytoscape Control Panel under "Style" button.
}
\description{
This function is able to create a Cytoscape visual style for two PANDA networks output.
}
