% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPandaStyle.R
\name{createPandaStyle}
\alias{createPandaStyle}
\title{Create a Cytoscape visual style for PANDA network}
\usage{
createPandaStyle(style_name = "PandaStyle")
}
\arguments{
\item{style_name}{Character string indicating the style name. Defaults to "PandaStyle"}
}
\value{
A visual style in Cytoscape Control Panel under "Style" button.
}
\description{
This function is able to create a Cytoscape visual style for any PANDA network output.
}
\examples{
# Here we will load a customized visual style for our network, in which TF 
# nodes are orange circles, target gene nodes are blue squares, and edges 
# shade and width are the edge weight (likelyhood of regulatory interaction 
# between the TF and gene). You can further customize the network style 
# directly from Cytoscape.
\donttest{
createPandaStyle(style_name="PandaStyle")
}
}
