% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPandaStyle.R
\name{createPandaStyle}
\alias{createPandaStyle}
\title{Create a Cytoscape visual style for PANDA network}
\usage{
createPandaStyle(style.name = "PandaStyle")
}
\arguments{
\item{style.name}{Character string indicating the style name. Defaults to "PandaStyle"}
}
\value{
a visual style in Cytoscape Control Panel under "Style" button.
}
\description{
This function is able to create a Cytoscape visual style for any PANDA network output.
}
\examples{
# refer to the input datasets files of control TB dataset in inst/extdat as example
control_expression_file_path <- system.file("extdata", "expr10.txt", package = "netZoo", mustWork = TRUE)
motif_file_path <- system.file("extdata", "chip.txt", package = "netZoo", mustWork = TRUE)
ppi_file_path <- system.file("extdata", "ppi.txt", package = "netZoo", mustWork = TRUE)

# Run PANDA algorithm and access PANDA output
panda.net <- runPanda(e = control_expression_file_path, m = motif_file_path, ppi = ppi_file_path, rm_missing = TRUE)$panda

}
