% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{domonster}
\alias{domonster}
\title{MONSTER quick-start with pre-made regulatory networks}
\usage{
domonster(exp_graph, control_graph, nullPerms = 1000, numMaxCores = 3, ...)
}
\arguments{
\item{exp_graph}{matrix or PANDA object (generated by \code{netzooR::panda}); if matrix, should be the adjacency matrix for the network with regulators in rows and genes in columns, both named. This should be the network for the experimental (case) group.}

\item{control_graph}{matrix or PANDA object (generated by \code{netzooR::panda}); if matrix, should be the adjacency matrix for the network with regulators in rows and genes in columns, both named. This should be the network for the control (reference) group.}

\item{nullPerms}{numeric; defaults to 1000. Number of null permutations to perform. See \code{monster} for more details.}

\item{numMaxCores}{numeric; defaults to 3. Maximum number of cores to use; will be the minimum of this number and the actual available cores. See \code{monster} for more details.}

\item{...}{other arguments for \code{monster} may be passed.}
}
\value{
\code{monster} object
}
\description{
This function is a wrapper to simplify usage of the \code{monster} function in
the case where the pair of regulatory networks to be contrasted have already
been estimated, either as \code{panda} objects, or represented as an
adjacency matrix with regulators in rows and genes in columns.
}
\examples{

\donttest{

# Generating PANDA networks for demonstration:
# For the purposes of this example, first partition the pandaToyData samples, then perform panda:
pandaResult_exp <- panda(pandaToyData$motif, pandaToyData$expression[,1:25], pandaToyData$ppi)
pandaResult_control <- panda(pandaToyData$motif, pandaToyData$expression[,26:50], pandaToyData$ppi)

# function takes both panda objects and matrices, or a mixture
monster_res1 <- domonster(pandaResult_exp, pandaResult_control, numMaxCores = 1)
monster_res2 <- domonster(pandaResult_exp@regNet, pandaResult_control@regNet, numMaxCores = 1)
monster_res3 <- domonster(pandaResult_exp@regNet, pandaResult_control, numMaxCores = 1)
}
}
