% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAGON.R
\name{dragon}
\alias{dragon}
\title{Run DRAGON in R.}
\usage{
dragon(layer1, layer2, pval = F, gradient = "finite_difference")
}
\arguments{
\item{layer1}{: first layer of omics data}

\item{layer2}{: second layer of omics data}

\item{pval}{: calculate p-values for network edges}

\item{gradient}{: method for estimating parameters of p-value distribution. default = "finite_difference"; other option = "exact"}
}
\value{
cov

prec

ggm
}
\description{
Description: Estimates a multi-omic Gaussian graphical model for two input layers of paired omic data.
}
