% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRAGON.R
\name{dragon}
\alias{dragon}
\title{Run DRAGON in R.}
\usage{
dragon(
  layer1,
  layer2,
  pval = FALSE,
  gradient = "finite_difference",
  verbose = FALSE
)
}
\arguments{
\item{layer1}{: first layer of omics data; rows: samples (order must match layer2), columns: variables}

\item{layer2}{: second layer of omics data; rows: samples (order must match layer1), columns: variables.}

\item{pval}{: calculate p-values for network edges. Not yet implemented in R; available in netZooPy.}

\item{gradient}{: method for estimating parameters of p-value distribution, applies only if p-val == TRUE. default = "finite_difference"; other option = "exact"}

\item{verbose}{: verbosity level (TRUE/FALSE)}
}
\value{
A list of model results. cov : the shrunken covariance matrix
\itemize{
 \item{\code{cov}}{  the shrunken covariance matrix}
 \item{\code{prec}}{  the shrunken precision matrix}
 \item{\code{ggm}}{ the shrunken Gaussian graphical model; matrix of partial correlations. Self-edges (diagonal elements) are set to zero.}
 \item{\code{lambdas}}{  Vector of omics-specific tuning parameters (lambda1, lambda2) for \code{layer1} and \code{layer2}}
 \item{\code{gammas}}{  Reparameterized tuning parameters; gamma = 1 - lambda^2}
 \item{\code{risk_grid}}{  Risk grid, for assessing optimization. Grid boundaries are in terms of gamma.}
}
}
\description{
Description: Estimates a multi-omic Gaussian graphical model for two input layers of paired omic data.
}
