% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon.size-data.R
\docType{data}
\name{exon.size}
\alias{exon.size}
\title{Gene length}
\format{A integer vector of size 23459, with gene symbols as names}
\usage{
data(exon.size)
}
\description{
A vector of gene lengths. This will be used to normalize the gene mutation scores by the gene's length. This example is based on hg19 gene symbols. The gene length is based on the number of non-overlapping exons.
Data were downloaded and pre-processed as described in
\href{https://doi.org/10.1101/228031}{Kuijjer et al.}
}
\keyword{datasets}
