% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARN.R
\name{filterMissingGenes}
\alias{filterMissingGenes}
\title{Filter genes not expressed in any sample}
\usage{
filterMissingGenes(obj, threshold = 0)
}
\arguments{
\item{obj}{ExpressionSet object.}

\item{threshold}{Minimum sum of gene counts across samples -- defaults to zero.}
}
\value{
Filtered ExpressionSet object
}
\description{
The main use case for this function is the removal of missing genes.
}
