% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARN.R
\name{filterSamples}
\alias{filterSamples}
\title{Filter samples}
\usage{
filterSamples(obj, ids, groups = colnames(obj), keepOnly = FALSE)
}
\arguments{
\item{obj}{ExpressionSet object.}

\item{ids}{Names found within the groups labels corresponding to samples to be removed}

\item{groups}{Vector of labels for each sample or a column name of the phenoData slot
for the ids to filter. Default is the column names.}

\item{keepOnly}{Filter or keep only the samples with those labels.}
}
\value{
Filtered ExpressionSet object
}
\description{
Filter samples
}
