% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLOBFISH.R
\name{GenerateNullPANDADistribution}
\alias{GenerateNullPANDADistribution}
\title{Generate a null distribution of edge scores for PANDA-like networks; that is,
the set of edges where (1) the TF does not have a binding motif in the gene region,
(2) the TF does not form a complex with any other TF that has a binding motif in
the gene region, and (3) the genes regulated by the TF are not coexpressed with the
gene in question. We obtain this by inputting an empty prior and an identity coexpression
matrix.}
\usage{
GenerateNullPANDADistribution(
  ngenes = 20000,
  nTranscriptionFactors = 600,
  sampSize = 20,
  numberOfPandas = 10
)
}
\arguments{
\item{ngenes}{Number of genes to simulate}

\item{nTranscriptionFactors}{Number of transcription factors to simulate}

\item{sampSize}{Number of samples to simulate}

\item{numberOfPandas}{Number of null PANDA networks to generate}
}
\description{
Generate a null distribution of edge scores for PANDA-like networks; that is,
the set of edges where (1) the TF does not have a binding motif in the gene region,
(2) the TF does not form a complex with any other TF that has a binding motif in
the gene region, and (3) the genes regulated by the TF are not coexpressed with the
gene in question. We obtain this by inputting an empty prior and an identity coexpression
matrix.
}
