% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLOBFISH.R
\name{GenerateNullPANDADistribution}
\alias{GenerateNullPANDADistribution}
\title{Generate a null distribution of edge scores for PANDA-like networks; that is,
the set of edges where (1) the TF does not have a binding motif in the gene region,
(2) the TF does not form a complex with any other TF that has a binding motif in
the gene region, and (3) the genes regulated by the TF are not coexpressed with the
gene in question. We obtain this by inputting an empty prior and an identity coexpression
matrix.}
\usage{
GenerateNullPANDADistribution(
  ppiFile,
  motifFile,
  sampSize = 20,
  numberOfPandas = 10
)
}
\arguments{
\item{ppiFile}{The location of the protein-protein interaction network between transcription factors.
This should be a TSV file where the first two columns are the transcription
factors and the third is whether there is a PPI between them.}

\item{sampSize}{Number of samples to simulate}

\item{numberOfPandas}{Number of null PANDA networks to generate}

\item{motif}{The location of the motif prior from genes to transcription factors. This should
be a TSV file where the first column is the transcription factors, the
second is the genes, and the third is whether the transcription factor's
binding motif is in the gene promoter region.}
}
\description{
Generate a null distribution of edge scores for PANDA-like networks; that is,
the set of edges where (1) the TF does not have a binding motif in the gene region,
(2) the TF does not form a complex with any other TF that has a binding motif in
the gene region, and (3) the genes regulated by the TF are not coexpressed with the
gene in question. We obtain this by inputting an empty prior and an identity coexpression
matrix.
}
