% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{GeneratePermutationNull}
\alias{GeneratePermutationNull}
\title{Generates null models for MONSTER by permuting the expression levels
for the case and control data.}
\usage{
GeneratePermutationNull(expr, iterations, mode)
}
\arguments{
\item{expr}{Gene expression dataset}

\item{iterations}{Number of null models to generate}

\item{mode}{"buildNet" or "regNet"}
}
\value{
A list of matrices, each one corresponding to one null model
}
\description{
Generates null models for MONSTER by permuting the expression levels
for the case and control data.
}
