% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-data.R
\docType{data}
\name{genes}
\alias{genes}
\title{Example of a gene list}
\format{A character vector of length 2352}
\usage{
data(genes)
}
\description{
List of cancer-associated genes to subset the mutation data to, as described in
\href{https://doi.org/10.1101/228031}{Kuijjer et al.}
}
\keyword{datasets}
