% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{hcl.heatmap.plot}
\alias{hcl.heatmap.plot}
\title{Transformation matrix plot}
\usage{
hcl.heatmap.plot(x, method = "pearson")
}
\arguments{
\item{x}{monster Object}

\item{method}{distance metric for hierarchical clustering.
Default is "Pearson correlation"}
}
\value{
ggplot2 object for transition matrix heatmap
}
\description{
This function plots a hierachically clustered heatmap and 
corresponding dendrogram of a transaction matrix
}
\examples{
data(yeast)
cc.net <- monsterNI(yeast$motif,yeast$exp.cc)
sr.net <- monsterNI(yeast$motif,yeast$exp.sr)
transformation.matrix(cc.net, sr.net)
}
\keyword{keywords}

