% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_functions.R
\name{lioness}
\alias{lioness}
\title{Compute LIONESS (Linear Interpolation to Obtain Network Estimates for Single Samples)}
\usage{
lioness(motif, expr, ppi = NULL, network.inference.method, ...)
}
\arguments{
\item{motif}{A motif dataset, a data.frame, matrix or exprSet containing 3 columns.
Each row describes an motif associated with a transcription factor (column 1) a
gene (column 2) and a score (column 3) for the motif.}

\item{expr}{A mandatory expression dataset, as a genes (rows) by samples (columns) data.frame}

\item{ppi}{A Protein-Protein interaction dataset, a data.frame containing 3 columns.
Each row describes a protein-protein interaction between transcription factor 1(column 1),
transcription factor 2 (column 2) and a score (column 3) for the interaction.}

\item{network.inference.method}{String specifying choice of network inference method. Default is "panda".
Options include "pearson".}

\item{...}{additional arguments for panda analysis}
}
\value{
A list of length N, containing objects of class "panda" 
corresponding to each of the N samples in the expression data set.\cr
"regNet" is the regulatory network\cr
"coregNet" is the coregulatory network\cr
"coopNet" is the cooperative network
}
\description{
Compute LIONESS (Linear Interpolation to Obtain Network Estimates for Single Samples)
}
\examples{

# refer to the input datasets files of control in inst/extdat as example
control_expression_file_path <- system.file("extdata", "expr10_matched.txt", package = "netZooR", mustWork = TRUE)
motif_file_path <- system.file("extdata", "chip_matched.txt", package = "netZooR", mustWork = TRUE)
ppi_file_path <- system.file("extdata", "ppi_matched.txt", package = "netZooR", mustWork = TRUE)

# Run LIONESS algorithm
control_lioness_result <- lioness(e = control_expression_file_path, m = motif_file_path, ppi = ppi_file_path, rm_missing = TRUE,start_sample=1, end_sample=2)

}
\references{
Kuijjer, M.L., Tung, M., Yuan, G., Quackenbush, J. and Glass, K., 2015. 
Estimating sample-specific regulatory networks. arXiv preprint arXiv:1505.06440.
}
\keyword{keywords}

