\name{MONSTER-package}
\alias{MONSTER-package}
\alias{MONSTER}
\docType{package}
\title{
\packageTitle{MONSTER}
}
\description{
\packageDescription{MONSTER}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MONSTER}
\packageIndices{MONSTER}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MONSTER}

Maintainer: \packageMaintainer{MONSTER}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
##
}
