% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monster.dTFIPlot}
\alias{monster.dTFIPlot}
\title{This function plots the Off diagonal mass of an 
observed Transition Matrix compared to a set of null TMs}
\usage{
monster.dTFIPlot(
  monsterObj,
  rescale = FALSE,
  plot.title = NA,
  highlight.tfs = NA,
  nTFs = -1
)
}
\arguments{
\item{monsterObj}{monsterAnalysis Object}

\item{rescale}{logical indicating whether to reorder transcription
factors according to their statistical significance and to 
rescale the values observed to be standardized by the null
distribution}

\item{plot.title}{String specifying the plot title}

\item{highlight.tfs}{vector specifying a set of transcription 
factors to highlight in the plot}

\item{nTFs}{number of TFs to plot in x axis. -1 takes all TFs.}
}
\value{
ggplot2 object for transition matrix comparing observed 
distribution to that estimated under the null
}
\description{
This function plots the Off diagonal mass of an 
observed Transition Matrix compared to a set of null TMs
}
\examples{
# data(yeast)
# yeast$exp.cc[is.na(yeast$exp.cc)] <- mean(as.matrix(yeast$exp.cc),na.rm=TRUE)
# design <- c(rep(0,20),rep(NA,10),rep(1,20))
# monsterRes <- monster(yeast$exp.cc, design, yeast$motif, nullPerms=100, numMaxCores=4)#' 
data(monsterRes)
monster.dTFIPlot(monsterRes)
}
