% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monster.plot.monsterAnalysis}
\alias{monster.plot.monsterAnalysis}
\title{monster.plot.monsterAnalysis}
\usage{
monster.plot.monsterAnalysis(x, ...)
}
\arguments{
\item{x}{an object of class "monsterAnalysis"}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Plot of the dTFI for each TF against null distribution
}
\description{
plots the sum of squares of off diagonal mass (differential TF Involvement)
}
\examples{
data(yeast)
yeast$exp.cc[is.na(yeast$exp.cc)] <- mean(as.matrix(yeast$exp.cc),na.rm=T)
design <- c(rep(1,25),rep(0,10),rep(NA,15))
monsterRes <- monster(yeast$exp.cc, design,
yeast$motif, nullPerms=10, numMaxCores=1)
monster.plot.monsterAnalysis(monsterRes)
}
