% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monster}
\alias{monster}
\title{MOdeling Network State Transitions from Expression and Regulatory data (MONSTER)}
\usage{
monster(expr, design, motif = NULL, nullPerms = 100, numMaxCores = 1,
  outputDir = NA, ...)
}
\arguments{
\item{expr}{Gene Expression dataset}

\item{design}{Binary vector indicating case control partition}

\item{motif}{Regulatory data.frame}

\item{nullPerms}{number of random permutations to run (default 100).  Set to 0 to only 
calculate observed transition matrix}

\item{numMaxCores}{requires doParallel, foreach.  Runs MONSTER in parallel computing 
environment.  Set to 1 to avoid parallelization.}

\item{outputDir}{save MONSTER results in a directory}
}
\value{
An object of class "monster" containing results
}
\description{
This function runs the MONSTER algorithm
}
\examples{
data(yeast)
design <- c(rep(0,ncol(yeast$exp.cc)), rep(1,ncol(yeast$exp.sr)))
monsterRes <- monster(cbind(yeast$exp.cc, yeast$exp.sr), 
    design, yeast$motif, nullPerms=10, numMaxCores=4)
plot(monsterRes)


}
\keyword{keywords}

