% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monster}
\alias{monster}
\title{MOdeling Network State Transitions from Expression and Regulatory data (MONSTER)}
\usage{
monster(
  expr,
  design,
  motif,
  nullPerms = 100,
  ni_method = "BERE",
  ni.coefficient.cutoff = NA,
  numMaxCores = 1,
  outputDir = NA,
  alphaw = 0.5,
  mode = "buildNet",
  method = "ols",
  remove.diagonal = TRUE,
  logging = TRUE
)
}
\arguments{
\item{expr}{Gene Expression dataset, can be matrix or data.frame of expression values or ExpressionSet. 
If `mode` is set to `regNet`, MONSTER will be run in pre-computed gene regulatory network mode, in which case
gene regulatory networks can be passed for this argument. See also `domonster` to use this mode.}

\item{design}{Binary vector indicating case control partition. 1 for case and 0 for control.}

\item{motif}{Regulatory data.frame consisting of three columns.  For each row, a transcription factor (column 1) 
regulates a gene (column 2) with a defined strength (column 3), usually taken to be 0 or 1.
May also be NA, if MONSTER is being run on pre-computed gene regulatory networks, passed in the `expr` argument.
See also `domonster` to use this mode.}

\item{nullPerms}{number of random permutations to run (default 100).  Set to 0 to only 
calculate observed transition matrix. When mode is is 'buildNet' it randomly permutes the case and control expression
samples, if mode is 'regNet' it will randomly permute the case and control networks.}

\item{ni_method}{String to indicate algorithm method.  Must be one of "BERE","pearson",or "lda". Default is "BERE"}

\item{ni.coefficient.cutoff}{numeric to specify a p-value cutoff at the network
inference step.  Default is NA, indicating inclusion of all coefficients.}

\item{numMaxCores}{requires doParallel, foreach.  Runs MONSTER in parallel computing 
environment.  Set to 1 to avoid parallelization, NA will take the default parallel pool in the computer.}

\item{outputDir}{character vector specifying a directory or path in which 
which to save MONSTER results, default is NA and results are not saved.}

\item{alphaw}{A weight parameter between 0 and 1 specifying proportion of weight 
to give to indirect compared to direct evidence. The default is 0.5 to give an 
equal weight to direct and indirect evidence.}

\item{mode}{A parameter telling whether to build the regulatory networks ('buildNet') or to use provided regulatory networks
('regNet'). If set to 'regNet', then the parameters motif, ni_method, ni.coefficient.cutoff, and alphaw will be set to NA. Gene regulatory
networks are supplied in the 'expr' variable as a TF-by-Gene matrix, by concatenating the TF-by-Gene matrices of case and control, expr has size nTFs x 2nGenes.}

\item{method}{Method to use in computing the transition matrix. These include "ols" (default),"kabsch","L1", and "orig" (SVD).}

\item{remove.diagonal}{#' Logical for returning a result containing 0s across the diagonal (default = TRUE).}

\item{logging}{Whether or not to print logging messages for MONSTER (default is TRUE)}
}
\value{
An object of class "monsterAnalysis" containing results
}
\description{
This function runs the MONSTER algorithm.  Biological states are characterized by distinct patterns 
of gene expression that reflect each phenotype's active cellular processes. 
Driving these phenotypes are gene regulatory networks in which transcriptions factors control 
when and to what degree individual genes are expressed. Phenotypic transitions, such as those that 
occur when disease arises from healthy tissue, are associated with changes in these  networks. 
MONSTER is an approach to understanding these transitions. MONSTER models phenotypic-specific 
regulatory networks and then estimates a "transition matrix" that converts one state to another. 
By examining the properties of the transition matrix, we can gain insight into regulatory 
changes associated with phenotypic state transition.
Important note: the direct regulatory network observed from gene expression is currently
implemented as a regular correlation as opposed to the partial correlation described 
in the paper.
There are 2 modes to run MONSTER:
(1) MONSTER can internally estimate a gene regulatory network, in which case 
it will take as input an expression data plus a motif network
(2) MONSTER can run on pre-computed gene regulatory networks (e.g., from PANDA),
in which case the `motif` argument is set to `NA`, and the `expr` argument will
contain the concatenated gene regulatory networks. This mode can be selected
by setting the `mode` argument to `regNet`.
Alternatively, see the `domonster` function for a quick-start way to run this mode.
Citation: Schlauch, Daniel, et al. "Estimating drivers of cell state transitions using gene regulatory network models." 
BMC systems biology 11.1 (2017): 139. https://doi.org/10.1186/s12918-017-0517-y
}
\examples{
# Example with the network reconstruction step
data(yeast)
design <- c(rep(0,20),rep(NA,10),rep(1,20))
yeast$exp.cc[is.na(yeast$exp.cc)] <- mean(as.matrix(yeast$exp.cc),na.rm=TRUE)
# Example with provided networks
\donttest{
pandaResult <- panda(pandaToyData$motif, pandaToyData$expression, pandaToyData$ppi)
case=pandaResult@regNet
nelemReg=dim(pandaResult@regNet)[1]*dim(pandaResult@regNet)[2]
nGenes=length(colnames(pandaResult@regNet))
control=matrix(rexp(nelemReg, rate=.1), ncol=nGenes)
colnames(control) = colnames(case)
rownames(control) = rownames(case) 
expr = as.data.frame(cbind(control,case))
design=c(rep(0,nGenes),rep(1, nGenes))
monsterRes <- monster(expr, design, motif=NA, nullPerms=10, numMaxCores=1, mode='regNet')
}
# alternatively, if a gene regulatory network has already been estimated,
# see the domonster function for quick start
}
