% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monsterCalculateTmPValues}
\alias{monsterCalculateTmPValues}
\title{Calculate p-values for a tranformation matrix}
\usage{
monsterCalculateTmPValues(monsterObj, method = "z-score")
}
\arguments{
\item{monsterObj}{monsterAnalysis Object}

\item{method}{one of 'z-score' or 'non-parametric'}
}
\value{
vector of p-values for each transcription factor
}
\description{
This function calculates the significance of an observed
transition matrix given a set of null transition matrices
}
\examples{
# data(yeast)
# design <- c(rep(0,20),rep(NA,10),rep(1,20))
# yeast$exp.cc[is.na(yeast$exp.cc)] <- mean(as.matrix(yeast$exp.cc),na.rm=TRUE)
# monsterRes <- monster(yeast$exp.cc, design, yeast$motif, nullPerms=100, numMaxCores=4)
data(monsterRes)
monsterCalculateTmPValues(monsterRes)
}
