% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monsterCalculateTmStats}
\alias{monsterCalculateTmStats}
\title{Calculate statistics for a tranformation matrix}
\usage{
monsterCalculateTmStats(monsterObj, method = "z-score")
}
\arguments{
\item{monsterObj}{monsterAnalysis Object}

\item{method}{one of 'z-score' or 'non-parametric'}
}
\value{
p-values, t-values, off-diagonal squared mass for the observed transition matrix,
and the null off-diagonal squared mass matrix
}
\description{
This function powers both the p-value and t-value calculations
for a transformation matrix.  It calculates the off-diagonal squared mass
for each transition matrix, and then calculates the p-values and t-values
for the observed transition matrix compared to the null transition matrices.
It is used by the monsterCalculateTmPValues function and by the monsterdTFIPlot
}
\examples{
# data(yeast)
# design <- c(rep(0,20),rep(NA,10),rep(1,20))
# yeast$exp.cc[is.na(yeast$exp.cc)] <- mean(as.matrix(yeast$exp.cc),na.rm=TRUE)
# monsterRes <- monster(yeast$exp.cc, design, yeast$motif, nullPerms=100, numMaxCores=4)
data(monsterRes)
monsterCalculateTmStats(monsterRes)
}
