% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONSTER.R
\name{monsterHclHeatmapPlot}
\alias{monsterHclHeatmapPlot}
\title{Transformation matrix plot}
\usage{
monsterHclHeatmapPlot(monsterObj, method = "pearson")
}
\arguments{
\item{monsterObj}{monsterAnalysis Object}

\item{method}{distance metric for hierarchical clustering.
Default is "Pearson correlation"}
}
\value{
ggplot2 object for transition matrix heatmap
}
\description{
This function plots a hierachically clustered heatmap and 
corresponding dendrogram of a transaction matrix
}
\examples{
# data(yeast)
# design <- c(rep(0,20),rep(NA,10),rep(1,20))
# yeast$exp.cc[is.na(yeast$exp.cc)] <- mean(as.matrix(yeast$exp.cc),na.rm=TRUE)
# monsterRes <- monster(yeast$exp.cc, design, yeast$motif, nullPerms=10, numMaxCores=1)
data(monsterRes)
monsterHclHeatmapPlot(monsterRes)
}
