\name{monsterRes}
\docType{data}
\alias{monsterRes}
\title{MONSTER results from example cell-cycle yeast transition}
\description{
This data contains the MONSTER result from analysis of Yeast Cell cycle, included in data(yeast).  This result arbitrarily takes the first 20 gene expression samples in yeast$cc to be the baseline condition, and the final 20 samples to be the final condition.
}
\usage{
data(monsterRes)
}
\format{MONSTER obj}
\value{MONSTER obj}
\keyword{datasets}
