% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMBARFunctions.R
\docType{data}
\name{mut.ucec}
\alias{mut.ucec}
\title{Example of mutation data}
\format{A table with 248 rows and 19754 columns}
\usage{
data(mut.ucec)
}
\description{
Somatic mutations of Uterine Corpus Endometrial Carcinoma from The Cancer Genome Atlas.
Data were downloaded and pre-processed as described in
\href{https://doi.org/10.1101/228031}{Kuijjer et al.}
}
\keyword{datasets}
