% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{networkANOVA}
\alias{networkANOVA}
\title{Function to run ANOVA on nominal phenotypes with continuous network features.}
\usage{
networkANOVA(network, phenotype)
}
\arguments{
\item{network}{: A weighted matrix where rows are samples and columns are network features.}

\item{pheno}{: phenotype matrix
with rows as samples and columns as phenotype variables.}
}
\description{
Function to run ANOVA on nominal phenotypes with continuous network features.
}
