% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTTER.R
\name{otter}
\alias{otter}
\title{Run OTTER in R}
\usage{
otter(W, P, C, lambda = 0.0035, gamma = 0.335, Iter = 300,
  eta = 1e-05, bexp = 1)
}
\arguments{
\item{W}{: TF-gene regulatory network based on TF motifs as a
matrix of size (t,g), g=number of genes, t=number of TFs}

\item{P}{: TF-TF protein interaction network as a matrix of size (t,t)}

\item{C}{: gene coexpression as a matrix of size (g,g)}

\item{gamma}{: regularization parameter}

\item{Iter}{: number of iterations of the algorithm}

\item{eta}{: learning rate}

\item{bexp}{: exponent influencing learning rate (higher means smaller)}

\item{lambda:}{tuning parameter in [0,1] (higher gives more weight to C)}
}
\value{
W    : Predicted TF-gene complete regulatory network as an adjacency matrix of size (t,g).
}
\description{
Description:
              OTTER infers gene regulatory networks using TF DNA binding
              motif (W), TF PPI (P), and gene coexpression (C) through 
              minimzing the following objective:
                                 min f(W) 
              with f(W) = (1-lambda)*||WW' - P||^2 + lambda*||W'W - C||^2 + (gamma/2)*||W||^2
              
Authors: 
              Rebekka Burkholz 4/2020
}
\examples{

W=matrix(rexp(100, rate=.1), ncol=10)
C=matrix(rexp(100, rate=.1), ncol=10)
P=matrix(rexp(100, rate=.1), ncol=10)

# Run OTTER algorithm
W <- otter(W, P, C)
 
}
