% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{phenotype_anova}
\alias{phenotype_anova}
\title{Function to compute ANOVA statistics between nominal and continuous phenotypes.}
\usage{
phenotype_anova(
  phenotype,
  phenotypesToCompare,
  phenotypeType,
  phenotypesToCompareType
)
}
\arguments{
\item{phenotype}{The phenotype vector to test.}

\item{phenotypesToCompare}{A data frame containing all phenotypes against which to run associations.}

\item{phenotypeType}{Whether the phenotype is nominal or continuous.}

\item{phenotypesToCompareType}{A vector of types (nominal, continuous) corresponding to all
phenotypes against which to compare.}
}
\value{
A data frame with two vectors: "cor", which lists the ANOVA p-values, 
and "V" which is set to NA.
}
\description{
Function to compute ANOVA statistics between nominal and continuous phenotypes.
}
