% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{phenotype_chisq}
\alias{phenotype_chisq}
\title{Function to run associations between categorical (dichotomous or nominal) phenotypes.
If at least 5 samples exist in each group, run a Chi-square test and compute Cramer's V.
Otherwise, run a Fisher-Freeman-Halton Test.}
\usage{
phenotype_chisq(
  phenotype,
  phenotypesToCompare,
  phenotypeType,
  phenotypesToCompareType
)
}
\arguments{
\item{phenotype}{The phenotype vector to test.}

\item{phenotypesToCompare}{A data frame containing all phenotypes against which to run associations.}

\item{phenotypeType}{Whether the phenotype is dichotomous or nominal.}

\item{phenotypesToCompareType}{A vector of types (dichotomous, nominal) corresponding to all
phenotypes against which to compare.}
}
\value{
A data frame with two vectors: "cor", which lists the Chi-square or Fisher-Freeman-Halton Test p-values, 
and "V" which lists the Cramer's V statistics (will be NA if Fisher-Freeman-Halton Test is computed)
}
\description{
Function to run associations between categorical (dichotomous or nominal) phenotypes.
If at least 5 samples exist in each group, run a Chi-square test and compute Cramer's V.
Otherwise, run a Fisher-Freeman-Halton Test.
}
