% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{phenotype_ffs}
\alias{phenotype_ffs}
\title{Function to run associations between categorical (dichotomous or nominal) phenotypes using a Fisher-Freeman-Halton Test.}
\usage{
phenotype_ffs(
  phenotype,
  phenotypesToCompare,
  phenotypeType,
  phenotypesToCompareType
)
}
\arguments{
\item{phenotype}{The phenotype vector to test.}

\item{phenotypesToCompare}{A data frame containing all phenotypes against which to run associations.}

\item{phenotypeType}{Whether the phenotype is dichotomous or nominal.}

\item{phenotypesToCompareType}{A vector of types (dichotomous, nominal) corresponding to all
phenotypes against which to compare.}
}
\value{
A data frame with two vectors: "cor", which lists Fisher-Freeman-Halton Test p-values, 
and "V" which lists the Cramer's V statistics (NA)
}
\description{
Function to run associations between categorical (dichotomous or nominal) phenotypes using a Fisher-Freeman-Halton Test.
}
