% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{phenotype_ffs}
\alias{phenotype_ffs}
\title{Function to run associations between categorical (dichotomous or nominal) phenotypes using a Fisher-Freeman-Halton Test.}
\usage{
phenotype_ffs(tables)
}
\arguments{
\item{tables}{The list of tables to test, named by the phenotypes against which to compare.}
}
\value{
A data frame with two vectors: "cor", which lists Fisher-Freeman-Halton Test p-values, 
and "V" which lists the Cramer's V statistics (NA)
}
\description{
Function to run associations between categorical (dichotomous or nominal) phenotypes using a Fisher-Freeman-Halton Test.
}
