% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{phenotype_ttest}
\alias{phenotype_ttest}
\title{Function to compute Welch's t-test statistics between dichotomous and continuous phenotypes.}
\usage{
phenotype_ttest(
  phenotype,
  phenotypesToCompare,
  phenotypeType,
  phenotypesToCompareType
)
}
\arguments{
\item{phenotype}{The phenotype vector to test.}

\item{phenotypesToCompare}{A data frame containing all phenotypes against which to run associations.}

\item{phenotypeType}{Whether the phenotype is dichotomous or continuous.}

\item{phenotypesToCompareType}{A vector of types (dichotomous, continuous) corresponding to all
phenotypes against which to compare.}
}
\value{
A data frame with two vectors: "cor", which lists the t-test p-values, 
and "V" which is set to NA.
}
\description{
Function to compute Welch's t-test statistics between dichotomous and continuous phenotypes.
}
