% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARN.R
\name{plotCMDS}
\alias{plotCMDS}
\title{Plot classical MDS of dataset}
\usage{
plotCMDS(
  obj,
  comp = 1:2,
  normalized = FALSE,
  distFun = dist,
  distMethod = "euclidian",
  n = NULL,
  samples = TRUE,
  log = TRUE,
  plotFlag = TRUE,
  ...
)
}
\arguments{
\item{obj}{ExpressionSet object or objrix.}

\item{comp}{Which components to display.}

\item{normalized}{TRUE / FALSE, use the normalized matrix or raw counts.}

\item{distFun}{Distance function, default is dist.}

\item{distMethod}{The distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given.}

\item{n}{Number of features to make use of in calculating your distances.}

\item{samples}{Perform on samples or genes.}

\item{log}{TRUE/FALSE log2-transform raw counts.}

\item{plotFlag}{TRUE/FALSE whether to plot or not.}

\item{...}{Additional plot arguments.}
}
\value{
coordinates
}
\description{
This function plots the MDS coordinates for the "n" features of interest. Potentially uncovering batch
effects or feature relationships.
}
\examples{
\donttest{
u <- 'https://netzoo.s3.us-east-2.amazonaws.com/netZooR/unittest_datasets/yarn/'
skin <- tryCatch({tmp<-tempfile()
download.file(paste0(u,'skin.rdata'),tmp,quiet=TRUE)
local({load(tmp);get('skin')})}, error=function(e) NULL)
if(!is.null(skin)) res <- plotCMDS(skin,pch=21,bg=factor(pData(skin)$SMTSD))
}
}
