% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPANDAinCytoscape.R
\name{plotPANDAinCytoscape}
\alias{plotPANDAinCytoscape}
\title{Plot PANDA network in Cytoscape}
\usage{
plotPANDAinCytoscape(panda.net, network.name = "PANDA")
}
\arguments{
\item{panda.net}{Character string indicating the input PANDA network in data frame structure type.}

\item{network.name}{Character string indicating the name of Cytoscape network.}
}
\description{
This function is able to modify PANDA network and plot in Cytoscape.
}
\examples{
# refer to the input datasets files of control TB dataset in inst/extdat as example
control_expression_file_path <- system.file("extdata", "expr10.txt", package = "netZoo", mustWork = TRUE)
motif_file_path <- system.file("extdata", "chip.txt", package = "netZoo", mustWork = TRUE)
ppi_file_path <- system.file("extdata", "ppi.txt", package = "netZoo", mustWork = TRUE)

# Run PANDA algorithm
control_all_panda_result <- runPanda(e = control_expression_file_path, m = motif_file_path, ppi = ppi_file_path, rm_missing = TRUE )

# access PANDA regulatory network
control_net <- control_all_panda_result$panda

# select top 1000 edges in PANDA network by edge weight.
panda.net <- head(control_net[order(control_net$force,decreasing = T),], 1000)

# run this function to create a network in Cytoscape.
plotPANDAinCytoscape(panda.net, network.name="PANDA")

}
