% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGRET.R
\name{runEgret}
\alias{runEgret}
\title{Run EGRET in R}
\usage{
runEgret(b, v, q, m, e, p, g, t)
}
\arguments{
\item{b}{: Data frame of eQTL data, each row containing an eQTL which exist within motif regions adjacent to the eGene, with columns
TF, gene, variant position,variant chromosome, eQTL beta value.}

\item{v}{: Data frame of VCF file containing SNPs of the individual in question}

\item{q}{: Data frame of QBiC predictions of the effect of eQTL variants on TF binding. Each row represents an 
eQTL variant with a predicted negative (disruptive) effect on the binding of the TF corresponding to the motif in which the eQTL variant 
resides. Colums are: eQTL variant as chr[chrNum]_position, TF, adjacent eGene, QBiC binding effect size and QBiC binding effect (should be negative)}

\item{m}{: Motif prior data frame. Each row represents an edge in the bipartite motif prior, with columns TF, gene and edge weight. 
The edge weight should be 1 or 0 based on the presence/absence of the TF motif in the promoter region of the gene.}

\item{e}{: Gene expression data frame in which each row represents a gene and each column represents the expression of that gene in a sample.
The first column should contain gene IDs.}

\item{p}{: PPI network data frame. Each row represents an edgem with columns TF, TF and interaction weight.}

\item{g}{: Data frame mapping gene names to gene ids, with columns containing the gene ID the corresponding gene name.}

\item{t}{: A string containing a name for the EGRET run. Output files will be labelled with this tag.

Outputs:}
}
\value{
EGRET    : Predicted genotye-specific gene regulatory network saved as tag_egret.RData

BASELINE : A Baseline (PANDA) genotype-agnostic gene regulatory network saved as tag_panda.RData
}
\description{
Description:
              NOTE: Beta version. EGRET infers individual-specific gene regulatory networks using inidividual level data - 
              a genotype vcf file (v) and QBiC binding predictions (q) -  as well as population/reference level data - 
               eQTLs (b), a motif-gene prior (m), PPI network (p), and gene expression (e). An annotation file g is also used to
               map TF names to their corresponding ensemble ids.
}
\details{
Inputs:
}
\examples{

# Run EGRET algorithm
toy_qbic_path <- system.file("extdata", "toy_qbic.txt", package = "netZooR", mustWork = TRUE)
toy_genotype_path <- system.file("extdata", "toy_genotype.vcf", package = "netZooR", mustWork = TRUE)
toy_motif_path <- system.file("extdata", "toy_motif_prior.txt", package = "netZooR", mustWork = TRUE)
toy_expr_path <- system.file("extdata", "toy_expr.txt", package = "netZooR", mustWork = TRUE)
toy_ppi_path <- system.file("extdata", "toy_ppi_prior.txt", package = "netZooR", mustWork = TRUE)
toy_eqtl_path <- system.file("extdata", "toy_eQTL.txt", package = "netZooR", mustWork = TRUE)
toy_map_path <- system.file("extdata", "toy_map.txt", package = "netZooR", mustWork = TRUE)
qbic <- read.table(file = toy_qbic_path, header = FALSE)
vcf <- read.table(toy_genotype_path, header = FALSE, sep = "\t", stringsAsFactors = FALSE, colClasses = c("character", "numeric", "character", "character", "character", "character", "character", "character", "character", "character"))
motif <- read.table(toy_motif_path, sep = "\t", header = FALSE)
expr <- read.table(toy_expr_path, header = FALSE, sep = "\t", row.names = 1)
ppi <- read.table(toy_ppi_path, header = FALSE, sep = "\t")
qtl <- read.table(toy_eqtl_path, header = FALSE)
nameGeneMap <- read.table(toy_map_path, header = FALSE)
tag <- "my_toy_egret_run"
\donttest{
runEgret(qtl,vcf,qbic,motif,expr,ppi,nameGeneMap,tag)
}
}
