% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{seahorse}
\alias{seahorse}
\title{Run SEAHORSE in R
Author(s): Enakshi Saha
Description:
              SEAHORSE computes gene-gene coexpression matrix,
              associations between a set of given phenotypes and each gene,
              performs gene set enrichment analysis (GSEA) for each phenotype,
              using the measures of association.
              GSEA is performed through R package "fgsea".}
\usage{
seahorse(
  expression,
  phenotype,
  phenotype_dictionary,
  pathways,
  compute_cor = TRUE,
  assoc_method = "pearson",
  transform_expression = TRUE,
  pval_adj_method = "none"
)
}
\arguments{
\item{expression}{: gene expression matrix (normalized, and filtered) 
with rows as genes and columns as samples.
Row and column names must be present.
Row names must be HGNC symbols.
Column names must match the row names of the phenotype matrix.}

\item{phenotype}{: phenotype matrix
with rows as samples and columns as phenotype variables.}

\item{phenotype_dictionary}{: a vector of strings
containing type of each phenotype.
Types can be "dichotomous", "nominal", or "continuous"}

\item{pathways}{: a list of pathways (e.g. KEGG, GO, Reactome etc. 
downloaded from http://www.gsea-msigdb.org/gsea/msigdb/human/collections.jsp)}

\item{compute_cor}{: Whether or not to compute the correlation matrix. Default is TRUE.}

\item{assoc_method}{: The method used to infer associations between phenotypes and genes. Default
is "pearson". Other options are "spearman", "kendall", and "linear". The "pearson", "kendall", and "spearman" options
compute correlations between each phenotype and gene independently for numeric phenotypes and
compute an ANOVA for categorical phenotypes. linear" computes a linear regression
model of the form gene ~ phenotype1 + phenotype2 + ... + phenotypeN and returns the p-values
for each coefficient.}

\item{transform_expression}{: Whether or not to transform gene expression data into logCPM.
This parameter is only used for the "linear" association method. Use if your data are raw RNA-seq
counts. Default is TRUE.}

\item{pval_adj_method}{Wrapper for p.adjust. Defaults to "none" (no adjustment). Other options are
"holm", "hochberg", "hommel", "bonferroni", "BH", "BY", and "fdr".
Outputs:}
}
\value{
results    : a list containing three objects
        results$coexpression: a gene x gene Pearson correlation matrix.
        results$phenotype_association : a list containing a vector for each phenotype
        results$GSEA: a list containing a matrix of GSEA results for each phenotype
}
\description{
Run SEAHORSE in R
Author(s): Enakshi Saha
Description:
              SEAHORSE computes gene-gene coexpression matrix,
              associations between a set of given phenotypes and each gene,
              performs gene set enrichment analysis (GSEA) for each phenotype,
              using the measures of association.
              GSEA is performed through R package "fgsea".
}
\examples{

expression_data = data.frame(matrix(rexp(200, rate=.1), ncol=10, nrow = 20))
rownames(expression_data) = paste("gene", 1:20, sep = "")
colnames(expression_data) = paste("sample", 1:10, sep = "")

phenotype_data = data.frame(matrix(0, ncol=2, nrow = 10))
colnames(phenotype_data) = c("sex", "height")
rownames(phenotype_data) = colnames(expression_data)
phenotype_data$sex = c(rep("male", nrow(phenotype_data)/2), rep("female", nrow(phenotype_data)/2))
phenotype_data$height = 65 + sample.int(10, nrow(phenotype_data), replace = TRUE)

phenotype_dictionary = c("dichotomous", "continuous")

pathways = list()
pathways$pathway1 = sample(rownames(expression_data), 5)
pathways$pathway2 = sample(rownames(expression_data), 3)
pathways$pathway3 = sample(rownames(expression_data), 7)

# Run seahorse
results <- seahorse(expression_data, phenotype_data, phenotype_dictionary, pathways)
 
}
