% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{seahorseFormatForUI}
\alias{seahorseFormatForUI}
\title{Generates the following files which are needed for use with the SEAHORSE UI:
- all_gsea_results.tsv.gz: A list of pathway enrichment scores with respect to all
phenotypes and all files, formatted as pathway	pval	padj	log2err	ES	NES	size	ranks	leadingEdge	varname	tissue
- data_dictionary.tsv.gz: A list of phenotype names, descriptions, and variable types formatted as
VARNAME VARDESC VARMETA TYPE
- geneexpression2geneexpression.tsv.gz: Gene-gene association statistics, formatted as Gene A	Gene B	Tissue	Correlation
- geneexpression_data.tsv.gz: Gene expression data, formatted as ENSG    SAMPID  GENE_EXPRESSION
- human_ensembl2symbol_map.tsv.gz: A mapping from Ensembl IDs to HGNC symbols and Entrez IDs,
formatted as ALIAS	ENSEMBL	SYMBOL	ENTREZID
- metadata.tsv.gz: The phenotype data, formatted as SAMPID	tissue	VARNAME	VALUE
- metadata2expression.tsv.gz: Gene-phenotype association statistics, formatted as 
VARNAME GENE    tissue  TEST    TESTSTAT        TESTPVALUE
- metadata2metadata.tsv.gz: Phenotype-phenotype association statistics, formatted as 
VARNAME1        VARNAME2        tissue  TEST    TESTSTAT        TESTPVALUE}
\usage{
seahorseFormatForUI(
  input_directory,
  result_directory,
  output_directory,
  data_dictionary,
  pathways,
  geneGeneCutoff = -2
)
}
\arguments{
\item{input_directory}{: A directory containing all inputs to SEAHORSE, stored as RDS files,
where the input values contain slots for expression, phenotype, and phenotype dictionary.}

\item{result_directory}{: A directory containing all SEAHORSE results from the expression data,
phenotype data, phenotype dictionary, and pathways provided, stored as RDS files. Files must follow
the same naming convention as in the result directory.}

\item{output_directory}{: The directory where you wish to store the files generated by this function.}

\item{data_dictionary}{: The data dictionary.}

\item{pathways}{: The pathways used to run SEAHORSE}

\item{geneGeneCutoff}{: The correlation cutoff for gene-gene associations. Only correlations greater
than this cutoff will be saved.}
}
\description{
Generates the following files which are needed for use with the SEAHORSE UI:
- all_gsea_results.tsv.gz: A list of pathway enrichment scores with respect to all
phenotypes and all files, formatted as pathway	pval	padj	log2err	ES	NES	size	ranks	leadingEdge	varname	tissue
- data_dictionary.tsv.gz: A list of phenotype names, descriptions, and variable types formatted as
VARNAME VARDESC VARMETA TYPE
- geneexpression2geneexpression.tsv.gz: Gene-gene association statistics, formatted as Gene A	Gene B	Tissue	Correlation
- geneexpression_data.tsv.gz: Gene expression data, formatted as ENSG    SAMPID  GENE_EXPRESSION
- human_ensembl2symbol_map.tsv.gz: A mapping from Ensembl IDs to HGNC symbols and Entrez IDs,
formatted as ALIAS	ENSEMBL	SYMBOL	ENTREZID
- metadata.tsv.gz: The phenotype data, formatted as SAMPID	tissue	VARNAME	VALUE
- metadata2expression.tsv.gz: Gene-phenotype association statistics, formatted as 
VARNAME GENE    tissue  TEST    TESTSTAT        TESTPVALUE
- metadata2metadata.tsv.gz: Phenotype-phenotype association statistics, formatted as 
VARNAME1        VARNAME2        tissue  TEST    TESTSTAT        TESTPVALUE
}
