% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARN.R
\docType{data}
\name{skin}
\alias{skin}
\title{Skin RNA-seq data from the GTEx consortium}
\format{
An object of class \code{"ExpressionSet"}, see \code{\linkS4class{ExpressionSet}}.
}
\source{
GTEx Portal
}
\usage{
data(skin)
}
\value{
ExpressionSet object
}
\description{
Skin RNA-seq data from the GTEx consortium. V6 release. Random selection of 20 skin samples.
13 of the samples are fibroblast cells, 5 Skin sun exposed, 2 sun unexposed.
}
\examples{
\donttest{data(skin);
checkMisAnnotation(skin,"GENDER");}
}
\references{
GTEx Consortium, 2015. The Genotype-Tissue Expression (GTEx) pilot analysis: Multitissue gene regulation in humans. Science, 348(6235), pp.648-660.
(\href{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC4547484/}{PubMed})
}
\keyword{datasets}
