% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{small1976}
\alias{small1976}
\title{Pollinator-plant interactions}
\format{A data frame with 442 rows and 3 variables}
\usage{
data(small1976)
}
\description{
A dataset containing the number of interactions 34 plants and 13 pollinators. The variables are as follows:
}
\details{
\itemize{
  \item pollinator. Species name of insect pollinator
  \item plant. Species name of plant
  \item interactions. Number of visitors caught on each plant species
}
}
\references{
\url{https://www.nceas.ucsb.edu/interactionweb/html/small_1976.html}
}
\keyword{datasets}

