% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcePPI.R
\name{sourcePPI}
\alias{sourcePPI}
\title{Source the Protein-Protein interaction in STRING database}
\usage{
sourcePPI(TF, STRING.version = "10", species.index, ...)
}
\arguments{
\item{TF}{a data frame with one column indicating the TF of interest}

\item{STRING.version}{a numeric vector indicating the STRING version. Default valuve is 10}

\item{species.index}{a numeric vector indicating NCBI taxonomy identifiers}

\item{...}{any dditional arguments passed to}
}
\value{
A PPI data.frame which contains three columns: "from" and "to" indicating the direction of protein-protein interaction, and "score" indicating the interaction score between two proteins.
}
\description{
This function uses a list of Transcription Factors (TF) of interest to source the Protein-Protein interactions (PPI) from STRING database using all types of interactions not only the physical subnetwork
Important: this function produces a simple unweighted network for tutorial purposes, and does not support weighted PPI edges for the moment.
For more complex PPI network modeling, consider pulling the PPI network directly from STRINGdb directly or through their R package.
}
\examples{
# the example motif file
motif_file_path <- system.file("extdata", "chip_matched.txt", package = "netZooR", mustWork = TRUE)
motif <- read.table(motif_file_path, sep="\t")
# create a TF data frame with one column
TF <-data.frame(motif[,1])
# create PPI data frame by searching in STRING version 10 
# and specifying specie to "Mycobacterium tuberculosis H37Rv".
# STRING verison 11 is only accessible to R 4.0.
\donttest{
if(R.Version()$major=="3"){PPI <- sourcePPI(TF, STRING.version="10", 
species.index=83332, score_threshold=0)}
if(R.Version()$major=="4"){PPI <- sourcePPI(TF, STRING.version="11", 
species.index=83332, score_threshold=0)}
}
# write out locally then can be used in \code{\link{pandaPy}}.

}
