% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcePPI.R
\name{sourcePPI}
\alias{sourcePPI}
\title{Source the Protein-Protein interaction in STRING database}
\usage{
sourcePPI(TF, species.index, score_threshold = 0)
}
\arguments{
\item{TF}{Data frame with one column indicating the TF of interest}

\item{species.index}{Numeric vector indicating NCBI taxonomy identifiers}

\item{score_threshold}{Numeric vector indicating the threshold for the combined scores of the interactions.Default to 0.}
}
\value{
A PPI data frame
}
\description{
This function is able to use a list of Transcription Factors(TF) of interest to source the Protein-Protein interactions (PPI)in STRING database
}
\examples{
# the example motif file
motif_file_path <- system.file("extdata", "chip.txt", package = "netZoo", mustWork = TRUE)
motif <- read.table(motif_file_path, sep="\\t")
# create a TF data frame with one column
TF <- motif[,1]
# create PPI data frame
PPI <- sourcePPI(TF,species.index=83332, score_streshold=0)
# write locally then use in \\code{\\link{runPanda}}.

}
