% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIGER.R
\name{tiger}
\alias{tiger}
\title{TIGER main function}
\usage{
tiger(
  expr,
  prior,
  method = "VB",
  TFexpressed = TRUE,
  signed = TRUE,
  baseline = TRUE,
  psis_loo = FALSE,
  seed = 123,
  out_path = NULL,
  out_size = 300,
  a_sigma = 1,
  b_sigma = 1,
  a_alpha = 1,
  b_alpha = 1,
  sigmaZ = 10,
  sigmaB = 1,
  tol = 0.005
)
}
\arguments{
\item{expr}{A normalized log-transformed gene expressison matrix. Rows are genes
and columns are sampeles (cells).}

\item{prior}{A prior regulatory network in adjacency matrix format. Rows are TFs
and columns target genes.}

\item{method}{Method used for Bayesian inference. "VB" or "MCMC". Defaults to "VB".}

\item{TFexpressed}{TF mRNA needs to be expressed or not. Defaults to TRUE.}

\item{signed}{Prior network is signed or not. Defaults to TRUE.}

\item{baseline}{Include baseline or not. Defaults to TRUE.}

\item{psis_loo}{Use pareto smoothed importance sampling leave-one-out cross
validation to check model fitting or not. Defaults to FALSE.}

\item{seed}{Seed for reproducible results. Defaults to 123.}

\item{out_path}{(Optional) output path for CmdStanVB or CmdStanMCMC object. Defaults to NULL.}

\item{out_size}{Posterior sampling size. Default = 300.}

\item{a_sigma}{Hyperparameter of error term. Default = 1.}

\item{b_sigma}{Hyperparameter of error term. Default = 1.}

\item{a_alpha}{Hyperparameter of edge weight W. Default = 1.}

\item{b_alpha}{Hyperparameter of edge weight W. Default = 1.}

\item{sigmaZ}{Standard deviation of TF activity Z. Default = 10.}

\item{sigmaB}{Standard deviation of baseline term. Default = 1.}

\item{tol}{Convergence tolerance on ELBO.. Default = 0.005.}
}
\value{
A TIGER list object.
* W is the estimated regulatory network, but different from prior network,
rows are genes and columns are TFs.
* Z is the estimated TF activities, rows are TFs and columns are samples.
* TF.name, TG.name, and sample.name are the used TFs, target genes and samples.
* If psis_loo is TRUE, loocv is a table of psis_loo result for model checking.
* If psis_loo is TRUE, elpd_loo is the Bayesian LOO estimate of the expected log pointwise predictive
density, which can be used for Bayesian stacking to handle multi-modality later.
}
\description{
TIGER main function
}
\examples{
if (requireNamespace("cmdstanr", quietly = TRUE)) {
   data(TIGER_expr)
   data(TIGER_prior)
   tiger(TIGER_expr,TIGER_prior)
}
}
