% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{transformation.matrix}
\alias{transformation.matrix}
\title{Bi-partite network analysis tools}
\usage{
transformation.matrix(network.1, network.2, by.tfs = TRUE,
  standardize = FALSE, remove.diagonal = TRUE, method = "ols")
}
\arguments{
\item{by.tfs}{logical indicating a transcription factor based transformation.    If 
false, gives gene by gene transformation matrix}

\item{remove.diagonal}{logical for returning a result containing 0s across the diagonal}

\item{method}{character specifying which algorithm to use, default='kabsch'}

\item{net1}{starting network, a genes by transcription factors data.frame with scores 
for the existence of edges between}

\item{net2}{final network, a genes by transcription factors data.frame with scores 
for the existence of edges between}
}
\value{
matrix object corresponding to transition matrix
}
\description{
This function analyzes a bi-partite network, such as a Transcription factor 
to gene network derived from the PANDA algorithm.
}
\examples{
data(yeast)
cc.net <- monsterNI(yeast$motif,yeast$exp.cc)
sr.net <- monsterNI(yeast$motif,yeast$exp.sr)
transformation.matrix(cc.net, sr.net)
}
\keyword{keywords}

