% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{transitionNetworkPlot}
\alias{transitionNetworkPlot}
\title{This function uses igraph to plot the transition matrix as a network}
\usage{
transitionNetworkPlot(monsterObj, numEdges = 100, numTopTFs = 10)
}
\arguments{
\item{monsterObj}{Monster Object}

\item{numEdges}{The number of edges to display}

\item{numTopTFs}{The number of TFs to display, ranked by largest dTFI}
}
\value{
igraph object for transition matrix
}
\description{
This function uses igraph to plot the transition matrix as a network
}
\examples{
data(yeast)
monsterRes <- monster(yeast$exp.ko,
    c(rep(1,42),rep(0,49),rep(NA,15)),
    yeast$motif, nullPerms=10, numMaxCores=4)
transitionNetworkPlot(monsterRes)

}
\keyword{keywords}

