% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.diff.panda.in.cytoscape.R
\name{vis.diff.panda.in.cytoscape}
\alias{vis.diff.panda.in.cytoscape}
\title{Plot two PANDA networsk in Cytoscape}
\usage{
vis.diff.panda.in.cytoscape(
  merged_panda,
  condition_name = "cond.1",
  network_name = "diff.PANDA"
)
}
\arguments{
\item{merged_panda}{vector indicating the merged PANDA networks in data frame structure type.}

\item{condition_name}{string vector indicating the same condition name used in \code{\link{panda.diff.edges}}.}

\item{network_name}{Character string indicating the name of Cytoscape network.}
}
\value{
Plot two PANDA networsk in Cytoscape
}
\description{
This function is able to modify PANDA network and plot in Cytoscape. Please make sure that Cytoscape 
is installed and open it before calling this function.
}
\examples{
\dontrun{
# create a merged PANDA network from two conditions by running \code{\link{panda.diff.edges}}
merged.panda <- panda.diff.edges(panda.net1, panda.net2, condition_name="condition1")
# then plot two PANDA network in Cytoscape.
vis.diff.panda.in.cytoscape(merged.panda,condition_name = "condition1", network_name="diff.PANDA" )
}
}
