% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.diff.panda.in.cytoscape.R
\name{vis.diff.panda.in.cytoscape}
\alias{vis.diff.panda.in.cytoscape}
\title{Plot two PANDA networsk in Cytoscape}
\usage{
vis.diff.panda.in.cytoscape(merged_panda, condition_name = "cond.1",
  network_name = "diff.PANDA")
}
\arguments{
\item{merged_panda}{vector indicating the merged PANDA networks in data frame structure type.}

\item{condition_name}{string vector indicating the same condition name used in \code{\link{panda.diff.edges}}.}

\item{network_name}{Character string indicating the name of Cytoscape network.}
}
\description{
This function is able to modify PANDA network and plot in Cytoscape.
}
\examples{
\dontrun{
# run this function to create a network in Cytoscape.
vis.diff.panda.in.cytoscape(panda.net1, panda.net1, network.name="diff.PANDA")
}
}
