% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.panda.in.cytoscape.R
\name{vis.panda.in.cytoscape}
\alias{vis.panda.in.cytoscape}
\title{Plot PANDA network in Cytoscape}
\usage{
vis.panda.in.cytoscape(panda.net, network.name = "PANDA")
}
\arguments{
\item{panda.net}{Character string indicating the input PANDA network in data frame structure type.}

\item{network.name}{Character string indicating the name of Cytoscape network.}
}
\description{
This function is able to modify PANDA network and plot in Cytoscape.
}
\examples{
\dontrun{
# refer to the input datasets files of control TB dataset in inst/extdat as example
control_expression_file_path <- system.file("extdata", "expr10_matched.txt", package = "netZooR", mustWork = TRUE)
motif_file_path <- system.file("extdata", "chip_matched.txt", package = "netZooR", mustWork = TRUE)
ppi_file_path <- system.file("extdata", "ppi_matched.txt", package = "netZooR", mustWork = TRUE)

# Run PANDA algorithm
control_all_panda_result <- panda.fast(e = control_expression_file_path, m = motif_file_path, ppi = ppi_file_path, rm_missing = TRUE )

# access PANDA regulatory network
control_net <- control_all_panda_result$panda

# select top 1000 edges in PANDA network by edge weight.
panda.net <- head(control_net[order(control_net$force,decreasing = TRUE),], 1000)

# run this function to create a network in Cytoscape.
vis.panda.in.cytoscape(panda.net, network.name="PANDA")
}
}
