% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visPandaInCytoscape.R
\name{visPandaInCytoscape}
\alias{visPandaInCytoscape}
\title{Plot PANDA network in Cytoscape}
\usage{
visPandaInCytoscape(panda.net, network_name = "PANDA")
}
\arguments{
\item{panda.net}{Character string indicating the input PANDA network in data frame structure type.}

\item{network_name}{Character string indicating the name of Cytoscape network.}
}
\value{
PANDA network in Cytoscape
}
\description{
This function is able to modify PANDA network and plot in Cytoscape. Please make sure that Cytoscape 
is installed and open it before calling this function.
}
\examples{
\donttest{
# refer to the input datasets files of control TB dataset in inst/extdat as example
control_expression_file_path <- system.file("extdata", "expr10_matched.txt", 
package = "netZooR", mustWork = TRUE)
motif_file_path <- system.file("extdata", "chip_matched.txt", package = "netZooR", mustWork = TRUE)
ppi_file_path <- system.file("extdata", "ppi_matched.txt", package = "netZooR", mustWork = TRUE)

# Run PANDA algorithm
control_all_panda_result <- pandaPy(expr = control_expression_file_path, motif = motif_file_path, 
ppi = ppi_file_path, modeProcess="legacy", remove_missing = TRUE )

# access PANDA regulatory network
control_net <- control_all_panda_result$panda

# select top 1000 edges in PANDA network by edge weight.
panda.net <- head(control_net[order(control_net$Score,decreasing = TRUE),], 1000)

# run this function to create a network in Cytoscape.
visPandaInCytoscape(panda.net, network_name="PANDA")
}
}
