% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{writeGeneGene}
\alias{writeGeneGene}
\title{Generates the gene correlation results in the format expected by the UI.
Format is Gene A	Gene B	Tissue	Correlation}
\usage{
writeGeneGene(inFiles, file, removeLowerTri = TRUE, cutoff = -2)
}
\arguments{
\item{inFiles}{The SEAHORSE files}

\item{file}{The file where results should be stored. .tsv.gz will be appended.}

\item{removeLowerTri}{Whether or not to remove the lower triangular matrix}

\item{cutoff}{The correlation cutoff for filtering. Default is -2 (all retained).
(default is TRUE)}
}
\description{
Generates the gene correlation results in the format expected by the UI.
Format is Gene A	Gene B	Tissue	Correlation
}
