% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{writePhenotypeGeneAssociations}
\alias{writePhenotypeGeneAssociations}
\title{Generates the phenotype data in the format expected by the UI.
Format is VARNAME GENE    tissue  TEST    TESTSTAT        TESTPVALUE}
\usage{
writePhenotypeGeneAssociations(inFiles, resultFiles, file)
}
\arguments{
\item{inFiles}{The SEAHORSE inputs}

\item{resultFiles}{The SEAHORSE results}

\item{file}{The file where results should be stored. .tsv.gz will be appended.}
}
\description{
Generates the phenotype data in the format expected by the UI.
Format is VARNAME GENE    tissue  TEST    TESTSTAT        TESTPVALUE
}
