% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{writePhenotypePhenotypeAssociations}
\alias{writePhenotypePhenotypeAssociations}
\title{Generates the phenotype results in the format expected by the UI.
Format is VARNAME1        VARNAME2        tissue  TEST    TESTSTAT        TESTPVALUE}
\usage{
writePhenotypePhenotypeAssociations(inFiles, file)
}
\arguments{
\item{inFiles}{The SEAHORSE results}

\item{file}{The file where results should be stored. .tsv.gz will be appended.}
}
\description{
Generates the phenotype results in the format expected by the UI.
Format is VARNAME1        VARNAME2        tissue  TEST    TESTSTAT        TESTPVALUE
}
