% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{writePhenotypePhenotypeAssociations}
\alias{writePhenotypePhenotypeAssociations}
\title{Generates the phenotype results in the format expected by the UI.
Format is VARNAME1        VARNAME2        tissue  TEST    TESTSTAT        TESTPVALUE}
\usage{
writePhenotypePhenotypeAssociations(
  inFiles,
  file,
  corCutoff = -2,
  padjCutoff = 2
)
}
\arguments{
\item{inFiles}{The SEAHORSE results}

\item{file}{The file where results should be stored. .tsv.gz will be appended.}

\item{corCutoff}{The cutoff for correlation. Default is -2 (all values will be retained.)}

\item{padjCutoff}{The cutoff for adjusted p-value. Default is 2 (all values will be retained.)}
}
\description{
Generates the phenotype results in the format expected by the UI.
Format is VARNAME1        VARNAME2        tissue  TEST    TESTSTAT        TESTPVALUE
}
