% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{writeSigTable}
\alias{writeSigTable}
\title{Write the table containing the top results across tissues.}
\usage{
writeSigTable(
  seahorseResultDir,
  seahorseInputDir,
  resultType,
  consolidatedResultFile,
  pathways = NULL,
  corCutoffSignificance = 0.9,
  padjCutoffSignificance = 0.05
)
}
\arguments{
\item{seahorseResultDir}{The directory containing the SEAHORSE results across tissues}

\item{seahorseInputDir}{The directory containing input files for SEAHORSE, with the file
names corresponding to the file names in seahorseResultDir and each file formatted as
a named list of the following: expression, phenotype, dict.}

\item{resultType}{The name of the result component ("coexpression", "phenotype_association",
"phenocor", or "GSEA")}

\item{consolidatedResultFile}{The path where you wish to save your consolidated file.}

\item{pathways}{The pathways in the format used by SEAHORSE}

\item{corCutoffSignificance}{The correlation cutoff significance value. Default is 0.9.}

\item{padjCutoffSignificance}{The significance cutoff for adjusted p-value.
Default is 0.05.}
}
\description{
Write the table containing the top results across tissues.
}
