% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAHORSE.R
\name{writeTable}
\alias{writeTable}
\title{Write table of SEAHORSE results filtered by a variable}
\usage{
writeTable(
  result,
  phenotype = NULL,
  variable,
  variableType,
  resultType,
  tmpFile,
  resultFile,
  dictionary = NULL,
  pathways = NULL
)
}
\arguments{
\item{result}{The full SEAHORSE result}

\item{phenotype}{The phenotype data in the format used by SEAHORSE}

\item{variable}{The variable of interest by which to filter (a phenotype or gene)}

\item{variableType}{Either "gene" or "phenotype"}

\item{resultType}{The name of the result component ("coexpression", "phenotype_association",
"phenocor", or "GSEA")}

\item{tmpFile}{The temporary file you wish to create for constructing the table.}

\item{resultFile}{The location where you wish to save your result file.}

\item{dictionary}{The phenotype dictionary in the format used by SEAHORSE}

\item{pathways}{The pathways in the format used by SEAHORSE}
}
\description{
Write table of SEAHORSE results filtered by a variable
}
